Consolidate onboarding component icons for single sources
