--- packages/js/onboarding/src/components/WCPayConnectCard/Icons/affirm.js
@@ -1,39 +1,0 @@
-/**
- * External dependencies
- */
-import { createElement } from '@wordpress/element';
-
-export const Affirm = () => (
-	/* eslint-disable */
-	<svg
-		xmlns="http://www.w3.org/2000/svg"
-		viewBox="0 0 64 40"
-		fill="none"
-	>
-		<path fill="#fff" d="M0 0h64v40H0z"/>
-		<mask
-			id="a"
-			width={42}
-			height={18}
-			x={11}
-			y={10}
-			maskUnits="userSpaceOnUse"
-			style={{
-				maskType: "luminance",
-			}}
-		>
-			<path fill="#fff" d="M53 10.62H11V27.38h42V10.62Z"/>
-		</mask>
-		<g mask="url(#a)">
-			<path
-				fill="#101820"
-				d="M13.754 25.935c-.52 0-.784-.256-.784-.676 0-.784.874-1.048 2.469-1.218 0 1.045-.707 1.894-1.69 1.894h.005Zm.686-5.875c-1.14 0-2.45.536-3.162 1.103l.65 1.372c.57-.522 1.493-.97 2.325-.97.79 0 1.228.265 1.228.799 0 .357-.29.539-.836.61-2.045.265-3.647.83-3.647 2.404 0 1.249.889 2.005 2.278 2.005.992 0 1.873-.551 2.293-1.275v1.078h1.848V22.67c0-1.862-1.292-2.613-2.976-2.613l-.001.004ZM32.991 20.263v6.915h1.978v-3.332c0-1.583.958-2.048 1.626-2.048.3-.003.595.084.846.25l.361-1.828a2.24 2.24 0 0 0-.882-.16c-1.016 0-1.655.45-2.077 1.365v-1.162h-1.852ZM46.967 20.06c-1.045 0-1.828.617-2.234 1.213-.377-.77-1.176-1.213-2.137-1.213-1.044 0-1.769.58-2.103 1.248v-1.046h-1.907v6.916h1.98v-3.56c0-1.275.669-1.891 1.293-1.891.566 0 1.085.365 1.085 1.309v4.142h1.976v-3.56c0-1.293.653-1.891 1.306-1.891.524 0 1.078.38 1.078 1.295v4.156h1.976v-4.781c0-1.553-1.046-2.337-2.309-2.337M28.179 20.263h-1.792v-.702c0-.916.523-1.177.972-1.177.308.005.61.08.882.221l.61-1.395s-.618-.403-1.742-.403c-1.264 0-2.702.712-2.702 2.948v.508h-2.994v-.702c0-.916.522-1.177.973-1.177.307 0 .61.076.882.221l.61-1.395c-.364-.212-.95-.403-1.742-.403-1.264 0-2.701.712-2.701 2.948v.508H18.29v1.525h1.148v5.39h1.976v-5.39h2.997v5.39h1.976v-5.39h1.792v-1.525ZM31.332 20.263h-1.974v6.912h1.974v-6.912Z"
-			/>
-			<path
-				fill="#4A4AF4"
-				d="M29.43 19.088h1.934c1.128-3.547 4.955-6.665 9.507-6.665 5.535 0 10.319 4.215 10.319 10.775a13.48 13.48 0 0 1-.55 3.98h1.878l.018-.065c.317-1.279.473-2.592.465-3.91 0-7.315-5.332-12.58-12.126-12.58-5.338 0-10.095 3.704-11.443 8.467l-.001-.002Z"
-			/>
-		</g>
-	</svg>
-	/* eslint-enable */
-);
