--- packages/js/onboarding/src/components/WCPayConnectCard/Icons/afterpay.js
@@ -1,23 +1,0 @@
-/**
- * External dependencies
- */
-import { createElement } from '@wordpress/element';
-
-export const Afterpay = () => (
-	/* eslint-disable */
-	<svg
-		xmlns="http://www.w3.org/2000/svg"
-		data-name="Layer 1"
-		viewBox="0 0 38 24"
-	>
-		<path
-			d="M35 0H3a3 3 0 0 0-3 3v18a3 3 0 0 0 3 3h32a3 3 0 0 0 3-3V3a3 3 0 0 0-3-3Z"
-			style={{
-				fill: "#b2fce4",
-			}}
-		/>
-		<path
-			d="m36.12 10-1.06-.61L34 8.75a1.07 1.07 0 0 0-1.6.93v.13a.23.23 0 0 0 .1.19l.5.28a.2.2 0 0 0 .28-.08.19.19 0 0 0 0-.1v-.32a.21.21 0 0 1 .2-.22h.12l1 .57 1 .56a.22.22 0 0 1 .07.3l-.07.07-1 .56-1 .57a.21.21 0 0 1-.29-.07.24.24 0 0 1 0-.12v-.16a1.07 1.07 0 0 0-1.6-.93l-1.08.62-1.06.61a1.08 1.08 0 0 0-.4 1.46 1 1 0 0 0 .4.4l1.06.61 1.08.61a1.07 1.07 0 0 0 1.6-.93v-.13a.23.23 0 0 0-.1-.19l-.5-.29a.2.2 0 0 0-.28.08.23.23 0 0 0 0 .11v.32a.21.21 0 0 1-.2.22.27.27 0 0 1-.12 0l-1-.57-1-.56A.22.22 0 0 1 30 13l.07-.07 1-.56 1-.57a.21.21 0 0 1 .29.07.24.24 0 0 1 0 .12v.16a1.07 1.07 0 0 0 1.6.93l1.08-.62 1.06-.61a1.06 1.06 0 0 0 .34-1.47 1 1 0 0 0-.32-.38Zm-6.98.18-2.49 5.14h-1l.93-1.92-1.46-3.22h1.06l.94 2.15 1-2.15ZM4.25 12a1 1 0 1 0-1 1.05 1 1 0 0 0 1-1V12m0 1.83v-.48a1.49 1.49 0 0 1-1.16.54 1.79 1.79 0 0 1-1.75-1.83V12a1.82 1.82 0 0 1 1.75-1.89 1.47 1.47 0 0 1 1.14.53v-.46h.89v3.65ZM9.52 13c-.32 0-.4-.12-.4-.42V11h.57v-.79h-.57v-.92H8.2v.89H7v-.37c0-.3.12-.42.44-.42h.2v-.71h-.43c-.76 0-1.12.25-1.12 1v.49h-.51V11h.51v2.86H7V11h1.2v1.79c0 .75.28 1.07 1 1.07h.5V13Zm3.3-1.33a.87.87 0 0 0-.91-.75.89.89 0 0 0-.92.75Zm-1.82.57a.91.91 0 0 0 .94.85 1 1 0 0 0 .86-.47h.94a1.79 1.79 0 0 1-1.82 1.27A1.8 1.8 0 0 1 10 12.22V12a1.89 1.89 0 1 1 3.78 0 .89.89 0 0 1 0 .23Zm8.66-.24a1 1 0 1 0 0 .05V12m-2.89 3.32v-5.14h.89v.47a1.49 1.49 0 0 1 1.16-.54 1.79 1.79 0 0 1 1.76 1.82V12a1.82 1.82 0 0 1-1.75 1.89 1.44 1.44 0 0 1-1.11-.49v1.92ZM23.82 12a1 1 0 1 0-1.05 1.05h.05a1 1 0 0 0 1-1V12m0 1.83v-.48a1.46 1.46 0 0 1-1.15.54 1.79 1.79 0 0 1-1.76-1.82V12a1.82 1.82 0 0 1 1.75-1.89 1.45 1.45 0 0 1 1.13.53v-.46h.9v3.65Zm-8.67-3.29a.91.91 0 0 1 .79-.43.88.88 0 0 1 .39.08v.94a1.15 1.15 0 0 0-.65-.17.64.64 0 0 0-.5.71v2.16h-.93v-3.65h.9Z"/>
-	</svg>
-	/* eslint-enable */
-);
