--- packages/js/onboarding/src/components/WCPayConnectCard/Icons/amex.js
@@ -1,22 +1,0 @@
-/**
- * External dependencies
- */
-import { createElement } from '@wordpress/element';
-
-export const Amex = () => (
-	/* eslint-disable */
-	<svg viewBox="0 0 64 40" fill="none" xmlns="http://www.w3.org/2000/svg">
-		<g clipPath="url(#clip0_7623_49052)">
-			<rect width="64" height="40" rx="3" fill="white"/>
-			<path
-				d="M56.5333 9.81335L57.8133 6.33335L64 6.38669V0.0533428L0 0V40L64 40.0533V33.7333L58.1467 33.68L56.1067 31.2934L53.9867 33.68H38.3067V20.9734H33.1333L39.6 6.33335H45.8933L47.4133 9.65334V6.33335H55.2267L56.5333 9.81335ZM52.12 12.12L52.1067 10.72L52.64 12.12L55.24 19.0667H57.8267L60.44 12.12L60.9467 10.7333V19.0667L64 19.12V8.45334L59.1467 8.4L57.0933 13.8134L56.5467 15.28L55.9867 13.8134L53.92 8.4H49.4V19.0667H52.12V12.12ZM46.2667 19.0667H49.4L44.7067 8.4H41.0667L36.3466 19.0667H39.44L40.2667 17.0134H45.44L46.2667 19.0667ZM42.32 12.0134L42.8533 10.68L43.3867 12.0134L44.4933 14.7066H41.2133L42.32 12.0134ZM40.4933 20.9866V31.6134H49.4V29.3066H43.2133V27.4534H49.28V25.16H43.2133V23.2934H49.4V20.9866H40.4933ZM59.3067 31.6134H62.84L57.8533 26.28L62.84 20.9866H59.36L56.1467 24.44L52.9467 20.9866H49.4L54.3733 26.32L49.4 31.6134H52.84L56.08 28.1467L59.3067 31.6134ZM60.6533 26.2666L64 29.3867V23.2934L60.6533 26.2666Z"
-				fill="#006FCF"/>
-		</g>
-		<defs>
-			<clipPath id="clip0_7623_49052">
-				<rect width="64" height="40" rx="3" fill="white"/>
-			</clipPath>
-		</defs>
-	</svg>
-	/* eslint-enable */
-);
