--- packages/js/onboarding/src/components/WCPayConnectCard/Icons/cb.js
@@ -1,24 +1,0 @@
-/**
- * External dependencies
- */
-import { createElement } from '@wordpress/element';
-
-export const CB = () => (
-	/* eslint-disable */
-	<svg viewBox="0 0 39 25" fill="none" xmlns="http://www.w3.org/2000/svg">
-		<rect x="1.09985" y="0.723633" width="37.4" height="23" rx="1.5" fill="url(#paint0_linear_685_110690)"/>
-		<rect x="1.09985" y="0.723633" width="37.4" height="23" rx="1.5" stroke="#DDDDDD"/>
-		<path fillRule="evenodd" clipRule="evenodd"
-			  d="M13.307 11.7158H20.0383C19.9783 10.2608 19.6393 8.76069 18.6111 7.79714C17.3863 6.64937 15.241 6.22363 13.3241 6.22363C11.3289 6.22363 9.1224 6.6906 7.89114 7.92209C6.82939 8.98291 6.59985 10.6897 6.59985 12.2233C6.59985 13.8297 7.04892 15.6982 8.20072 16.7744C9.4255 17.9196 11.4101 18.2236 13.3241 18.2236C15.1833 18.2236 17.1974 17.8803 18.4151 16.7931C19.6293 15.7072 20.0489 13.878 20.0489 12.2233V12.2156H13.307V11.7158ZM20.4897 12.2158V17.9553H29.8536V17.9476C31.2238 17.8729 32.3138 16.6375 32.3138 15.1207C32.3138 13.6032 31.2238 12.2886 29.8536 12.2133V12.2158H20.4897ZM29.7471 6.46388C31.0825 6.46388 32.1425 7.62259 32.1425 9.08467C32.1425 10.4688 31.169 11.5992 29.9307 11.7158H20.4897V6.45551H29.4516C29.5068 6.44876 29.571 6.45351 29.6339 6.45814C29.673 6.46103 29.7116 6.46388 29.7471 6.46388Z"
-			  fill="#FEFEFE"/>
-		<defs>
-			<linearGradient id="paint0_linear_685_110690" x1="11.0155" y1="-2.81556" x2="3.21751" y2="23.5991"
-							gradientUnits="userSpaceOnUse">
-				<stop stopColor="#222E72"/>
-				<stop offset="0.591647" stopColor="#40CBFF"/>
-				<stop offset="1" stopColor="#3CB792"/>
-			</linearGradient>
-		</defs>
-	</svg>
-	/* eslint-enable */
-);
