--- packages/js/onboarding/src/components/WCPayConnectCard/Icons/clearpay.js
@@ -1,15 +1,0 @@
-/**
- * External dependencies
- */
-import { createElement } from '@wordpress/element';
-
-export const Clearpay = () => (
-	/* eslint-disable */
-	<svg viewBox="0 0 64 40" fill="none" xmlns="http://www.w3.org/2000/svg">
-		<path fill="#fff" d="M0 0h64v40H0z"/>
-		<path
-			d="m48.47 17.208-3.827 7.897h-1.586l1.43-2.95-2.247-4.947h1.629l1.444 3.314 1.576-3.314h1.581ZM33.905 20.012c0-.906-.684-1.605-1.524-1.605s-1.525.675-1.525 1.605c0 .92.685 1.605 1.525 1.605.84 0 1.524-.699 1.524-1.605Zm-4.441 5.093v-7.897h1.378v.727a2.228 2.228 0 0 1 1.77-.83c1.515 0 2.704 1.245 2.704 2.893 0 1.647-1.213 2.902-2.737 2.902-.708 0-1.288-.278-1.695-.75v2.95h-1.42v.005ZM40.29 20.012c0-.944-.684-1.605-1.524-1.605s-1.524.675-1.524 1.605c0 .92.684 1.605 1.524 1.605s1.525-.666 1.525-1.605Zm.01 2.803v-.726a2.245 2.245 0 0 1-1.77.816c-1.538 0-2.704-1.232-2.704-2.893 0-1.647 1.213-2.903 2.737-2.903.718 0 1.322.316 1.737.807v-.708h1.379v5.607H40.3ZM26.976 17.756s.35-.652 1.213-.652c.368 0 .604.128.604.128v1.43s-.519-.321-.995-.255c-.477.066-.78.5-.78 1.086v3.322h-1.425v-5.607h1.379v.548h.004ZM59.175 16.892l-1.633-.934-1.657-.95c-1.095-.627-2.463.161-2.463 1.426v.213a.32.32 0 0 0 .165.283l.77.439a.319.319 0 0 0 .476-.279v-.505c0-.25.269-.406.486-.283l1.51.869 1.506.863a.324.324 0 0 1 0 .562l-1.506.864-1.51.868a.325.325 0 0 1-.486-.283v-.25c0-1.265-1.369-2.058-2.464-1.426l-1.657.949-1.633.934c-1.1.633-1.1 2.224 0 2.856l1.633.935 1.657.948c1.095.628 2.464-.16 2.464-1.425v-.212a.32.32 0 0 0-.165-.284l-.77-.439a.319.319 0 0 0-.476.279v.505c0 .25-.27.406-.487.283l-1.51-.868-1.506-.864a.324.324 0 0 1 0-.562l1.506-.863 1.51-.869a.325.325 0 0 1 .487.283v.25c0 1.265 1.368 2.058 2.463 1.426l1.657-.949 1.633-.934c1.1-.638 1.1-2.224 0-2.856ZM9.749 20.724C9.447 22.014 8.338 22.9 6.903 22.9 5.246 22.9 4 21.654 4 20.007c0-1.647 1.265-2.903 2.926-2.903 1.402 0 2.497.888 2.813 2.153H8.281a1.524 1.524 0 0 0-1.345-.84c-.83 0-1.525.684-1.525 1.59s.694 1.59 1.525 1.59a1.47 1.47 0 0 0 1.354-.873H9.75ZM10.405 22.81V14.9h1.411v7.91h-1.411ZM13.94 20.38c.1.83.694 1.298 1.444 1.298.595 0 1.053-.279 1.322-.727h1.444c-.335 1.19-1.402 1.95-2.803 1.95-1.695 0-2.88-1.19-2.88-2.88s1.256-2.912 2.913-2.912c1.67 0 2.879 1.232 2.879 2.912 0 .123-.01.246-.033.359H13.94Zm2.813-.878c-.099-.727-.694-1.166-1.388-1.166-.693 0-1.264.425-1.41 1.166h2.798ZM23.3 22.81v-.726c-.416.505-1.03.816-1.77.816-1.535 0-2.7-1.231-2.7-2.893 0-1.647 1.208-2.903 2.732-2.903.718 0 1.322.312 1.737.808v-.709h1.379v5.603h-1.379v.005Zm-.015-2.803c0-.94-.684-1.605-1.524-1.605s-1.525.67-1.525 1.605c0 .92.684 1.605 1.525 1.605.84 0 1.524-.66 1.524-1.605Z"
-			fill="#000"/>
-	</svg>
-	/* eslint-enable */
-);
