--- packages/js/onboarding/src/components/WCPayConnectCard/Icons/jcb.js
@@ -1,41 +1,0 @@
-/**
- * External dependencies
- */
-import { createElement } from '@wordpress/element';
-
-export const JCB = () => (
-	/* eslint-disable */
-	<svg viewBox="0 0 64 40" fill="none" xmlns="http://www.w3.org/2000/svg">
-		<rect width="64" height="40" fill="#0E4C96"/>
-		<path fillRule="evenodd" clipRule="evenodd"
-			  d="M52.6712 29.4208C52.6712 30.2653 52.5048 31.1015 52.1816 31.8817C51.8585 32.6619 51.3848 33.3709 50.7876 33.968C50.1905 34.5652 49.4816 35.0388 48.7014 35.362C47.9211 35.6852 47.0849 35.8515 46.2404 35.8515H11.3289V10.4316C11.3287 9.58701 11.495 8.75069 11.8181 7.97038C12.1412 7.19007 12.6149 6.48105 13.2121 5.88382C13.8092 5.28658 14.5182 4.81282 15.2985 4.48959C16.0787 4.16636 16.915 4 17.7596 4H52.6712V29.4208Z"
-			  fill="white"/>
-		<path fillRule="evenodd" clipRule="evenodd"
-			  d="M41.2893 20.4839C42.286 20.5061 43.2895 20.4403 44.2828 20.518C45.2881 20.7057 45.5304 22.2289 44.637 22.7281C44.0277 23.0558 43.3032 22.8493 42.6419 22.9073H41.2893V20.4839ZM44.8588 17.7413C45.0807 18.5229 44.3263 19.2261 43.5737 19.1177H41.2893C41.3047 18.3804 41.2578 17.58 41.3124 16.8819C42.2271 16.9075 43.1513 16.829 44.061 16.9221C44.2569 16.9711 44.4353 17.0737 44.5762 17.2184C44.7171 17.3631 44.815 17.5441 44.8588 17.7413ZM50.3569 6.14446C50.3995 7.63779 50.3628 9.20963 50.3748 10.7337C50.3722 16.9289 50.3816 23.1241 50.3705 29.3184C50.3304 31.6395 48.273 33.6568 45.9673 33.7037C43.659 33.7122 41.3508 33.7046 39.0434 33.7071V24.3418C41.5581 24.329 44.0746 24.3682 46.5877 24.3222C47.7542 24.2488 49.0316 23.4799 49.0854 22.1965C49.2228 20.908 48.0076 20.0162 46.8539 19.8754C46.4102 19.8635 46.4238 19.7457 46.8539 19.6945C47.9547 19.4565 48.8183 18.3181 48.4949 17.1772C48.2184 15.9774 46.8923 15.5132 45.7898 15.5158C43.5413 15.5004 41.2927 15.5132 39.0434 15.5089C39.0579 13.7605 39.0126 12.0103 39.0673 10.2635C39.2456 7.98339 41.3542 6.1035 43.6283 6.14446H50.3569Z"
-			  fill="url(#paint0_linear_17_2768)"/>
-		<path fillRule="evenodd" clipRule="evenodd"
-			  d="M13.6311 10.4999C13.6883 8.18221 15.755 6.18114 18.0573 6.14786C20.3571 6.14018 22.6568 6.14615 24.9557 6.14445C24.9497 13.8995 24.9685 21.6555 24.9463 29.4114C24.8575 31.7009 22.8138 33.6636 20.5363 33.7037C18.2323 33.7122 15.9283 33.7054 13.6243 33.7071V24.026C15.8626 24.5551 18.2092 24.7795 20.4919 24.4288C21.8572 24.2095 23.3497 23.5396 23.8114 22.124C24.1518 20.9122 23.9607 19.6382 24.011 18.3949V15.5089H20.0601C20.043 17.4178 20.0968 19.3302 20.0319 21.2374C19.9253 22.409 18.7647 23.1539 17.6597 23.1147C16.2884 23.1283 13.5722 22.1206 13.5722 22.1206C13.5654 18.5434 13.6115 14.0651 13.6311 10.4999Z"
-			  fill="url(#paint1_linear_17_2768)"/>
-		<path fillRule="evenodd" clipRule="evenodd"
-			  d="M26.4294 16.6737C26.2221 16.7181 26.3868 15.9654 26.3347 15.6796C26.3484 13.8748 26.3048 12.0683 26.3586 10.2652C26.5361 7.97568 28.6626 6.09067 30.9444 6.14357H37.6653C37.6593 13.9004 37.6781 21.6555 37.6567 29.4106C37.5671 31.7009 35.5234 33.6636 33.2459 33.7037C30.9419 33.7122 28.6379 33.7054 26.3339 33.7071V23.1002C27.9083 24.3921 30.0459 24.5935 32.0068 24.5961C33.4848 24.5961 34.9534 24.3682 36.3887 24.0269V22.0847C34.7708 22.8911 32.8695 23.4031 31.0767 22.9397C29.8265 22.6283 28.9186 21.4199 28.9391 20.1289C28.794 18.7866 29.5808 17.37 30.9 16.9715C32.5384 16.4578 34.3219 16.8503 35.8579 17.5168C36.1864 17.6892 36.5201 17.9025 36.3887 17.353V15.8255C33.821 15.2145 31.0895 14.9892 28.5099 15.6548C27.7632 15.8647 27.0361 16.1839 26.4294 16.6746V16.6737Z"
-			  fill="url(#paint2_linear_17_2768)"/>
-		<defs>
-			<linearGradient id="paint0_linear_17_2768" x1="39.3796" y1="1384.35" x2="1172.57" y2="1384.35"
-							gradientUnits="userSpaceOnUse">
-				<stop stopColor="#007B40"/>
-				<stop offset="1" stopColor="#55B330"/>
-			</linearGradient>
-			<linearGradient id="paint1_linear_17_2768" x1="18.9229" y1="1384.36" x2="1152.05" y2="1384.36"
-							gradientUnits="userSpaceOnUse">
-				<stop stopColor="#1D2970"/>
-				<stop offset="1" stopColor="#006DBA"/>
-			</linearGradient>
-			<linearGradient id="paint2_linear_17_2768" x1="27.5695" y1="1384.42" x2="1160.79" y2="1384.42"
-							gradientUnits="userSpaceOnUse">
-				<stop stopColor="#6E2B2F"/>
-				<stop offset="1" stopColor="#E30138"/>
-			</linearGradient>
-		</defs>
-	</svg>
-	/* eslint-enable */
-);
