--- packages/js/onboarding/src/components/WCPayConnectCard/Icons/unionpay.js
@@ -1,31 +1,0 @@
-/**
- * External dependencies
- */
-import { createElement } from '@wordpress/element';
-
-export const UnionPay = () => (
-	/* eslint-disable */
-	<svg viewBox="0 0 64 40" fill="none" xmlns="http://www.w3.org/2000/svg">
-		<rect width="64" height="40" fill="white"/>
-		<g clipPath="url(#clip0_17_2784)">
-			<path
-				d="M17.3827 5H29.3414C31.0102 5 32.0489 6.36125 31.6589 8.03563L26.0939 31.94C25.7002 33.6088 24.0314 34.97 22.3589 34.97H10.4002C8.73142 34.97 7.69267 33.6088 8.08267 31.94L13.6514 8.03563C14.0414 6.36125 15.712 5 17.3827 5Z"
-				fill="#E21836"/>
-			<path
-				d="M28.3456 5H42.0988C43.7675 5 43.0138 6.36125 42.6238 8.03563L37.055 31.94C36.665 33.6088 36.7869 34.97 35.1144 34.97H21.3613C19.6888 34.97 18.6538 33.6088 19.0475 31.94L24.6144 8.03563C25.0063 6.36125 26.675 5 28.3438 5"
-				fill="#00447C"/>
-			<path
-				d="M41.5531 5H53.5119C55.1825 5 56.2212 6.36125 55.8294 8.03563L50.2606 31.94C49.8687 33.6088 48.1981 34.97 46.5256 34.97H34.5725C32.9 34.97 31.8631 33.6088 32.2531 31.94L37.8219 8.03563C38.21 6.36125 39.8806 5 41.5531 5Z"
-				fill="#007B84"/>
-			<path
-				d="M20.5063 12.6613C19.2763 12.6725 18.9125 12.6613 18.7963 12.6331C18.7532 12.845 17.9225 16.6738 17.9207 16.6756C17.7407 17.45 17.6113 18.0031 17.1707 18.3594C16.9229 18.5649 16.6093 18.674 16.2875 18.6669C15.7438 18.6669 15.425 18.395 15.3707 17.8813L15.3613 17.705L15.5263 16.6625C15.5263 16.6625 16.3982 13.175 16.5538 12.7138L16.5669 12.6613C14.8719 12.6763 14.57 12.6613 14.5494 12.6331C14.5382 12.6706 14.4969 12.8863 14.4969 12.8863L13.6063 16.8181L13.5294 17.1519L13.3813 18.2431C13.3725 18.525 13.4378 18.8043 13.5707 19.0531C13.9757 19.7619 15.1325 19.8688 15.785 19.8688C16.6288 19.8688 17.4182 19.6888 17.9525 19.3625C18.8788 18.815 19.1225 17.9581 19.3382 17.1969L19.4375 16.8069C19.4375 16.8069 20.3357 13.1844 20.4875 12.7138C20.49 12.6954 20.4957 12.6776 20.5044 12.6613H20.5063ZM23.5588 15.5825C23.3432 15.5825 22.9475 15.635 22.5913 15.8094C22.4638 15.8769 22.3419 15.9519 22.2125 16.0288L22.3288 15.6088L22.265 15.5394C21.5132 15.6913 21.3444 15.7119 20.6488 15.8094L20.5907 15.8469C20.51 16.5181 20.4369 17.0206 20.1388 18.3369C20.0244 18.8225 19.9063 19.3119 19.7882 19.7956L19.82 19.8575C20.5325 19.82 20.7482 19.82 21.3669 19.8294L21.4175 19.775C21.4963 19.3719 21.5075 19.2781 21.68 18.4625C21.7625 18.0744 21.9332 17.225 22.0175 16.9213C22.1573 16.8428 22.3118 16.7943 22.4713 16.7788C22.8182 16.7788 22.7769 17.0825 22.7619 17.2025C22.6956 17.6842 22.6054 18.1622 22.4919 18.635L22.4057 19.0006C22.3457 19.2725 22.2782 19.535 22.2182 19.8031L22.2444 19.8575C22.9475 19.82 23.1613 19.82 23.7613 19.8294L23.8325 19.775C23.9394 19.1469 23.9713 18.9781 24.1644 18.0613L24.2619 17.6394C24.4494 16.8144 24.5432 16.3981 24.4007 16.0569C24.2507 15.6763 23.8907 15.5825 23.5569 15.5825H23.5588ZM26.9694 16.445C26.5944 16.5181 26.3563 16.565 26.12 16.595C25.8857 16.6344 25.6569 16.67 25.295 16.7188L25.2669 16.745L25.2407 16.7638C25.2032 17.0338 25.1769 17.2663 25.1282 17.54C25.085 17.8213 25.0213 18.1419 24.9144 18.6031C24.8725 18.8071 24.8155 19.0077 24.7438 19.2031C24.6988 19.3288 24.6482 19.4488 24.5563 19.7956L24.5769 19.8275L24.5957 19.8575C24.9332 19.8406 25.1544 19.8294 25.3794 19.8275C25.6063 19.82 25.8407 19.8275 26.2044 19.8294L26.2363 19.805L26.27 19.775C26.3225 19.4619 26.33 19.3775 26.3638 19.2256C26.3938 19.0625 26.45 18.8356 26.5832 18.23C26.6469 17.945 26.7182 17.6619 26.7838 17.3713L26.9919 16.5125L26.9825 16.4788L26.9694 16.4469V16.445ZM26.9769 15.2825C26.6375 15.0819 26.0413 15.1456 25.6419 15.4231C25.2407 15.6931 25.1957 16.0794 25.5332 16.2819C25.8669 16.4769 26.465 16.4188 26.8625 16.1394C27.2619 15.8619 27.3125 15.4813 26.975 15.2825H26.9769ZM29.0282 19.9194C29.7144 19.9194 30.4194 19.7319 30.9482 19.1694C31.355 18.7138 31.5425 18.0369 31.6063 17.7575C31.8182 16.8331 31.6532 16.4019 31.4469 16.1394C31.1338 15.7381 30.5807 15.6088 30.0069 15.6088C29.6619 15.6088 28.8407 15.6444 28.1975 16.235C27.7363 16.6625 27.5225 17.2419 27.395 17.7969C27.2657 18.3631 27.1175 19.3831 28.0532 19.7619C28.3438 19.8856 28.76 19.9194 29.0282 19.9194ZM28.9757 17.84C29.1332 17.1406 29.3207 16.5519 29.7969 16.5519C30.1719 16.5519 30.1982 16.9888 30.0313 17.6919C30.0013 17.8475 29.8644 18.4269 29.6807 18.6744C29.5513 18.8563 29.3994 18.9669 29.2307 18.9669C29.18 18.9669 28.8819 18.9669 28.8782 18.5244C28.8763 18.305 28.9194 18.0819 28.9757 17.84ZM33.3238 19.8294L33.3763 19.775C33.4513 19.3719 33.4644 19.2781 33.6332 18.4625C33.7175 18.0744 33.8919 17.225 33.9744 16.9213C34.13 16.85 34.28 16.7788 34.43 16.7788C34.775 16.7788 34.7319 17.0825 34.7188 17.2025C34.7057 17.405 34.58 18.065 34.4488 18.635L34.3663 19.0006C34.3025 19.2725 34.235 19.535 34.175 19.8031L34.2013 19.8594C34.9063 19.82 35.1125 19.82 35.7144 19.8294L35.7894 19.775C35.8944 19.145 35.9225 18.9781 36.1213 18.0613L36.215 17.6394C36.4025 16.8144 36.5 16.3981 36.3594 16.0569C36.2057 15.6763 35.8419 15.5825 35.5157 15.5825C35.0324 15.59 34.5631 15.7458 34.1713 16.0288L34.28 15.6088L34.2219 15.5394C33.4682 15.6913 33.2975 15.7119 32.6019 15.8094L32.5494 15.8469C32.4482 16.6851 32.2973 17.5166 32.0975 18.3369C31.9832 18.8225 31.865 19.3119 31.7469 19.7956L31.7788 19.8575C32.4913 19.82 32.705 19.82 33.3238 19.8294ZM38.5007 19.8575C38.5457 19.6419 38.8082 18.3575 38.8119 18.3575C38.8119 18.3575 39.0369 17.42 39.0481 17.3844C39.0481 17.3844 39.1194 17.2869 39.1907 17.2475H39.2938C40.2725 17.2475 41.3788 17.2475 42.245 16.61C42.8481 16.1492 43.2643 15.486 43.4169 14.7425C43.4638 14.5513 43.4994 14.3225 43.4994 14.0938C43.4994 13.7938 43.4375 13.4938 43.2632 13.2631C42.8225 12.6444 41.9432 12.6331 40.9269 12.6294L40.4263 12.6331C39.125 12.65 38.6038 12.6444 38.39 12.6181L38.3375 12.8806L37.8725 15.0444L36.7044 19.85C37.8407 19.835 38.3057 19.835 38.5007 19.8575ZM39.3632 16.0231L39.8582 13.8763L39.8732 13.7656L39.8807 13.6813L40.0775 13.7019L41.1238 13.7919C41.5269 13.9475 41.6919 14.3506 41.5775 14.8756C41.5297 15.1045 41.4315 15.3197 41.2901 15.5059C41.1487 15.6921 40.9676 15.8444 40.76 15.9519C40.4319 16.1169 40.0288 16.13 39.6125 16.13H39.3444L39.3632 16.0231ZM42.4475 17.8794C42.3163 18.4381 42.1663 19.4563 43.0982 19.82C43.3963 19.9456 43.6625 19.985 43.9325 19.97C44.2194 19.955 44.4838 19.8106 44.7294 19.6044L44.6638 19.8594L44.705 19.9156C45.3763 19.8856 45.5844 19.8856 46.3119 19.8913L46.3775 19.8406C46.4825 19.2181 46.5838 18.6125 46.8594 17.4181C46.9925 16.8463 47.1275 16.2819 47.2644 15.7119L47.2438 15.65C46.4938 15.7888 46.2932 15.8188 45.5713 15.92L45.515 15.965L45.4944 16.1338C45.3731 15.9348 45.1881 15.7828 44.9694 15.7025C44.6469 15.575 43.8932 15.7381 43.2444 16.3288C42.7907 16.7488 42.5713 17.3263 42.4475 17.8794ZM44.0225 17.9131C44.1838 17.2269 44.3694 16.6438 44.8475 16.6438C45.1475 16.6438 45.3088 16.925 45.275 17.3994C45.2507 17.5175 45.2225 17.6431 45.1888 17.7838C45.1419 17.99 45.0894 18.1906 45.0388 18.3931C44.9969 18.5195 44.9375 18.6394 44.8625 18.7494C44.7388 18.9256 44.4463 19.0325 44.2794 19.0325C44.2307 19.0325 43.9363 19.0325 43.925 18.5975C43.9317 18.3663 43.965 18.1367 44.0244 17.9131H44.0225ZM52.2538 15.6444L52.1938 15.5769C51.4513 15.7269 51.3181 15.7513 50.6356 15.8431L50.585 15.8938L50.5775 15.9256L50.5757 15.9144C50.0675 17.0863 50.0825 16.8331 49.6681 17.7556L49.6644 17.6431L49.5613 15.6444L49.4956 15.5769C48.7175 15.7269 48.6988 15.7513 47.9806 15.8431L47.9244 15.8938C47.9169 15.9181 47.9169 15.9444 47.9131 15.9725L47.9169 15.9838C48.0069 16.4413 47.9863 16.34 48.0763 17.0638C48.1175 17.42 48.1738 17.7763 48.215 18.1269C48.2863 18.7138 48.3275 19.0044 48.4119 19.9006C47.9281 20.6994 47.8119 21.0031 47.345 21.7063L47.3488 21.7119L47.0207 22.2331C46.9832 22.2875 46.9494 22.325 46.9007 22.34C46.8321 22.3668 46.7584 22.3777 46.685 22.3719H46.5032L46.2332 23.2719L47.1613 23.2888C47.705 23.285 48.05 23.03 48.2338 22.6888L48.8188 21.6875H48.8094L48.8694 21.6163C49.2632 20.7725 52.2519 15.6444 52.2519 15.6444H52.2538ZM42.4475 27.4719H42.0538L43.5125 22.6475H43.9963L44.15 22.1506L44.165 22.7038C44.1463 23.045 44.4144 23.3488 45.1213 23.2981H45.9369L46.2182 22.37H45.9107C45.7344 22.37 45.6519 22.325 45.6631 22.2294L45.6481 21.6669H44.1369V21.6706C43.6475 21.68 42.1869 21.7175 41.8925 21.7963C41.6302 21.878 41.3822 22 41.1575 22.1581L41.3057 21.6613H39.8919L39.5975 22.6475L38.1182 27.5469H37.8313L37.55 28.4694H40.3682L40.2744 28.7769H41.6619L41.7538 28.4694H42.1419L42.4475 27.4719ZM41.2925 23.6281C41.0719 23.701 40.8554 23.7855 40.6438 23.8813L41.0188 22.6475H42.1438L41.8719 23.5475C41.8719 23.5475 41.525 23.5663 41.2907 23.6281H41.2925ZM41.3132 25.3906C41.3132 25.3906 40.9607 25.4338 40.7282 25.4863C40.4994 25.5556 40.0719 25.775 40.0719 25.775L40.4582 24.4925H41.5869L41.3132 25.3906ZM40.6832 27.4831H39.5582L39.8844 26.4013H41.0075L40.6832 27.4831ZM43.3981 24.4925H45.0238L44.7894 25.2481H43.1431L42.8957 26.0731H44.3375L43.2463 27.6106C43.1713 27.7231 43.1019 27.7625 43.025 27.7944C42.95 27.8319 42.8488 27.8788 42.7325 27.8788H42.3332L42.0575 28.7844H43.1038C43.6475 28.7844 43.9663 28.5369 44.2044 28.2125L44.9525 27.1888L45.1138 28.2275C45.1272 28.3036 45.1585 28.3754 45.2051 28.437C45.2517 28.4986 45.3123 28.5483 45.3819 28.5819C45.4869 28.6344 45.5938 28.7244 45.7475 28.7375C45.9107 28.745 46.0288 28.7506 46.1075 28.7506H46.6213L46.9307 27.7381H46.7282C46.6119 27.7381 46.4094 27.7175 46.3775 27.6819C46.3438 27.6369 46.3438 27.5694 46.325 27.4644L46.1619 26.4219H45.4944L45.7869 26.075H47.4294L47.6825 25.2481H46.1619L46.3981 24.4925H47.9169L48.1981 23.5588H43.6757L43.3981 24.4925ZM29.6769 27.695L30.0557 26.4331H31.6138L31.8988 25.4956H30.3388L30.5769 24.7175H32.1013L32.3825 23.81H28.5688L28.2932 24.7175H29.1594L28.9288 25.4938H28.0588L27.7719 26.45H28.6382L28.1319 28.1188C28.0644 28.3381 28.1638 28.4225 28.2275 28.5256C28.2932 28.625 28.3588 28.6925 28.5069 28.73C28.6374 28.764 28.7715 28.7823 28.9063 28.7844H30.6632L30.9763 27.7438L30.1982 27.8525C30.0482 27.8525 29.63 27.8338 29.6769 27.695ZM29.855 21.6556L29.4613 22.3681C29.4086 22.4846 29.3296 22.5873 29.2307 22.6681C29.1707 22.7056 29.0507 22.7206 28.8782 22.7206H28.6719L28.3963 23.6338H29.0807C29.4107 23.6338 29.6619 23.5119 29.7838 23.4519C29.9132 23.3825 29.9469 23.4219 30.0463 23.3263L30.2788 23.1256H32.4163L32.6975 22.175H31.1357L31.4075 21.6556H29.855ZM33.0107 27.7119C32.9732 27.6613 32.9994 27.5675 33.0557 27.3744L33.6388 25.4431H35.7182C36.02 25.4375 36.2394 25.4338 36.3819 25.4244C36.5338 25.4075 36.7007 25.3531 36.8807 25.2556C37.0298 25.1784 37.1554 25.0624 37.2444 24.92C37.4057 24.6699 37.5266 24.3959 37.6025 24.1081L38.3375 21.6613L36.1813 21.6744C36.1813 21.6744 35.5175 21.7719 35.225 21.8806C34.9288 22.0006 34.5088 22.3381 34.5088 22.3381L34.7038 21.6669H33.3725L31.505 27.8506C31.453 28.0207 31.4129 28.1943 31.385 28.37C31.3813 28.4825 31.5275 28.595 31.6213 28.6794C31.7338 28.7638 31.8988 28.7506 32.0563 28.7638C32.225 28.7769 32.4594 28.7825 32.7875 28.7825H33.8113L34.1244 27.7231L33.2075 27.8094C33.11 27.8094 33.0388 27.7569 33.0107 27.7119ZM34.0157 24.1363H36.1982L36.0594 24.5713C36.0407 24.5806 35.9938 24.5488 35.7725 24.575H33.8825L34.0157 24.1363ZM34.4544 22.6775H36.6538L36.4963 23.2025C36.4963 23.2025 35.4594 23.1913 35.2925 23.2213C34.5613 23.3488 34.1357 23.7388 34.1357 23.7388L34.4544 22.6775ZM36.11 26.0281C36.0913 26.0919 36.0632 26.1331 36.0219 26.1613C35.9788 26.1913 35.9057 26.2006 35.7988 26.2006H35.4894L35.5082 25.6719H34.2144L34.16 28.2594C34.16 28.4469 34.1769 28.5556 34.3138 28.6419C34.4507 28.7506 34.8707 28.7638 35.4388 28.7638H36.2488L36.5413 27.7944L35.8363 27.8319L35.6019 27.8469C35.57 27.8338 35.5382 27.8206 35.5044 27.7869C35.4744 27.7569 35.4257 27.7756 35.4332 27.5881L35.4388 26.9244L36.1794 26.8944C36.5788 26.8944 36.7494 26.7631 36.8938 26.6394C37.0344 26.5213 37.0794 26.3844 37.1319 26.2006L37.2557 25.6138H36.2394L36.1082 26.0263L36.11 26.0281Z"
-				fill="#FEFEFE"/>
-		</g>
-		<defs>
-			<clipPath id="clip0_17_2784">
-				<rect width="48" height="30" fill="white" transform="translate(8 5)"/>
-			</clipPath>
-		</defs>
-	</svg>
-	/* eslint-enable */
-);
