+++ packages/js/onboarding/src/components/WCPayConnectCard/index.ts
@@ -1,3 +1,2 @@
 export * from './WCPayConnectCard';
-export * from './Icons';
 export * from './PaymentMethodsIcons';
