+++ packages/js/onboarding/src/images/cards/applepay.js
@@ -3,20 +3,13 @@
  */
 import { createElement } from '@wordpress/element';
 
-export default () => (
-	<svg
-		width="64"
-		height="40"
-		viewBox="0 0 64 40"
-		fill="none"
-		xmlns="http://www.w3.org/2000/svg"
-	>
-		<rect width="64" height="40" fill="white" />
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M18.92 13.32C18.42 13.92 17.6 14.41 16.79 14.34C16.69 13.52 17.09 12.64 17.55 12.1C18.05 11.48 18.95 11.03 19.67 11C19.75 11.86 19.42 12.7 18.92 13.32ZM19.66 14.5C19.0027 14.5012 18.3538 14.6481 17.76 14.93C17.4955 15.0585 17.2118 15.143 16.92 15.18C16.67 15.18 16.37 15.06 16.03 14.92C15.59 14.75 15.09 14.54 14.56 14.55C13.9502 14.5664 13.3555 14.7433 12.8358 15.0628C12.3161 15.3823 11.8899 15.8332 11.6 16.37C10.33 18.57 11.27 21.85 12.5 23.65C13.1 24.54 13.82 25.51 14.77 25.48C15.19 25.46 15.49 25.34 15.8 25.2C16.16 25.05 16.53 24.89 17.11 24.89C17.67 24.89 18.03 25.04 18.37 25.19C18.69 25.34 19 25.47 19.47 25.46C20.45 25.45 21.07 24.57 21.67 23.68C22.32 22.73 22.6 21.79 22.64 21.65V21.64L22.63 21.62C22.41 21.52 20.75 20.76 20.73 18.7C20.7357 18.1518 20.8788 17.6138 21.146 17.1351C21.4133 16.6565 21.7963 16.2525 22.26 15.96L22.29 15.94C21.9942 15.5116 21.6023 15.1583 21.1457 14.9083C20.689 14.6583 20.1803 14.5185 19.66 14.5ZM26.46 25.37V12.02H31.42C32.0003 11.994 32.5796 12.0899 33.1205 12.3013C33.6615 12.5128 34.1522 12.8352 34.5611 13.2478C34.9699 13.6604 35.2878 14.154 35.4943 14.697C35.7008 15.2399 35.7913 15.82 35.76 16.4C35.76 19 33.94 20.8 31.35 20.8H28.5V25.37H26.45H26.46ZM28.51 13.77H30.87C32.65 13.77 33.67 14.73 33.67 16.41C33.67 18.1 32.65 19.07 30.87 19.07H28.5V13.77H28.51ZM42.53 23.77C42.2261 24.3011 41.7838 24.7397 41.2502 25.0391C40.7165 25.3385 40.1116 25.4874 39.5 25.47C37.6 25.47 36.27 24.32 36.27 22.59C36.27 20.89 37.56 19.89 39.94 19.75L42.49 19.6V18.86C42.49 17.77 41.79 17.18 40.54 17.18C39.5 17.18 38.75 17.72 38.6 18.54H36.75C36.81 16.81 38.42 15.56 40.6 15.56C42.94 15.56 44.47 16.8 44.47 18.72V25.37H42.57V23.77H42.53ZM40.05 23.89C38.96 23.89 38.27 23.36 38.27 22.54C38.27 21.7 38.94 21.22 40.22 21.14L42.49 20.99V21.75C42.49 23 41.44 23.89 40.05 23.89ZM50.76 25.89C49.94 28.23 49 29 47 29C46.7392 29.0015 46.4785 28.9848 46.22 28.95V27.35L46.82 27.38C47.72 27.38 48.24 27 48.55 25.99L48.74 25.4L45.26 15.68H47.41L49.82 23.58H49.86L52.28 15.68H54.36L50.76 25.89Z"
-			fill="black"
-		/>
+export const ApplePay = () => (
+	/* eslint-disable */
+	<svg viewBox="0 0 64 40" fill="none" xmlns="http://www.w3.org/2000/svg">
+		<rect width="64" height="40" fill="white"/>
+		<path fillRule="evenodd" clipRule="evenodd"
+			  d="M18.92 13.32C18.42 13.92 17.6 14.41 16.79 14.34C16.69 13.52 17.09 12.64 17.55 12.1C18.05 11.48 18.95 11.03 19.67 11C19.75 11.86 19.42 12.7 18.92 13.32ZM19.66 14.5C19.0027 14.5012 18.3538 14.6481 17.76 14.93C17.4955 15.0585 17.2118 15.143 16.92 15.18C16.67 15.18 16.37 15.06 16.03 14.92C15.59 14.75 15.09 14.54 14.56 14.55C13.9502 14.5664 13.3555 14.7433 12.8358 15.0628C12.3161 15.3823 11.8899 15.8332 11.6 16.37C10.33 18.57 11.27 21.85 12.5 23.65C13.1 24.54 13.82 25.51 14.77 25.48C15.19 25.46 15.49 25.34 15.8 25.2C16.16 25.05 16.53 24.89 17.11 24.89C17.67 24.89 18.03 25.04 18.37 25.19C18.69 25.34 19 25.47 19.47 25.46C20.45 25.45 21.07 24.57 21.67 23.68C22.32 22.73 22.6 21.79 22.64 21.65V21.64L22.63 21.62C22.41 21.52 20.75 20.76 20.73 18.7C20.7357 18.1518 20.8788 17.6138 21.146 17.1351C21.4133 16.6565 21.7963 16.2525 22.26 15.96L22.29 15.94C21.9942 15.5116 21.6023 15.1583 21.1457 14.9083C20.689 14.6583 20.1803 14.5185 19.66 14.5ZM26.46 25.37V12.02H31.42C32.0003 11.994 32.5796 12.0899 33.1205 12.3013C33.6615 12.5128 34.1522 12.8352 34.5611 13.2478C34.9699 13.6604 35.2878 14.154 35.4943 14.697C35.7008 15.2399 35.7913 15.82 35.76 16.4C35.76 19 33.94 20.8 31.35 20.8H28.5V25.37H26.45H26.46ZM28.51 13.77H30.87C32.65 13.77 33.67 14.73 33.67 16.41C33.67 18.1 32.65 19.07 30.87 19.07H28.5V13.77H28.51ZM42.53 23.77C42.2261 24.3011 41.7838 24.7397 41.2502 25.0391C40.7165 25.3385 40.1116 25.4874 39.5 25.47C37.6 25.47 36.27 24.32 36.27 22.59C36.27 20.89 37.56 19.89 39.94 19.75L42.49 19.6V18.86C42.49 17.77 41.79 17.18 40.54 17.18C39.5 17.18 38.75 17.72 38.6 18.54H36.75C36.81 16.81 38.42 15.56 40.6 15.56C42.94 15.56 44.47 16.8 44.47 18.72V25.37H42.57V23.77H42.53ZM40.05 23.89C38.96 23.89 38.27 23.36 38.27 22.54C38.27 21.7 38.94 21.22 40.22 21.14L42.49 20.99V21.75C42.49 23 41.44 23.89 40.05 23.89ZM50.76 25.89C49.94 28.23 49 29 47 29C46.7392 29.0015 46.4785 28.9848 46.22 28.95V27.35L46.82 27.38C47.72 27.38 48.24 27 48.55 25.99L48.74 25.4L45.26 15.68H47.41L49.82 23.58H49.86L52.28 15.68H54.36L50.76 25.89Z"
+			  fill="black"/>
 	</svg>
+	/* eslint-enable */
 );
