+++ packages/js/onboarding/src/images/cards/cb.js
@@ -3,42 +3,22 @@
  */
 import { createElement } from '@wordpress/element';
 
-export default () => (
-	<svg
-		width="52"
-		height="35"
-		viewBox="0 0 52 35"
-		fill="none"
-		xmlns="http://www.w3.org/2000/svg"
-	>
-		<rect
-			x="1.18945"
-			y="0.5"
-			width="49.6897"
-			height="34"
-			rx="2.5"
-			fill="url(#paint0_linear)"
-			stroke="#F1F1F1"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M19.7636 16.9816H26.6269C26.5657 15.4963 26.2201 13.9649 25.1717 12.9813C23.9229 11.8096 21.7355 11.375 19.781 11.375C17.7466 11.375 15.4968 11.8517 14.2414 13.1088C13.1588 14.1918 12.9248 15.9341 12.9248 17.4997C12.9248 19.1395 13.3827 21.0469 14.5571 22.1456C15.8059 23.3147 17.8294 23.625 19.781 23.625C21.6767 23.625 23.7302 23.2746 24.9718 22.1647C26.2099 21.0561 26.6377 19.1888 26.6377 17.4997V17.4918H19.7636V16.9816ZM27.0876 17.4921V23.3511H36.6352V23.3432C38.0322 23.267 39.1436 22.0059 39.1436 20.4575C39.1436 18.9084 38.0322 17.5664 36.6352 17.4895V17.4921H27.0876ZM36.5263 11.6203C37.8879 11.6203 38.9687 12.8032 38.9687 14.2957C38.9687 15.7087 37.9762 16.8626 36.7135 16.9816H27.0873V11.6118H36.2251C36.2813 11.6049 36.3468 11.6097 36.4108 11.6144C36.4508 11.6174 36.4901 11.6203 36.5263 11.6203Z"
-			fill="#FEFEFE"
-		/>
+export const CB = () => (
+	/* eslint-disable */
+	<svg viewBox="0 0 39 25" fill="none" xmlns="http://www.w3.org/2000/svg">
+		<rect x="1.09985" y="0.723633" width="37.4" height="23" rx="1.5" fill="url(#paint0_linear_685_110690)"/>
+		<rect x="1.09985" y="0.723633" width="37.4" height="23" rx="1.5" stroke="#DDDDDD"/>
+		<path fillRule="evenodd" clipRule="evenodd"
+			  d="M13.307 11.7158H20.0383C19.9783 10.2608 19.6393 8.76069 18.6111 7.79714C17.3863 6.64937 15.241 6.22363 13.3241 6.22363C11.3289 6.22363 9.1224 6.6906 7.89114 7.92209C6.82939 8.98291 6.59985 10.6897 6.59985 12.2233C6.59985 13.8297 7.04892 15.6982 8.20072 16.7744C9.4255 17.9196 11.4101 18.2236 13.3241 18.2236C15.1833 18.2236 17.1974 17.8803 18.4151 16.7931C19.6293 15.7072 20.0489 13.878 20.0489 12.2233V12.2156H13.307V11.7158ZM20.4897 12.2158V17.9553H29.8536V17.9476C31.2238 17.8729 32.3138 16.6375 32.3138 15.1207C32.3138 13.6032 31.2238 12.2886 29.8536 12.2133V12.2158H20.4897ZM29.7471 6.46388C31.0825 6.46388 32.1425 7.62259 32.1425 9.08467C32.1425 10.4688 31.169 11.5992 29.9307 11.7158H20.4897V6.45551H29.4516C29.5068 6.44876 29.571 6.45351 29.6339 6.45814C29.673 6.46103 29.7116 6.46388 29.7471 6.46388Z"
+			  fill="#FEFEFE"/>
 		<defs>
-			<linearGradient
-				id="paint0_linear"
-				x1="14.4385"
-				y1="-4.43215"
-				x2="2.09335"
-				y2="33.4202"
-				gradientUnits="userSpaceOnUse"
-			>
-				<stop stopColor="#222E72" />
-				<stop offset="0.591647" stopColor="#40CBFF" />
-				<stop offset="1" stopColor="#3CB792" />
+			<linearGradient id="paint0_linear_685_110690" x1="11.0155" y1="-2.81556" x2="3.21751" y2="23.5991"
+							gradientUnits="userSpaceOnUse">
+				<stop stopColor="#222E72"/>
+				<stop offset="0.591647" stopColor="#40CBFF"/>
+				<stop offset="1" stopColor="#3CB792"/>
 			</linearGradient>
 		</defs>
 	</svg>
+	/* eslint-enable */
 );
