+++ packages/js/onboarding/src/images/cards/discover.js
@@ -3,1396 +3,1428 @@
  */
 import { createElement } from '@wordpress/element';
 
-export default () => (
+export const Discover = () => (
+	/* eslint-disable */
 	<svg
 		xmlns="http://www.w3.org/2000/svg"
 		viewBox="0 0 95.69 62.85"
-		height="40"
-		width="61"
 	>
 		<defs>
-			<clipPath id="clip-path">
+			<clipPath id="c">
 				<path
-					fill="none"
 					d="M53.91 31.28a6.56 6.56 0 1 1-6.56-6.55 6.56 6.56 0 0 1 6.56 6.55Z"
+					className="cls-1"
 				/>
 			</clipPath>
-			<clipPath id="clip-path-2">
+			<clipPath id="d">
 				<path
-					fill="none"
 					d="M90.39 2.3H5.3a3 3 0 0 0-3 3v52.25a3 3 0 0 0 3 3h85.09a3 3 0 0 0 3-3V5.3a3 3 0 0 0-3-3Zm-1.48 29.13a28.13 28.13 0 0 1-28.13 28.12H5.3a2 2 0 0 1-2-2V5.3a2 2 0 0 1 2-2h55.48a28.13 28.13 0 0 1 28.13 28.13Z"
+					className="cls-1"
 				/>
 			</clipPath>
-			<clipPath id="clip-path-3">
-				<circle fill="none" cx="47.88" cy="31.43" r="54.04" />
+			<clipPath id="e">
+				<circle cx={47.88} cy={31.43} r={54.04} className="cls-1"/>
 			</clipPath>
 			<linearGradient
-				id="linear-gradient"
-				x1="50.71"
-				y1="36.53"
-				x2="46.17"
-				y2="29.43"
+				id="a"
+				x1={50.71}
+				x2={46.17}
+				y1={36.53}
+				y2={29.43}
 				gradientUnits="userSpaceOnUse"
 			>
-				<stop offset="0" stopColor="#ff9200" />
-				<stop offset=".21" stopColor="#ff8d00" />
-				<stop offset=".45" stopColor="#ff8000" />
-				<stop offset=".62" stopColor="#ff7100" />
-				<stop offset=".82" stopColor="#ff5e00" />
-				<stop offset="1" stopColor="#ff5200" />
+				<stop offset={0} stopColor="#ff9200"/>
+				<stop offset={0.21} stopColor="#ff8d00"/>
+				<stop offset={0.45} stopColor="#ff8000"/>
+				<stop offset={0.62} stopColor="#ff7100"/>
+				<stop offset={0.82} stopColor="#ff5e00"/>
+				<stop offset={1} stopColor="#ff5200"/>
 			</linearGradient>
 			<linearGradient
-				id="linear-gradient-2"
-				x1="49.99"
-				y1="36.44"
-				x2="43.36"
-				y2="23.47"
+				id="b"
+				x1={49.99}
+				x2={43.36}
+				y1={36.44}
+				y2={23.47}
 				gradientUnits="userSpaceOnUse"
 			>
-				<stop offset="0" stopColor="#ff7100" stopOpacity="0" />
-				<stop offset=".03" stopColor="#ff6e00" stopOpacity=".03" />
-				<stop offset=".21" stopColor="#ff5a00" stopOpacity=".25" />
-				<stop offset=".39" stopColor="#ff4900" stopOpacity=".42" />
-				<stop offset=".56" stopColor="#ff3c00" stopOpacity=".56" />
-				<stop offset=".72" stopColor="#f30" stopOpacity=".66" />
-				<stop offset=".86" stopColor="#ff2d00" stopOpacity=".72" />
-				<stop offset=".98" stopColor="#ff2b00" stopOpacity=".74" />
+				<stop offset={0} stopColor="#ff7100" stopOpacity={0}/>
+				<stop offset={0.03} stopColor="#ff6e00" stopOpacity={0.03}/>
+				<stop offset={0.21} stopColor="#ff5a00" stopOpacity={0.25}/>
+				<stop offset={0.39} stopColor="#ff4900" stopOpacity={0.42}/>
+				<stop offset={0.56} stopColor="#ff3c00" stopOpacity={0.56}/>
+				<stop offset={0.72} stopColor="#f30" stopOpacity={0.66}/>
+				<stop offset={0.86} stopColor="#ff2d00" stopOpacity={0.72}/>
+				<stop offset={0.98} stopColor="#ff2b00" stopOpacity={0.74}/>
 			</linearGradient>
 		</defs>
-		<g style={ { isolation: 'isolate' } }>
-			<g id="Mark">
+		<g
+			style={{
+				isolation: "isolate",
+			}}
+		>
+			<path
+				fill="#fff"
+				d="M3.3 5.3v52.25a2 2 0 0 0 2 2h55.48a28.13 28.13 0 0 0 28.13-28.12A28.13 28.13 0 0 0 60.78 3.3H5.3a2 2 0 0 0-2 2Z"
+			/>
+			<path
+				d="M10.06 25.1H6.51v12.38H10a6.45 6.45 0 0 0 4.42-1.43 6.2 6.2 0 0 0 2.28-4.75c0-3.65-2.7-6.2-6.64-6.2Zm2.82 9.3a4.57 4.57 0 0 1-3.31 1h-.64v-8.2h.64a4.5 4.5 0 0 1 3.31 1 4.21 4.21 0 0 1 1.34 3.08 4.26 4.26 0 0 1-1.34 3.12Zm4.94-9.3h2.41v12.38h-2.41zm8.31 4.75c-1.45-.53-1.87-.89-1.87-1.55s.76-1.38 1.8-1.38a2.47 2.47 0 0 1 1.94 1l1.27-1.66a5.41 5.41 0 0 0-3.64-1.37 3.64 3.64 0 0 0-3.86 3.55c0 1.7.78 2.58 3.05 3.39a9.05 9.05 0 0 1 1.66.71 1.46 1.46 0 0 1 .73 1.28 1.76 1.76 0 0 1-1.88 1.74 2.84 2.84 0 0 1-2.63-1.65l-1.56 1.5a4.8 4.8 0 0 0 4.28 2.36 4 4 0 0 0 4.27-4.06c0-1.96-.81-2.85-3.56-3.86Zm4.33 1.45A6.4 6.4 0 0 0 37 37.76a6.65 6.65 0 0 0 3-.76v-2.8a3.88 3.88 0 0 1-2.92 1.35 4 4 0 0 1-4.16-4.27A4.12 4.12 0 0 1 37 27a4 4 0 0 1 3 1.4v-2.81a6.15 6.15 0 0 0-3-.77 6.51 6.51 0 0 0-6.54 6.48Zm28.7 2.12-3.3-8.32h-2.64l5.26 12.7h1.3l5.34-12.7h-2.61l-3.35 8.32zm7.06 4.06h6.84v-2.09h-4.43v-3.34h4.27v-2.1h-4.27V27.2h4.43v-2.1h-6.84v12.38zm16.41-8.72c0-2.32-1.6-3.66-4.39-3.66h-3.58v12.38h2.41v-5h.32l3.34 5h3l-3.9-5.21a3.3 3.3 0 0 0 2.8-3.51Zm-4.85 2h-.71v-3.71h.75c1.5 0 2.32.63 2.32 1.84s-.82 1.91-2.36 1.91Zm6.77-5.01c0-.21-.15-.33-.41-.33h-.35v1.09h.26v-.42l.3.42h.32l-.36-.45a.3.3 0 0 0 .24-.31Zm-.46.15v-.28c.13 0 .2 0 .2.14s-.07.14-.2.14Z"
+				className="cls-4"
+			/>
+			<path
+				d="M84.19 25a1 1 0 0 0 0 1.9 1 1 0 0 0 0-1.9Zm0 1.73a.78.78 0 0 1 0-1.56.78.78 0 0 1 0 1.56Z"
+				className="cls-4"
+			/>
+			<path
+				fill="url(#a)"
+				d="M53.91 31.28a6.56 6.56 0 1 1-6.55-6.55 6.56 6.56 0 0 1 6.55 6.55Z"
+			/>
+			<path
+				fill="url(#b)"
+				d="M53.91 31.28a6.56 6.56 0 1 1-6.55-6.55 6.56 6.56 0 0 1 6.55 6.55Z"
+				opacity={0.65}
+				style={{
+					mixBlendMode: "multiply",
+				}}
+			/>
+			<g
+				clipPath="url(#c)"
+				opacity={0.5}
+				style={{
+					mixBlendMode: "multiply",
+				}}
+			>
 				<path
-					d="M3.3 5.3v52.25a2 2 0 0 0 2 2h55.48a28.13 28.13 0 0 0 28.13-28.12A28.13 28.13 0 0 0 60.78 3.3H5.3a2 2 0 0 0-2 2Z"
-					style={ { fill: '#fff' } }
+					fill="#020000"
+					d="M47.38 24.63a6.69 6.69 0 1 1-6.68 6.68 6.67 6.67 0 0 1 6.68-6.68Z"
 				/>
 				<path
-					fill="#261f20"
-					d="M10.06 25.1H6.51v12.38H10a6.45 6.45 0 0 0 4.42-1.43 6.2 6.2 0 0 0 2.28-4.75c0-3.65-2.7-6.2-6.64-6.2Zm2.82 9.3a4.57 4.57 0 0 1-3.31 1h-.64v-8.2h.64a4.5 4.5 0 0 1 3.31 1 4.21 4.21 0 0 1 1.34 3.08 4.26 4.26 0 0 1-1.34 3.12ZM17.82 25.1h2.41v12.38h-2.41zM26.13 29.85c-1.45-.53-1.87-.89-1.87-1.55s.76-1.38 1.8-1.38a2.47 2.47 0 0 1 1.94 1l1.27-1.66a5.41 5.41 0 0 0-3.64-1.37 3.64 3.64 0 0 0-3.86 3.55c0 1.7.78 2.58 3.05 3.39a9.05 9.05 0 0 1 1.66.71 1.46 1.46 0 0 1 .73 1.28 1.76 1.76 0 0 1-1.88 1.74 2.84 2.84 0 0 1-2.63-1.65l-1.56 1.5a4.8 4.8 0 0 0 4.28 2.36 4 4 0 0 0 4.27-4.06c0-1.96-.81-2.85-3.56-3.86ZM30.46 31.3A6.4 6.4 0 0 0 37 37.76a6.65 6.65 0 0 0 3-.76v-2.8a3.88 3.88 0 0 1-2.92 1.35 4 4 0 0 1-4.16-4.27A4.12 4.12 0 0 1 37 27a4 4 0 0 1 3 1.4v-2.81a6.15 6.15 0 0 0-3-.77 6.51 6.51 0 0 0-6.54 6.48ZM59.16 33.42l-3.3-8.32h-2.64l5.26 12.7h1.3l5.34-12.7h-2.61l-3.35 8.32zM66.22 37.48h6.84v-2.09h-4.43v-3.34h4.27v-2.1h-4.27V27.2h4.43v-2.1h-6.84v12.38zM82.63 28.76c0-2.32-1.6-3.66-4.39-3.66h-3.58v12.38h2.41v-5h.32l3.34 5h3l-3.9-5.21a3.3 3.3 0 0 0 2.8-3.51Zm-4.85 2h-.71v-3.71h.75c1.5 0 2.32.63 2.32 1.84s-.82 1.91-2.36 1.91ZM84.55 25.75c0-.21-.15-.33-.41-.33h-.35v1.09h.26v-.42l.3.42h.32l-.36-.45a.3.3 0 0 0 .24-.31Zm-.46.15v-.28c.13 0 .2 0 .2.14s-.07.14-.2.14Z"
+					fill="#030100"
+					d="M47.39 24.64a6.69 6.69 0 1 1-6.68 6.68 6.68 6.68 0 0 1 6.68-6.68Z"
 				/>
 				<path
-					fill="#261f20"
-					d="M84.19 25a1 1 0 0 0 0 1.9 1 1 0 0 0 0-1.9Zm0 1.73a.78.78 0 0 1 0-1.56.78.78 0 0 1 0 1.56Z"
+					fill="#050100"
+					d="M47.4 24.65a6.69 6.69 0 1 1-6.69 6.68 6.67 6.67 0 0 1 6.69-6.68Z"
 				/>
 				<path
-					d="M53.91 31.28a6.56 6.56 0 1 1-6.55-6.55 6.56 6.56 0 0 1 6.55 6.55Z"
-					style={ { fill: `url(#linear-gradient)` } }
+					fill="#070200"
+					d="M47.4 24.66a6.68 6.68 0 1 1-6.68 6.68 6.68 6.68 0 0 1 6.68-6.68Z"
 				/>
 				<path
-					d="M53.91 31.28a6.56 6.56 0 1 1-6.55-6.55 6.56 6.56 0 0 1 6.55 6.55Z"
-					style={ {
-						opacity: 0.65,
-						fill: `url(#linear-gradient-2)`,
-						mixBlendMode: 'multiply',
-					} }
+					fill="#080200"
+					d="M47.41 24.67a6.68 6.68 0 1 1-6.68 6.68 6.68 6.68 0 0 1 6.68-6.68Z"
 				/>
-				<g style={ { opacity: 0.5, mixBlendMode: 'multiply' } }>
-					<g style={ { clipPath: `url('#clip-path')` } }>
-						<path
-							d="M47.38 24.63a6.69 6.69 0 1 1-6.68 6.68 6.67 6.67 0 0 1 6.68-6.68Z"
-							style={ { fill: '#020000' } }
-						/>
-						<path
-							d="M47.39 24.64a6.69 6.69 0 1 1-6.68 6.68 6.68 6.68 0 0 1 6.68-6.68Z"
-							style={ { fill: '#030100' } }
-						/>
-						<path
-							d="M47.4 24.65a6.69 6.69 0 1 1-6.69 6.68 6.67 6.67 0 0 1 6.69-6.68Z"
-							style={ { fill: '#050100' } }
-						/>
-						<path
-							d="M47.4 24.66a6.68 6.68 0 1 1-6.68 6.68 6.68 6.68 0 0 1 6.68-6.68Z"
-							style={ { fill: '#070200' } }
-						/>
-						<path
-							d="M47.41 24.67a6.68 6.68 0 1 1-6.68 6.68 6.68 6.68 0 0 1 6.68-6.68Z"
-							style={ { fill: '#080200' } }
-						/>
-						<path
-							d="M47.42 24.68a6.68 6.68 0 1 1-6.68 6.68 6.68 6.68 0 0 1 6.68-6.68Z"
-							style={ { fill: '#0a0300' } }
-						/>
-						<path
-							d="M47.43 24.69a6.68 6.68 0 1 1-6.68 6.68 6.67 6.67 0 0 1 6.68-6.68Z"
-							style={ { fill: '#0c0300' } }
-						/>
-						<path
-							d="M47.43 24.7a6.68 6.68 0 0 1 6.71 6.64 6.75 6.75 0 0 1-6.71 6.72 6.68 6.68 0 0 1 0-13.36Z"
-							style={ { fill: '#0e0400' } }
-						/>
-						<path
-							d="M47.44 24.71a6.67 6.67 0 0 1 6.71 6.63 6.75 6.75 0 0 1-6.71 6.73 6.68 6.68 0 0 1 0-13.36Z"
-							style={ { fill: '#0f0400' } }
-						/>
-						<path
-							d="M47.45 24.73a6.67 6.67 0 0 1 6.71 6.62 6.76 6.76 0 0 1-6.71 6.73 6.68 6.68 0 1 1 0-13.35Z"
-							style={ { fill: '#110500' } }
-						/>
-						<path
-							d="M47.45 24.74a6.66 6.66 0 0 1 6.72 6.61 6.76 6.76 0 0 1-6.72 6.74 6.68 6.68 0 0 1 0-13.35Z"
-							style={ { fill: '#130500' } }
-						/>
-						<path
-							d="M47.46 24.75a6.66 6.66 0 0 1 6.72 6.6 6.76 6.76 0 0 1-6.72 6.74 6.67 6.67 0 1 1 0-13.34Z"
-							style={ { fill: '#140600' } }
-						/>
-						<path
-							d="M47.47 24.76a6.66 6.66 0 0 1 6.72 6.6 6.78 6.78 0 0 1-6.72 6.74 6.67 6.67 0 0 1 0-13.34Z"
-							style={ { fill: '#160600' } }
-						/>
-						<path
-							d="M47.48 24.77a6.66 6.66 0 0 1 6.72 6.59 6.78 6.78 0 0 1-6.72 6.75 6.67 6.67 0 0 1 0-13.34Z"
-							style={ { fill: '#180600' } }
-						/>
-						<path
-							d="M47.48 24.78a6.66 6.66 0 0 1 6.73 6.59 6.79 6.79 0 0 1-6.73 6.75 6.67 6.67 0 1 1 0-13.34Z"
-							style={ { fill: '#190700' } }
-						/>
-						<path
-							d="M47.49 24.79a6.66 6.66 0 0 1 6.73 6.58 6.8 6.8 0 0 1-6.73 6.76 6.67 6.67 0 0 1 0-13.34Z"
-							style={ { fill: '#1b0700' } }
-						/>
-						<path
-							d="M47.5 24.8a6.65 6.65 0 0 1 6.73 6.57 6.81 6.81 0 0 1-6.73 6.77 6.67 6.67 0 0 1 0-13.34Z"
-							style={ { fill: '#1d0800' } }
-						/>
-						<path
-							d="M47.5 24.81a6.66 6.66 0 0 1 6.74 6.57 6.81 6.81 0 0 1-6.74 6.77 6.67 6.67 0 0 1 0-13.34Z"
-							style={ { fill: '#1f0800' } }
-						/>
-						<path
-							d="M47.51 24.82a6.66 6.66 0 0 1 6.74 6.56 6.82 6.82 0 0 1-6.74 6.77 6.67 6.67 0 0 1 0-13.33Z"
-							style={ { fill: '#200900' } }
-						/>
-						<path
-							d="M47.52 24.84a6.65 6.65 0 0 1 6.74 6.55 6.82 6.82 0 0 1-6.74 6.77 6.66 6.66 0 0 1 0-13.32Z"
-							style={ { fill: '#220900' } }
-						/>
-						<path
-							d="M47.53 24.85a6.64 6.64 0 0 1 6.74 6.54 6.83 6.83 0 0 1-6.74 6.78 6.66 6.66 0 1 1 0-13.32Z"
-							style={ { fill: '#240a00' } }
-						/>
-						<path
-							d="M47.53 24.86a6.65 6.65 0 0 1 6.75 6.54 6.84 6.84 0 0 1-6.75 6.78 6.66 6.66 0 0 1 0-13.32Z"
-							style={ { fill: '#250a00' } }
-						/>
-						<path
-							d="M47.54 24.87a6.65 6.65 0 0 1 6.75 6.53 6.85 6.85 0 0 1-6.75 6.79 6.66 6.66 0 1 1 0-13.32Z"
-							style={ { fill: '#270b00' } }
-						/>
-						<path
-							d="M47.55 24.88a6.65 6.65 0 0 1 6.75 6.52 6.86 6.86 0 0 1-6.75 6.8 6.66 6.66 0 1 1 0-13.32Z"
-							style={ { fill: '#290b00' } }
-						/>
-						<path
-							d="M47.55 24.89a6.65 6.65 0 0 1 6.76 6.52 6.86 6.86 0 0 1-6.76 6.8 6.66 6.66 0 0 1 0-13.32Z"
-							style={ { fill: '#2a0c00' } }
-						/>
-						<path
-							d="M47.56 24.9a6.64 6.64 0 0 1 6.76 6.51 6.86 6.86 0 0 1-6.76 6.8 6.66 6.66 0 0 1 0-13.31Z"
-							style={ { fill: '#2c0c00' } }
-						/>
-						<path
-							d="M47.57 24.91a6.64 6.64 0 0 1 6.76 6.51 6.88 6.88 0 0 1-6.76 6.8 6.66 6.66 0 1 1 0-13.31Z"
-							style={ { fill: '#2e0d00' } }
-						/>
-						<path
-							d="M47.58 24.92a6.65 6.65 0 0 1 6.76 6.5 6.88 6.88 0 0 1-6.76 6.81 6.66 6.66 0 1 1 0-13.31Z"
-							style={ { fill: '#300d00' } }
-						/>
-						<path
-							d="M47.58 24.94a6.63 6.63 0 0 1 6.77 6.49 6.89 6.89 0 0 1-6.77 6.81 6.65 6.65 0 1 1 0-13.3Z"
-							style={ { fill: '#310d00' } }
-						/>
-						<path
-							d="M47.59 25a6.63 6.63 0 0 1 6.77 6.48 6.9 6.9 0 0 1-6.77 6.82 6.65 6.65 0 0 1 0-13.3Z"
-							style={ { fill: '#330e00' } }
-						/>
-						<path
-							d="M47.6 25a6.63 6.63 0 0 1 6.77 6.47 6.91 6.91 0 0 1-6.77 6.83 6.65 6.65 0 0 1 0-13.3Z"
-							style={ { fill: '#350e00' } }
-						/>
-						<path
-							d="M47.6 25a6.63 6.63 0 0 1 6.78 6.47 6.91 6.91 0 0 1-6.78 6.83 6.65 6.65 0 0 1 0-13.3Z"
-							style={ { fill: '#360f00' } }
-						/>
-						<path
-							d="M47.61 25a6.63 6.63 0 0 1 6.78 6.46 6.93 6.93 0 0 1-6.78 6.84 6.65 6.65 0 1 1 0-13.3Z"
-							style={ { fill: '#380f00' } }
-						/>
-						<path
-							d="M47.62 25a6.63 6.63 0 0 1 6.78 6.46 6.92 6.92 0 0 1-6.78 6.83 6.65 6.65 0 1 1 0-13.29Z"
-							style={ { fill: '#3a1000' } }
-						/>
-						<path
-							d="M47.63 25a6.62 6.62 0 0 1 6.78 6.45 6.93 6.93 0 0 1-6.78 6.84 6.65 6.65 0 1 1 0-13.29Z"
-							style={ { fill: '#3b1000' } }
-						/>
-						<path
-							d="M47.63 25a6.63 6.63 0 0 1 6.79 6.45 6.94 6.94 0 0 1-6.79 6.84 6.65 6.65 0 0 1 0-13.29Z"
-							style={ { fill: '#3d1100' } }
-						/>
-						<path
-							d="M47.64 25a6.63 6.63 0 0 1 6.79 6.44 7 7 0 0 1-6.79 6.85 6.65 6.65 0 0 1 0-13.29Z"
-							style={ { fill: '#3f1100' } }
-						/>
-						<path
-							d="M47.65 25a6.64 6.64 0 0 1 6.79 6.43 7 7 0 0 1-6.79 6.86 6.65 6.65 0 0 1 0-13.29Z"
-							style={ { fill: '#411200' } }
-						/>
-						<path
-							d="M47.65 25.05a6.62 6.62 0 0 1 6.8 6.42 7 7 0 0 1-6.8 6.86 6.64 6.64 0 0 1 0-13.28Z"
-							style={ { fill: '#421200' } }
-						/>
-						<path
-							d="M47.66 25.06a6.62 6.62 0 0 1 6.8 6.41 7 7 0 0 1-6.8 6.87 6.64 6.64 0 1 1 0-13.28Z"
-							style={ { fill: '#441300' } }
-						/>
-						<path
-							d="M47.67 25.07a6.62 6.62 0 0 1 6.8 6.41 7 7 0 0 1-6.8 6.86 6.64 6.64 0 1 1 0-13.27Z"
-							style={ { fill: '#461300' } }
-						/>
-						<path
-							d="M47.68 25.08a6.62 6.62 0 0 1 6.8 6.4 7 7 0 0 1-6.8 6.87 6.64 6.64 0 1 1 0-13.27Z"
-							style={ { fill: '#471300' } }
-						/>
-						<path
-							d="M47.68 25.09a6.61 6.61 0 0 1 6.81 6.39 7 7 0 0 1-6.81 6.88 6.64 6.64 0 0 1 0-13.27Z"
-							style={ { fill: '#491400' } }
-						/>
-						<path
-							d="M47.69 25.1a6.61 6.61 0 0 1 6.81 6.39 7 7 0 0 1-6.81 6.88 6.64 6.64 0 0 1 0-13.27Z"
-							style={ { fill: '#4b1400' } }
-						/>
-						<path
-							d="M47.7 25.11a6.61 6.61 0 0 1 6.81 6.38 7 7 0 0 1-6.81 6.89 6.64 6.64 0 1 1 0-13.27Z"
-							style={ { fill: '#4c1500' } }
-						/>
-						<path
-							d="M47.7 25.12c3.67.06 6.82 2.71 6.82 6.38a7 7 0 0 1-6.82 6.89 6.64 6.64 0 0 1 0-13.27Z"
-							style={ { fill: '#4e1500' } }
-						/>
-						<path
-							d="M47.71 25.13c3.66.06 6.82 2.71 6.82 6.37a7 7 0 0 1-6.82 6.9 6.64 6.64 0 0 1 0-13.27Z"
-							style={ { fill: '#501600' } }
-						/>
-						<path
-							d="M47.72 25.15c3.66.05 6.82 2.69 6.82 6.36a7 7 0 0 1-6.82 6.9 6.63 6.63 0 0 1 0-13.26Z"
-							style={ { fill: '#521600' } }
-						/>
-						<path
-							d="M47.73 25.16c3.66.05 6.82 2.69 6.82 6.35a7 7 0 0 1-6.82 6.9 6.63 6.63 0 1 1 0-13.25Z"
-							style={ { fill: '#531700' } }
-						/>
-						<path
-							d="M47.73 25.17c3.66 0 6.83 2.68 6.83 6.34a7 7 0 0 1-6.83 6.91 6.63 6.63 0 0 1 0-13.25Z"
-							style={ { fill: '#551700' } }
-						/>
-						<path
-							d="M47.74 25.18c3.66.06 6.83 2.68 6.83 6.34a7 7 0 0 1-6.83 6.91 6.63 6.63 0 1 1 0-13.25Z"
-							style={ { fill: '#571800' } }
-						/>
-						<path
-							d="M47.75 25.19c3.66.06 6.83 2.67 6.83 6.33a7.06 7.06 0 0 1-6.83 6.92 6.63 6.63 0 1 1 0-13.25Z"
-							style={ { fill: '#581800' } }
-						/>
-						<path
-							d="M47.75 25.2c3.66.06 6.84 2.67 6.84 6.33a7.07 7.07 0 0 1-6.84 6.92 6.63 6.63 0 0 1 0-13.25Z"
-							style={ { fill: '#5a1900' } }
-						/>
-						<path
-							d="M47.76 25.21c3.66.06 6.84 2.66 6.84 6.32a7.08 7.08 0 0 1-6.84 6.93 6.63 6.63 0 0 1 0-13.25Z"
-							style={ { fill: '#5c1900' } }
-						/>
-						<path
-							d="M47.77 25.22c3.65.06 6.84 2.66 6.84 6.32a7.08 7.08 0 0 1-6.84 6.93 6.63 6.63 0 0 1 0-13.25Z"
-							style={ { fill: '#5e1a00' } }
-						/>
-						<path
-							d="M47.78 25.23c3.65.07 6.84 2.65 6.84 6.31a7.08 7.08 0 0 1-6.84 6.93 6.62 6.62 0 0 1 0-13.24Z"
-							style={ { fill: '#5f1a00' } }
-						/>
-						<path
-							d="M47.78 25.24c3.66.07 6.85 2.65 6.85 6.3a7.09 7.09 0 0 1-6.85 6.94 6.62 6.62 0 0 1 0-13.24Z"
-							style={ { fill: '#611a00' } }
-						/>
-						<path
-							d="M47.79 25.26c3.65.06 6.85 2.63 6.85 6.29a7.1 7.1 0 0 1-6.85 6.94 6.62 6.62 0 1 1 0-13.23Z"
-							style={ { fill: '#631b00' } }
-						/>
-						<path
-							d="M47.8 25.27c3.65.06 6.85 2.63 6.85 6.28a7.11 7.11 0 0 1-6.85 7 6.62 6.62 0 1 1 0-13.23Z"
-							style={ { fill: '#641b00' } }
-						/>
-						<path
-							d="M47.8 25.28c3.66.07 6.86 2.62 6.86 6.28a7.12 7.12 0 0 1-6.86 7 6.62 6.62 0 0 1 0-13.23Z"
-							style={ { fill: '#661c00' } }
-						/>
-						<path
-							d="M47.81 25.29c3.65.07 6.86 2.62 6.86 6.27a7.14 7.14 0 0 1-6.86 7 6.62 6.62 0 0 1 0-13.23Z"
-							style={ { fill: '#681c00' } }
-						/>
-						<path
-							d="M47.82 25.3c3.65.07 6.86 2.61 6.86 6.27a7.14 7.14 0 0 1-6.86 7 6.62 6.62 0 0 1 0-13.23Z"
-							style={ { fill: '#691d00' } }
-						/>
-						<path
-							d="M47.83 25.31c3.65.07 6.86 2.61 6.86 6.26a7.14 7.14 0 0 1-6.86 7 6.57 6.57 0 0 1-6.62-6.61 6.64 6.64 0 0 1 6.62-6.65Z"
-							style={ { fill: '#6b1d00' } }
-						/>
-						<path
-							d="M47.83 25.32c3.65.07 6.87 2.6 6.87 6.25a7.15 7.15 0 0 1-6.87 7 6.57 6.57 0 0 1-6.61-6.61 6.64 6.64 0 0 1 6.61-6.64Z"
-							style={ { fill: '#6d1e00' } }
-						/>
-						<path
-							d="M47.84 25.33c3.65.08 6.87 2.6 6.87 6.25a7.15 7.15 0 0 1-6.87 7 6.57 6.57 0 0 1-6.61-6.61 6.64 6.64 0 0 1 6.61-6.64Z"
-							style={ { fill: '#6f1e00' } }
-						/>
-						<path
-							d="M47.85 25.34c3.65.08 6.87 2.59 6.87 6.24a7.16 7.16 0 0 1-6.87 7A6.57 6.57 0 0 1 41.24 32a6.64 6.64 0 0 1 6.61-6.66Z"
-							style={ { fill: '#701f00' } }
-						/>
-						<path
-							d="M47.85 25.35c3.65.08 6.88 2.59 6.88 6.24a7.17 7.17 0 0 1-6.88 7 6.56 6.56 0 0 1-6.6-6.59 6.63 6.63 0 0 1 6.6-6.65Z"
-							style={ { fill: '#721f00' } }
-						/>
-						<path
-							d="M47.86 25.37c3.65.07 6.88 2.57 6.88 6.22a7.18 7.18 0 0 1-6.88 7 6.61 6.61 0 0 1 0-13.21Z"
-							style={ { fill: '#742000' } }
-						/>
-						<path
-							d="M47.87 25.38c3.65.08 6.88 2.57 6.88 6.22a7.19 7.19 0 0 1-6.88 7 6.61 6.61 0 1 1 0-13.21Z"
-							style={ { fill: '#752000' } }
-						/>
-						<path
-							d="M47.88 25.39c3.64.08 6.88 2.56 6.88 6.21a7.2 7.2 0 0 1-6.88 7 6.61 6.61 0 1 1 0-13.21Z"
-							style={ { fill: '#772000' } }
-						/>
-						<path
-							d="M47.88 25.4c3.65.08 6.89 2.56 6.89 6.2a7.2 7.2 0 0 1-6.89 7 6.56 6.56 0 0 1-6.6-6.6 6.62 6.62 0 0 1 6.6-6.6Z"
-							style={ { fill: '#792100' } }
-						/>
-						<path
-							d="M47.89 25.41c3.64.08 6.89 2.55 6.89 6.2a7.2 7.2 0 0 1-6.89 7 6.55 6.55 0 0 1-6.6-6.6 6.62 6.62 0 0 1 6.6-6.6Z"
-							style={ { fill: '#7a2100' } }
-						/>
-						<path
-							d="M47.9 25.42c3.64.08 6.89 2.55 6.89 6.19a7.21 7.21 0 0 1-6.89 7A6.54 6.54 0 0 1 41.3 32a6.62 6.62 0 0 1 6.6-6.58Z"
-							style={ { fill: '#7c2200' } }
-						/>
-						<path
-							d="M47.9 25.43c3.65.09 6.89 2.54 6.89 6.19a7.21 7.21 0 0 1-6.89 7A6.55 6.55 0 0 1 41.31 32a6.62 6.62 0 0 1 6.59-6.57Z"
-							style={ { fill: '#7e2200' } }
-						/>
-						<path
-							d="M47.91 25.44c3.64.09 6.89 2.54 6.89 6.18a7.22 7.22 0 0 1-6.89 7 6.56 6.56 0 0 1-6.6-6.6 6.64 6.64 0 0 1 6.6-6.58Z"
-							style={ { fill: '#802300' } }
-						/>
-						<path
-							d="M47.92 25.45c3.64.09 6.89 2.53 6.89 6.17a7.23 7.23 0 0 1-6.89 7 6.56 6.56 0 0 1-6.6-6.6 6.64 6.64 0 0 1 6.6-6.57Z"
-							style={ { fill: '#812300' } }
-						/>
-						<path
-							d="M47.93 25.47c3.64.08 6.89 2.52 6.89 6.16a7.24 7.24 0 0 1-6.89 7 6.55 6.55 0 0 1-6.6-6.6 6.62 6.62 0 0 1 6.6-6.56Z"
-							style={ { fill: '#832400' } }
-						/>
-						<path
-							d="M47.93 25.48c3.64.08 6.9 2.51 6.9 6.15a7.24 7.24 0 0 1-6.9 7 6.54 6.54 0 0 1-6.59-6.59 6.61 6.61 0 0 1 6.59-6.56Z"
-							style={ { fill: '#852400' } }
-						/>
-						<path
-							d="M47.94 25.49c3.64.09 6.9 2.51 6.9 6.15a7.25 7.25 0 0 1-6.9 7 6.54 6.54 0 0 1-6.59-6.59 6.61 6.61 0 0 1 6.59-6.56Z"
-							style={ { fill: '#862500' } }
-						/>
-						<path
-							d="M48 25.5c3.64.09 6.9 2.5 6.9 6.14a7.26 7.26 0 0 1-6.9 7 6.53 6.53 0 0 1-6.59-6.59A6.61 6.61 0 0 1 48 25.5Z"
-							style={ { fill: '#882500' } }
-						/>
-						<path
-							d="M48 25.51c3.64.09 6.91 2.5 6.91 6.14a7.27 7.27 0 0 1-6.91 7 6.54 6.54 0 0 1-6.59-6.59A6.63 6.63 0 0 1 48 25.51Z"
-							style={ { fill: '#8a2600' } }
-						/>
-						<path
-							d="M48 25.52c3.64.09 6.91 2.49 6.91 6.13A7.28 7.28 0 0 1 48 38.7a6.54 6.54 0 0 1-6.59-6.59A6.63 6.63 0 0 1 48 25.52Z"
-							style={ { fill: '#8b2600' } }
-						/>
-						<path
-							d="M48 25.53c3.64.1 6.91 2.49 6.91 6.12A7.28 7.28 0 0 1 48 38.71a6.54 6.54 0 0 1-6.59-6.59A6.63 6.63 0 0 1 48 25.53Z"
-							style={ { fill: '#8d2600' } }
-						/>
-						<path
-							d="M48 25.54c3.63.1 6.91 2.48 6.91 6.12A7.29 7.29 0 0 1 48 38.72a6.53 6.53 0 0 1-6.59-6.59A6.62 6.62 0 0 1 48 25.54Z"
-							style={ { fill: '#8f2700' } }
-						/>
-						<path
-							d="M48 25.55c3.64.1 6.92 2.48 6.92 6.11A7.3 7.3 0 0 1 48 38.73a6.53 6.53 0 0 1-6.58-6.59A6.62 6.62 0 0 1 48 25.55Z"
-							style={ { fill: '#912700' } }
-						/>
-						<path
-							d="M48 25.56c3.63.1 6.92 2.47 6.92 6.11A7.3 7.3 0 0 1 48 38.73a6.54 6.54 0 0 1-6.59-6.58A6.63 6.63 0 0 1 48 25.56Z"
-							style={ { fill: '#922800' } }
-						/>
-						<path
-							d="M48 25.58c3.63.09 6.92 2.46 6.92 6.09A7.31 7.31 0 0 1 48 38.74a6.54 6.54 0 0 1-6.59-6.58A6.62 6.62 0 0 1 48 25.58Z"
-							style={ { fill: '#942800' } }
-						/>
-						<path
-							d="M48 25.59c3.64.1 6.93 2.45 6.93 6.09A7.32 7.32 0 0 1 48 38.75a6.53 6.53 0 0 1-6.58-6.58A6.62 6.62 0 0 1 48 25.59Z"
-							style={ { fill: '#962900' } }
-						/>
-						<path
-							d="M48 25.6c3.63.1 6.93 2.45 6.93 6.08A7.33 7.33 0 0 1 48 38.76a6.53 6.53 0 0 1-6.58-6.58A6.62 6.62 0 0 1 48 25.6Z"
-							style={ { fill: '#972900' } }
-						/>
-						<path
-							d="M48 25.61c3.63.1 6.93 2.44 6.93 6.07A7.34 7.34 0 0 1 48 38.77a6.52 6.52 0 0 1-6.58-6.58A6.62 6.62 0 0 1 48 25.61Z"
-							style={ { fill: '#992a00' } }
-						/>
-						<path
-							d="M48 25.62c3.63.1 6.93 2.43 6.93 6.07A7.35 7.35 0 0 1 48 38.78a6.52 6.52 0 0 1-6.58-6.58A6.62 6.62 0 0 1 48 25.62Z"
-							style={ { fill: '#9b2a00' } }
-						/>
-						<path
-							d="M48 25.63c3.63.11 6.94 2.43 6.94 6.06a7.35 7.35 0 0 1-6.94 7.1 6.53 6.53 0 0 1-6.58-6.58A6.62 6.62 0 0 1 48 25.63Z"
-							style={ { fill: '#9c2b00' } }
-						/>
-						<path
-							d="M48 25.64c3.63.11 6.94 2.42 6.94 6.06A7.35 7.35 0 0 1 48 38.79a6.52 6.52 0 0 1-6.58-6.57A6.62 6.62 0 0 1 48 25.64Z"
-							style={ { fill: '#9e2b00' } }
-						/>
-						<path
-							d="M48.05 25.65c3.63.11 6.94 2.42 6.94 6.05a7.36 7.36 0 0 1-6.94 7.1 6.52 6.52 0 0 1-6.58-6.57 6.62 6.62 0 0 1 6.58-6.58Z"
-							style={ { fill: '#a02c00' } }
-						/>
-						<path
-							d="M48.05 25.66c3.63.11 7 2.41 7 6.05a7.37 7.37 0 0 1-7 7.1 6.51 6.51 0 0 1-6.57-6.57 6.61 6.61 0 0 1 6.57-6.58Z"
-							style={ { fill: '#a22c00' } }
-						/>
-						<path
-							d="M48.06 25.67c3.63.11 6.95 2.41 6.95 6a7.38 7.38 0 0 1-6.95 7.11 6.51 6.51 0 0 1-6.57-6.57 6.61 6.61 0 0 1 6.57-6.54Z"
-							style={ { fill: '#a32d00' } }
-						/>
-						<path
-							d="M48.07 25.69c3.63.11 7 2.39 7 6a7.39 7.39 0 0 1-7 7.12 6.51 6.51 0 0 1-6.57-6.57 6.6 6.6 0 0 1 6.57-6.55Z"
-							style={ { fill: '#a52d00' } }
-						/>
-						<path
-							d="M48.08 25.7c3.62.11 7 2.39 7 6a7.4 7.4 0 0 1-7 7.12 6.53 6.53 0 0 1-6.58-6.57 6.61 6.61 0 0 1 6.58-6.55Z"
-							style={ { fill: '#a72d00' } }
-						/>
-						<path
-							d="M48.08 25.71c3.63.11 7 2.38 7 6a7.41 7.41 0 0 1-7 7.13 6.51 6.51 0 0 1-6.57-6.57 6.6 6.6 0 0 1 6.57-6.56Z"
-							style={ { fill: '#a82e00' } }
-						/>
-						<path
-							d="M48.09 25.72c3.63.11 7 2.38 7 6a7.41 7.41 0 0 1-7 7.12 6.5 6.5 0 0 1-6.57-6.56 6.6 6.6 0 0 1 6.57-6.56Z"
-							style={ { fill: '#aa2e00' } }
-						/>
-						<path
-							d="M48.1 25.73c3.62.11 7 2.37 7 6a7.42 7.42 0 0 1-7 7.13 6.5 6.5 0 0 1-6.57-6.56 6.6 6.6 0 0 1 6.57-6.57Z"
-							style={ { fill: '#ac2f00' } }
-						/>
-						<path
-							d="M48.1 25.74c3.63.12 7 2.37 7 6a7.43 7.43 0 0 1-7 7.14 6.5 6.5 0 0 1-6.56-6.56 6.6 6.6 0 0 1 6.56-6.58Z"
-							style={ { fill: '#ad2f00' } }
-						/>
-						<path
-							d="M48.11 25.75c3.62.12 7 2.36 7 6a7.44 7.44 0 0 1-7 7.14 6.49 6.49 0 0 1-6.56-6.56 6.6 6.6 0 0 1 6.56-6.58Z"
-							style={ { fill: '#af3000' } }
-						/>
-						<path
-							d="M48.12 25.76c3.62.12 7 2.36 7 6a7.44 7.44 0 0 1-7 7.15 6.5 6.5 0 0 1-6.57-6.56 6.62 6.62 0 0 1 6.57-6.59Z"
-							style={ { fill: '#b13000' } }
-						/>
-						<path
-							d="M48.13 25.77c3.62.12 7 2.35 7 6s-3.35 7.15-7 7.15a6.5 6.5 0 0 1-6.57-6.56 6.62 6.62 0 0 1 6.57-6.59Z"
-							style={ { fill: '#b33100' } }
-						/>
-						<path
-							d="M48.13 25.79c3.62.12 7 2.34 7 6a7.46 7.46 0 0 1-7 7.16 6.5 6.5 0 0 1-6.56-6.56 6.59 6.59 0 0 1 6.56-6.6Z"
-							style={ { fill: '#b43100' } }
-						/>
-						<path
-							d="M48.14 25.8c3.62.12 7 2.33 7 6s-3.36 7.16-7 7.16a6.49 6.49 0 0 1-6.56-6.56 6.59 6.59 0 0 1 6.56-6.6Z"
-							style={ { fill: '#b63200' } }
-						/>
-						<path
-							d="M48.15 25.81c3.62.12 7 2.33 7 6s-3.36 7.16-7 7.16a6.48 6.48 0 0 1-6.56-6.55 6.59 6.59 0 0 1 6.56-6.61Z"
-							style={ { fill: '#b83200' } }
-						/>
-						<path
-							d="M48.15 25.82c3.62.12 7 2.32 7 5.94s-3.36 7.17-7 7.17a6.48 6.48 0 0 1-6.55-6.55 6.59 6.59 0 0 1 6.55-6.56Z"
-							style={ { fill: '#b93300' } }
-						/>
-						<path
-							d="M55.15 31.77c0 3.62-3.37 7.17-7 7.17a6.49 6.49 0 0 1-6.56-6.55 6.61 6.61 0 0 1 6.56-6.56c3.63.17 7 2.32 7 5.94Z"
-							style={ { fill: '#b30' } }
-						/>
-						<path
-							d="M48.17 25.84c3.62.12 7 2.32 7 5.94s-3.38 7.22-7 7.22a6.49 6.49 0 0 1-6.56-6.56 6.6 6.6 0 0 1 6.56-6.6Z"
-							style={ { fill: '#bc3502', opacity: 0.99 } }
-						/>
-						<path
-							d="M48.17 25.85c3.62.12 7 2.33 7 5.95s-3.38 7.2-7 7.2a6.49 6.49 0 0 1-6.55-6.56 6.59 6.59 0 0 1 6.55-6.59Z"
-							style={ { fill: '#bc3705', opacity: 0.98 } }
-						/>
-						<path
-							d="M48.18 25.85c3.62.13 7 2.34 7 6S51.8 39 48.18 39a6.49 6.49 0 0 1-6.56-6.56 6.61 6.61 0 0 1 6.56-6.59Z"
-							style={ { fill: '#bd3907', opacity: 0.97 } }
-						/>
-						<path
-							d="M48.18 25.86c3.62.12 7 2.34 7 6a7.45 7.45 0 0 1-7 7.15 6.48 6.48 0 0 1-6.55-6.55 6.6 6.6 0 0 1 6.55-6.6Z"
-							style={ { fill: '#bd3a09', opacity: 0.96 } }
-						/>
-						<path
-							d="M48.19 25.87c3.62.12 7 2.35 7 6a7.44 7.44 0 0 1-7 7.14 6.49 6.49 0 0 1-6.56-6.55 6.61 6.61 0 0 1 6.56-6.59Z"
-							style={ { fill: '#be3c0c', opacity: 0.95 } }
-						/>
-						<path
-							d="M48.19 25.88c3.62.12 7 2.35 7 6a7.44 7.44 0 0 1-7 7.14 6.48 6.48 0 0 1-6.55-6.55 6.6 6.6 0 0 1 6.55-6.59Z"
-							style={ { fill: '#bf3e0e', opacity: 0.95 } }
-						/>
-						<path
-							d="M48.2 25.89c3.62.12 7 2.36 7 6a7.42 7.42 0 0 1-7 7.13 6.5 6.5 0 0 1-6.56-6.56 6.6 6.6 0 0 1 6.56-6.57Z"
-							style={ { fill: '#bf4010', opacity: 0.94 } }
-						/>
-						<path
-							d="M48.21 25.9c3.61.11 6.95 2.36 6.95 6a7.42 7.42 0 0 1-6.95 7.1 6.5 6.5 0 0 1-6.56-6.56 6.59 6.59 0 0 1 6.56-6.54Z"
-							style={ { fill: '#c04213', opacity: 0.93 } }
-						/>
-						<path
-							d="M48.21 25.9c3.62.12 7 2.37 7 6a7.42 7.42 0 0 1-7 7.13 6.49 6.49 0 0 1-6.55-6.56 6.59 6.59 0 0 1 6.55-6.57Z"
-							style={ { fill: '#c14415', opacity: 0.92 } }
-						/>
-						<path
-							d="M48.22 25.91c3.62.12 7 2.38 7 6a7.41 7.41 0 0 1-7 7.12 6.5 6.5 0 0 1-6.56-6.56 6.59 6.59 0 0 1 6.56-6.56Z"
-							style={ { fill: '#c14617', opacity: 0.91 } }
-						/>
-						<path
-							d="M48.22 25.92c3.62.11 7 2.38 7 6a7.4 7.4 0 0 1-7 7.11 6.48 6.48 0 0 1-6.55-6.55 6.59 6.59 0 0 1 6.55-6.56Z"
-							style={ { fill: '#c24719', opacity: 0.9 } }
-						/>
-						<path
-							d="M48.23 25.93c3.62.11 6.94 2.39 6.94 6A7.38 7.38 0 0 1 48.23 39a6.49 6.49 0 0 1-6.56-6.55 6.59 6.59 0 0 1 6.56-6.52Z"
-							style={ { fill: '#c2491c', opacity: 0.89 } }
-						/>
-						<path
-							d="M48.23 25.94c3.62.11 6.94 2.39 6.94 6a7.38 7.38 0 0 1-6.94 7.1 6.49 6.49 0 0 1-6.55-6.56 6.59 6.59 0 0 1 6.55-6.54Z"
-							style={ { fill: '#c34b1e', opacity: 0.88 } }
-						/>
-						<path
-							d="M48.24 26c3.62.11 6.93 2.39 6.93 6a7.36 7.36 0 0 1-6.93 7.1 6.51 6.51 0 0 1-6.56-6.56A6.59 6.59 0 0 1 48.24 26Z"
-							style={ { fill: '#c44d20', opacity: 0.87 } }
-						/>
-						<path
-							d="M48.25 26c3.61.1 6.93 2.4 6.93 6a7.36 7.36 0 0 1-6.93 7.09 6.49 6.49 0 0 1-6.56-6.56A6.59 6.59 0 0 1 48.25 26Z"
-							style={ { fill: '#c44f23', opacity: 0.86 } }
-						/>
-						<path
-							d="M48.25 26c3.62.11 6.93 2.41 6.93 6a7.36 7.36 0 0 1-6.93 7.09 6.51 6.51 0 0 1-6.56-6.56A6.61 6.61 0 0 1 48.25 26Z"
-							style={ { fill: '#c55125', opacity: 0.85 } }
-						/>
-						<path
-							d="M48.26 26c3.62.11 6.92 2.42 6.92 6a7.33 7.33 0 0 1-6.92 7.07 6.5 6.5 0 0 1-6.56-6.55A6.6 6.6 0 0 1 48.26 26Z"
-							style={ { fill: '#c65327', opacity: 0.85 } }
-						/>
-						<path
-							d="M48.26 26c3.62.11 6.92 2.42 6.92 6a7.33 7.33 0 0 1-6.92 7.07 6.48 6.48 0 0 1-6.55-6.55A6.59 6.59 0 0 1 48.26 26Z"
-							style={ { fill: '#c6542a', opacity: 0.84 } }
-						/>
-						<path
-							d="M48.27 26c3.62.1 6.91 2.42 6.91 6a7.32 7.32 0 0 1-6.91 7.07 6.51 6.51 0 0 1-6.56-6.56A6.59 6.59 0 0 1 48.27 26Z"
-							style={ { fill: '#c7562c', opacity: 0.83 } }
-						/>
-						<path
-							d="M48.27 26c3.62.1 6.91 2.43 6.91 6a7.31 7.31 0 0 1-6.91 7.06 6.5 6.5 0 0 1-6.55-6.56 6.58 6.58 0 0 1 6.55-6.5Z"
-							style={ { fill: '#c7582e', opacity: 0.82 } }
-						/>
-						<path
-							d="M48.28 26c3.62.1 6.91 2.43 6.91 6.05a7.31 7.31 0 0 1-6.91 7.06 6.51 6.51 0 0 1-6.56-6.56A6.59 6.59 0 0 1 48.28 26Z"
-							style={ { fill: '#c85a31', opacity: 0.81 } }
-						/>
-						<path
-							d="M48.28 26c3.62.1 6.91 2.45 6.91 6.07a7.3 7.3 0 0 1-6.91 7.05 6.5 6.5 0 0 1-6.55-6.56A6.59 6.59 0 0 1 48.28 26Z"
-							style={ { fill: '#c95c33', opacity: 0.8 } }
-						/>
-						<path
-							d="M48.29 26c3.62.1 6.9 2.45 6.9 6.07a7.29 7.29 0 0 1-6.9 7 6.5 6.5 0 0 1-6.56-6.55A6.6 6.6 0 0 1 48.29 26Z"
-							style={ { fill: '#c95e35', opacity: 0.79 } }
-						/>
-						<path
-							d="M48.3 26c3.61.1 6.89 2.45 6.89 6.07a7.27 7.27 0 0 1-6.89 7 6.49 6.49 0 0 1-6.56-6.55A6.59 6.59 0 0 1 48.3 26Z"
-							style={ { fill: '#ca6038', opacity: 0.78 } }
-						/>
-						<path
-							d="M48.3 26c3.62.1 6.89 2.46 6.89 6.08a7.27 7.27 0 0 1-6.89 7 6.51 6.51 0 0 1-6.56-6.56A6.59 6.59 0 0 1 48.3 26Z"
-							style={ { fill: '#ca613a', opacity: 0.77 } }
-						/>
-						<path
-							d="M48.31 26.05c3.62.09 6.88 2.46 6.88 6.08a7.26 7.26 0 0 1-6.88 7 6.51 6.51 0 0 1-6.56-6.56 6.59 6.59 0 0 1 6.56-6.52Z"
-							style={ { fill: '#cb633c', opacity: 0.76 } }
-						/>
-						<path
-							d="M48.31 26.06c3.62.09 6.89 2.47 6.89 6.09a7.26 7.26 0 0 1-6.89 7 6.5 6.5 0 0 1-6.55-6.56 6.58 6.58 0 0 1 6.55-6.53Z"
-							style={ { fill: '#cc653f', opacity: 0.75 } }
-						/>
-						<path
-							d="M48.32 26.06c3.62.1 6.88 2.48 6.88 6.1a7.25 7.25 0 0 1-6.88 7 6.51 6.51 0 0 1-6.56-6.56 6.6 6.6 0 0 1 6.56-6.54Z"
-							style={ { fill: '#cc6741', opacity: 0.75 } }
-						/>
-						<path
-							d="M48.32 26.07c3.62.09 6.88 2.48 6.88 6.1a7.23 7.23 0 0 1-6.88 7 6.5 6.5 0 0 1-6.55-6.55 6.59 6.59 0 0 1 6.55-6.55Z"
-							style={ { fill: '#cd6943', opacity: 0.74 } }
-						/>
-						<path
-							d="M48.33 26.08c3.62.09 6.87 2.49 6.87 6.11a7.22 7.22 0 0 1-6.87 7 6.5 6.5 0 0 1-6.56-6.55 6.59 6.59 0 0 1 6.56-6.56Z"
-							style={ { fill: '#ce6b46', opacity: 0.73 } }
-						/>
-						<path
-							d="M48.33 26.09c3.62.09 6.87 2.49 6.87 6.11a7.22 7.22 0 0 1-6.87 7 6.5 6.5 0 0 1-6.55-6.55 6.58 6.58 0 0 1 6.55-6.56Z"
-							style={ { fill: '#ce6c48', opacity: 0.72 } }
-						/>
-						<path
-							d="M48.34 26.1c3.62.09 6.86 2.5 6.86 6.12a7.21 7.21 0 0 1-6.86 7 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#cf6e4a', opacity: 0.71 } }
-						/>
-						<path
-							d="M48.35 26.11c3.62.08 6.86 2.5 6.86 6.12a7.21 7.21 0 0 1-6.86 7 6.51 6.51 0 0 1-6.56-6.56 6.58 6.58 0 0 1 6.56-6.56Z"
-							style={ { fill: '#cf704c', opacity: 0.7 } }
-						/>
-						<path
-							d="M48.35 26.11c3.62.09 6.86 2.51 6.86 6.13a7.2 7.2 0 0 1-6.86 7 6.5 6.5 0 0 1-6.55-6.56 6.58 6.58 0 0 1 6.55-6.57Z"
-							style={ { fill: '#d0724f', opacity: 0.69 } }
-						/>
-						<path
-							d="M48.36 26.12c3.62.09 6.85 2.52 6.85 6.14a7.19 7.19 0 0 1-6.85 7 6.52 6.52 0 0 1-6.56-6.56 6.59 6.59 0 0 1 6.56-6.58Z"
-							style={ { fill: '#d17451', opacity: 0.68 } }
-						/>
-						<path
-							d="M48.36 26.13c3.62.09 6.85 2.52 6.85 6.14a7.18 7.18 0 0 1-6.85 7 6.5 6.5 0 0 1-6.55-6.55 6.58 6.58 0 0 1 6.55-6.59Z"
-							style={ { fill: '#d17653', opacity: 0.67 } }
-						/>
-						<path
-							d="M48.37 26.14c3.62.08 6.84 2.53 6.84 6.15a7.16 7.16 0 0 1-6.84 7 6.51 6.51 0 0 1-6.56-6.55 6.59 6.59 0 0 1 6.56-6.6Z"
-							style={ { fill: '#d27856', opacity: 0.66 } }
-						/>
-						<path
-							d="M48.37 26.15c3.62.08 6.84 2.53 6.84 6.15a7.16 7.16 0 0 1-6.84 7 6.51 6.51 0 0 1-6.55-6.56 6.58 6.58 0 0 1 6.55-6.59Z"
-							style={ { fill: '#d27958', opacity: 0.65 } }
-						/>
-						<path
-							d="M48.38 26.16c3.62.08 6.84 2.53 6.84 6.15a7.15 7.15 0 0 1-6.84 7 6.52 6.52 0 0 1-6.56-6.56 6.59 6.59 0 0 1 6.56-6.59Z"
-							style={ { fill: '#d37b5a', opacity: 0.65 } }
-						/>
-						<path
-							d="M48.38 26.16c3.62.08 6.84 2.55 6.84 6.17a7.14 7.14 0 0 1-6.84 7 6.51 6.51 0 0 1-6.55-6.56 6.59 6.59 0 0 1 6.55-6.61Z"
-							style={ { fill: '#d47d5d', opacity: 0.64 } }
-						/>
-						<path
-							d="M48.39 26.17c3.62.08 6.83 2.55 6.83 6.17a7.14 7.14 0 0 1-6.83 6.95 6.53 6.53 0 0 1-6.56-6.56 6.6 6.6 0 0 1 6.56-6.56Z"
-							style={ { fill: '#d47f5f', opacity: 0.63 } }
-						/>
-						<path
-							d="M48.4 26.18c3.62.08 6.82 2.56 6.82 6.18a7.12 7.12 0 0 1-6.82 6.93 6.51 6.51 0 0 1-6.56-6.55 6.58 6.58 0 0 1 6.56-6.56Z"
-							style={ { fill: '#d58161', opacity: 0.62 } }
-						/>
-						<path
-							d="M48.4 26.19c3.62.08 6.82 2.56 6.82 6.18a7.12 7.12 0 0 1-6.82 6.93 6.5 6.5 0 0 1-6.55-6.55 6.57 6.57 0 0 1 6.55-6.56Z"
-							style={ { fill: '#d68364', opacity: 0.61 } }
-						/>
-						<path
-							d="M48.41 26.2c3.62.07 6.81 2.56 6.81 6.18a7.1 7.1 0 0 1-6.81 6.93 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#d68566', opacity: 0.6 } }
-						/>
-						<path
-							d="M48.41 26.21c3.62.07 6.82 2.57 6.82 6.19a7.11 7.11 0 0 1-6.82 6.92 6.56 6.56 0 0 1 0-13.11Z"
-							style={ { fill: '#d78668', opacity: 0.59 } }
-						/>
-						<path
-							d="M48.42 26.22c3.62.07 6.81 2.57 6.81 6.19a7.1 7.1 0 0 1-6.81 6.92 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#d7886b', opacity: 0.58 } }
-						/>
-						<path
-							d="M48.42 26.22c3.62.08 6.81 2.59 6.81 6.21a7.08 7.08 0 0 1-6.81 6.91 6.51 6.51 0 0 1-6.55-6.56 6.59 6.59 0 0 1 6.55-6.56Z"
-							style={ { fill: '#d88a6d', opacity: 0.57 } }
-						/>
-						<path
-							d="M48.43 26.23c3.62.07 6.8 2.59 6.8 6.21a7.07 7.07 0 0 1-6.8 6.9 6.52 6.52 0 0 1-6.56-6.55 6.59 6.59 0 0 1 6.56-6.56Z"
-							style={ { fill: '#d98c6f', opacity: 0.56 } }
-						/>
-						<path
-							d="M48.44 26.24c3.62.07 6.79 2.59 6.79 6.21a7.05 7.05 0 0 1-6.79 6.9 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#d98e72', opacity: 0.55 } }
-						/>
-						<path
-							d="M48.44 26.25c3.62.07 6.79 2.6 6.79 6.22a7.05 7.05 0 0 1-6.79 6.89 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#da9074', opacity: 0.55 } }
-						/>
-						<path
-							d="M48.45 26.26c3.62.06 6.78 2.6 6.78 6.22a7 7 0 0 1-6.78 6.89 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#db9276', opacity: 0.54 } }
-						/>
-						<path
-							d="M48.45 26.27c3.62.06 6.79 2.61 6.79 6.23a7 7 0 0 1-6.79 6.88 6.56 6.56 0 0 1 0-13.11Z"
-							style={ { fill: '#db9379', opacity: 0.53 } }
-						/>
-						<path
-							d="M48.46 26.27c3.62.07 6.78 2.62 6.78 6.24a7 7 0 0 1-6.78 6.88 6.56 6.56 0 0 1 0-13.12Z"
-							style={ { fill: '#dc957b', opacity: 0.52 } }
-						/>
-						<path
-							d="M48.46 26.28c3.62.07 6.78 2.62 6.78 6.24a7 7 0 0 1-6.78 6.87 6.56 6.56 0 0 1 0-13.11Z"
-							style={ { fill: '#dc977d', opacity: 0.51 } }
-						/>
-						<path
-							d="M48.47 26.29c3.62.06 6.77 2.63 6.77 6.25a7 7 0 0 1-6.77 6.86 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#dd9980', opacity: 0.5 } }
-						/>
-						<path
-							d="M48.47 26.3c3.62.06 6.77 2.63 6.77 6.25a7 7 0 0 1-6.77 6.86 6.56 6.56 0 0 1 0-13.11Z"
-							style={ { fill: '#de9b82', opacity: 0.49 } }
-						/>
-						<path
-							d="M48.48 26.31c3.62.06 6.76 2.64 6.76 6.26a7 7 0 0 1-6.76 6.85 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#de9d84', opacity: 0.48 } }
-						/>
-						<path
-							d="M48.49 26.32c3.62.06 6.76 2.64 6.76 6.26a7 7 0 0 1-6.76 6.85 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#df9f86', opacity: 0.47 } }
-						/>
-						<path
-							d="M48.49 26.32c3.62.06 6.76 2.65 6.76 6.27a7 7 0 0 1-6.76 6.85 6.56 6.56 0 0 1 0-13.12Z"
-							style={ { fill: '#dfa089', opacity: 0.46 } }
-						/>
-						<path
-							d="M48.5 26.33c3.62.06 6.75 2.66 6.75 6.28a7 7 0 0 1-6.75 6.83 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#e0a28b', opacity: 0.45 } }
-						/>
-						<path
-							d="M48.5 26.34c3.62.06 6.75 2.66 6.75 6.28a7 7 0 0 1-6.75 6.83 6.56 6.56 0 0 1 0-13.11Z"
-							style={ { fill: '#e1a48d', opacity: 0.45 } }
-						/>
-						<path
-							d="M48.51 26.35c3.62 0 6.74 2.67 6.74 6.29a7 7 0 0 1-6.74 6.82 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#e1a690', opacity: 0.44 } }
-						/>
-						<path
-							d="M48.51 26.36c3.62.05 6.74 2.67 6.74 6.29a7 7 0 0 1-6.74 6.82 6.56 6.56 0 0 1 0-13.11Z"
-							style={ { fill: '#e2a892', opacity: 0.43 } }
-						/>
-						<path
-							d="M48.52 26.37c3.62.05 6.74 2.67 6.74 6.29a6.94 6.94 0 0 1-6.74 6.82 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#e3aa94', opacity: 0.42 } }
-						/>
-						<path
-							d="M48.52 26.37c3.62.06 6.74 2.69 6.74 6.31a6.93 6.93 0 0 1-6.74 6.81 6.56 6.56 0 0 1 0-13.12Z"
-							style={ { fill: '#e3ac97', opacity: 0.41 } }
-						/>
-						<path
-							d="M48.53 26.38a6.54 6.54 0 0 1 6.73 6.31 6.93 6.93 0 0 1-6.73 6.81 6.56 6.56 0 0 1 0-13.12Z"
-							style={ { fill: '#e4ad99', opacity: 0.4 } }
-						/>
-						<path
-							d="M48.54 26.39a6.54 6.54 0 0 1 6.72 6.32 6.9 6.9 0 0 1-6.72 6.79 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#e4af9b', opacity: 0.39 } }
-						/>
-						<path
-							d="M48.54 26.4a6.54 6.54 0 0 1 6.72 6.32 6.9 6.9 0 0 1-6.72 6.79 6.56 6.56 0 0 1 0-13.11Z"
-							style={ { fill: '#e5b19e', opacity: 0.38 } }
-						/>
-						<path
-							d="M48.55 26.41a6.52 6.52 0 0 1 6.71 6.32 6.89 6.89 0 0 1-6.71 6.79 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#e6b3a0', opacity: 0.37 } }
-						/>
-						<path
-							d="M48.55 26.42a6.54 6.54 0 0 1 6.72 6.33 6.89 6.89 0 0 1-6.72 6.78 6.56 6.56 0 0 1 0-13.11Z"
-							style={ { fill: '#e6b5a2', opacity: 0.36 } }
-						/>
-						<path
-							d="M48.56 26.43a6.53 6.53 0 0 1 6.71 6.33 6.88 6.88 0 0 1-6.71 6.78 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#e7b7a5', opacity: 0.35 } }
-						/>
-						<path
-							d="M48.56 26.43a6.55 6.55 0 0 1 6.71 6.35 6.88 6.88 0 0 1-6.71 6.77 6.56 6.56 0 0 1 0-13.12Z"
-							style={ { fill: '#e8b9a7', opacity: 0.35 } }
-						/>
-						<path
-							d="M48.57 26.44a6.54 6.54 0 0 1 6.7 6.35 6.86 6.86 0 0 1-6.7 6.76 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#e8baa9', opacity: 0.34 } }
-						/>
-						<path
-							d="M48.58 26.45a6.53 6.53 0 0 1 6.69 6.35 6.84 6.84 0 0 1-6.69 6.76 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#e9bcac', opacity: 0.33 } }
-						/>
-						<path
-							d="M48.58 26.46a6.53 6.53 0 0 1 6.69 6.36 6.84 6.84 0 0 1-6.69 6.75 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#e9beae', opacity: 0.32 } }
-						/>
-						<path
-							d="M48.59 26.47a6.53 6.53 0 0 1 6.69 6.36 6.84 6.84 0 0 1-6.69 6.75 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#eac0b0', opacity: 0.31 } }
-						/>
-						<path
-							d="M48.59 26.48a6.54 6.54 0 0 1 6.69 6.37 6.83 6.83 0 0 1-6.69 6.74 6.56 6.56 0 0 1 0-13.11Z"
-							style={ { fill: '#ebc2b3', opacity: 0.3 } }
-						/>
-						<path
-							d="M48.6 26.48a6.54 6.54 0 0 1 6.68 6.38 6.82 6.82 0 0 1-6.68 6.74 6.56 6.56 0 0 1 0-13.12Z"
-							style={ { fill: '#ebc4b5', opacity: 0.29 } }
-						/>
-						<path
-							d="M48.6 26.49a6.54 6.54 0 0 1 6.68 6.38 6.81 6.81 0 0 1-6.68 6.73 6.56 6.56 0 0 1 0-13.11Z"
-							style={ { fill: '#ecc6b7', opacity: 0.28 } }
-						/>
-						<path
-							d="M48.61 26.5a6.54 6.54 0 0 1 6.67 6.39 6.8 6.8 0 0 1-6.67 6.72 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#ecc7b9', opacity: 0.27 } }
-						/>
-						<path
-							d="M48.61 26.51a6.54 6.54 0 0 1 6.67 6.39 6.79 6.79 0 0 1-6.67 6.72 6.56 6.56 0 0 1 0-13.11Z"
-							style={ { fill: '#edc9bc', opacity: 0.26 } }
-						/>
-						<path
-							d="M48.62 26.52a6.54 6.54 0 0 1 6.67 6.4 6.79 6.79 0 0 1-6.67 6.71 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#eecbbe', opacity: 0.25 } }
-						/>
-						<path
-							d="M48.63 26.53a6.53 6.53 0 0 1 6.66 6.4 6.78 6.78 0 0 1-6.66 6.71 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#eecdc0', opacity: 0.25 } }
-						/>
-						<path
-							d="M48.63 26.53a6.54 6.54 0 0 1 6.66 6.41 6.77 6.77 0 0 1-6.66 6.71 6.56 6.56 0 0 1 0-13.12Z"
-							style={ { fill: '#efcfc3', opacity: 0.24 } }
-						/>
-						<path
-							d="M48.64 26.54A6.54 6.54 0 0 1 55.29 33a6.75 6.75 0 0 1-6.65 6.69 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#f0d1c5', opacity: 0.23 } }
-						/>
-						<path
-							d="M48.64 26.55A6.54 6.54 0 0 1 55.29 33a6.75 6.75 0 0 1-6.65 6.69 6.56 6.56 0 0 1 0-13.11Z"
-							style={ { fill: '#f0d2c7', opacity: 0.22 } }
-						/>
-						<path
-							d="M48.65 26.56A6.54 6.54 0 0 1 55.29 33a6.74 6.74 0 0 1-6.64 6.68 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#f1d4ca', opacity: 0.21 } }
-						/>
-						<path
-							d="M48.65 26.57A6.54 6.54 0 0 1 55.3 33a6.74 6.74 0 0 1-6.65 6.68 6.56 6.56 0 0 1 0-13.11Z"
-							style={ { fill: '#f1d6cc', opacity: 0.2 } }
-						/>
-						<path
-							d="M48.66 26.58A6.54 6.54 0 0 1 55.3 33a6.73 6.73 0 0 1-6.64 6.68 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#f2d8ce', opacity: 0.19 } }
-						/>
-						<path
-							d="M48.66 26.58A6.55 6.55 0 0 1 55.3 33a6.72 6.72 0 0 1-6.64 6.67 6.56 6.56 0 0 1 0-13.12Z"
-							style={ { fill: '#f3dad1', opacity: 0.18 } }
-						/>
-						<path
-							d="M48.67 26.59A6.54 6.54 0 0 1 55.3 33a6.71 6.71 0 0 1-6.63 6.66 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#f3dcd3', opacity: 0.17 } }
-						/>
-						<path
-							d="M48.68 26.6a6.54 6.54 0 0 1 6.62 6.46 6.69 6.69 0 0 1-6.62 6.65 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#f4ded5', opacity: 0.16 } }
-						/>
-						<path
-							d="M48.68 26.61a6.54 6.54 0 0 1 6.62 6.46 6.69 6.69 0 0 1-6.62 6.65 6.56 6.56 0 0 1 0-13.11Z"
-							style={ { fill: '#f4dfd8', opacity: 0.15 } }
-						/>
-						<path
-							d="M48.69 26.62a6.54 6.54 0 0 1 6.62 6.46 6.69 6.69 0 0 1-6.62 6.65 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#f5e1da', opacity: 0.15 } }
-						/>
-						<path
-							d="M48.69 26.63a6.55 6.55 0 0 1 6.62 6.47 6.68 6.68 0 0 1-6.62 6.64 6.56 6.56 0 0 1 0-13.11Z"
-							style={ { fill: '#f6e3dc', opacity: 0.14 } }
-						/>
-						<path
-							d="M48.7 26.63a6.55 6.55 0 0 1 6.61 6.48 6.68 6.68 0 0 1-6.61 6.64 6.56 6.56 0 0 1 0-13.12Z"
-							style={ { fill: '#f6e5df', opacity: 0.13 } }
-						/>
-						<path
-							d="M48.7 26.64a6.56 6.56 0 0 1 6.61 6.49 6.67 6.67 0 0 1-6.61 6.63 6.56 6.56 0 0 1 0-13.12Z"
-							style={ { fill: '#f7e7e1', opacity: 0.12 } }
-						/>
-						<path
-							d="M48.71 26.65a6.55 6.55 0 0 1 6.6 6.49 6.65 6.65 0 0 1-6.6 6.62 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#f8e9e3', opacity: 0.11 } }
-						/>
-						<path
-							d="M48.72 26.66a6.54 6.54 0 0 1 6.59 6.49 6.63 6.63 0 0 1-6.59 6.62 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#f8ebe6', opacity: 0.1 } }
-						/>
-						<path
-							d="M48.72 26.67a6.55 6.55 0 0 1 6.6 6.5 6.64 6.64 0 0 1-6.6 6.61 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#f9ece8', opacity: 0.09 } }
-						/>
-						<path
-							d="M48.73 26.68a6.54 6.54 0 0 1 6.59 6.5 6.63 6.63 0 0 1-6.59 6.61 6.56 6.56 0 1 1 0-13.11Z"
-							style={ { fill: '#f9eeea', opacity: 0.08 } }
-						/>
-						<path
-							d="M48.73 26.69a6.55 6.55 0 0 1 6.59 6.51 6.62 6.62 0 0 1-6.59 6.6 6.56 6.56 0 0 1 0-13.11Z"
-							style={ { fill: '#faf0ec', opacity: 0.07 } }
-						/>
-						<path
-							d="M48.74 26.69a6.56 6.56 0 1 1-6.56 6.56 6.55 6.55 0 0 1 6.56-6.56Z"
-							style={ { fill: '#fbf2ef', opacity: 0.06 } }
-						/>
-						<path
-							d="M48.74 26.7a6.56 6.56 0 1 1-6.55 6.56 6.55 6.55 0 0 1 6.55-6.56Z"
-							style={ { fill: '#fbf4f1', opacity: 0.05 } }
-						/>
-						<path
-							d="M48.75 26.71a6.56 6.56 0 1 1-6.56 6.56 6.55 6.55 0 0 1 6.56-6.56Z"
-							style={ { fill: '#fcf6f3', opacity: 0.05 } }
-						/>
-						<path
-							d="M48.75 26.72a6.56 6.56 0 1 1-6.55 6.55 6.55 6.55 0 0 1 6.55-6.55Z"
-							style={ { fill: '#fdf8f6', opacity: 0.04 } }
-						/>
-						<path
-							d="M48.76 26.73a6.56 6.56 0 1 1-6.56 6.55 6.56 6.56 0 0 1 6.56-6.55Z"
-							style={ { fill: '#fdf9f8', opacity: 0.03 } }
-						/>
-						<path
-							d="M48.77 26.74a6.56 6.56 0 1 1-6.56 6.55 6.55 6.55 0 0 1 6.56-6.55Z"
-							style={ { fill: '#fefbfa', opacity: 0.02 } }
-						/>
-						<path
-							d="M48.77 26.74a6.56 6.56 0 1 1-6.55 6.56 6.56 6.56 0 0 1 6.55-6.56Z"
-							style={ { fill: '#fefdfd', opacity: 0.01 } }
-						/>
-					</g>
-				</g>
-				<g style={ { clipPath: `url(#clip-path-2)` } }>
-					<g style={ { clipPath: `url(#clip-path-3)` } }>
-						<path
-							transform="rotate(-30 47.879 31.429)"
-							style={ { fill: '#e04e15' } }
-							d="M-6.16-22.61h108.07V85.46H-6.16z"
-						/>
-						<path
-							transform="rotate(-30 48.999 33.37)"
-							style={ { fill: '#e04f15' } }
-							d="M-5.03-20.66h108.07V87.41H-5.03z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -10.95 29.8)"
-							style={ { fill: '#e15015' } }
-							d="M-3.91-18.71h108.07V89.36H-3.91z"
-						/>
-						<path
-							transform="rotate(-30 51.253 37.273)"
-							style={ { fill: '#e15115' } }
-							d="M-2.78-16.77h108.07V91.3H-2.78z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -12.59 31.44)"
-							style={ { fill: '#e25215' } }
-							d="M-1.66-14.82h108.07V93.25H-1.66z"
-						/>
-						<path
-							transform="rotate(-30 53.507 41.177)"
-							style={ { fill: '#e25316' } }
-							d="M-.53-12.87h108.07V95.2H-.53z"
-						/>
-						<path
-							transform="rotate(-30 54.627 43.117)"
-							style={ { fill: '#e35416' } }
-							d="M.59-10.92h108.07V97.15H.59z"
-						/>
-						<path
-							transform="rotate(-30 55.747 45.057)"
-							style={ { fill: '#e35516' } }
-							d="M1.72-8.97h108.07V99.1H1.72z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -15.89 34.74)"
-							style={ { fill: '#e35616' } }
-							d="M2.84-7.02h108.07v108.07H2.84z"
-						/>
-						<path
-							fill="#e45716"
-							d="M-15.812 29.176 77.78-24.86l54.035 93.591-93.59 54.035z"
-						/>
-						<path
-							fill="#e45716"
-							d="M-14.687 31.124 78.904-22.91l54.035 93.592-93.59 54.035z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -18.36 37.21)"
-							style={ { fill: '#e55816' } }
-							d="M6.22-1.18h108.07v108.07H6.22z"
-						/>
-						<path
-							transform="rotate(-30 61.375 54.805)"
-							style={ { fill: '#e55916' } }
-							d="M7.34.77h108.07v108.07H7.34z"
-						/>
-						<path
-							transform="rotate(-30 62.495 56.746)"
-							style={ { fill: '#e55a16' } }
-							d="M8.47 2.72h108.07v108.07H8.47z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -20.83 39.68)"
-							style={ { fill: '#e65b16' } }
-							d="M9.59 4.67h108.07v108.07H9.59z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -21.65 40.5)"
-							style={ { fill: '#e65c17' } }
-							d="M10.72 6.62h108.07v108.07H10.72z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -22.47 41.33)"
-							style={ { fill: '#e75d17' } }
-							d="M11.84 8.56h108.07v108.07H11.84z"
-						/>
-						<path
-							transform="rotate(-30 67.003 64.553)"
-							style={ { fill: '#e75e17' } }
-							d="M12.97 10.51h108.07v108.07H12.97z"
-						/>
-						<path
-							transform="rotate(-30 68.123 66.494)"
-							style={ { fill: '#e85f17' } }
-							d="M14.09 12.46h108.07v108.07H14.09z"
-						/>
-						<path
-							transform="rotate(-30 69.262 68.439)"
-							style={ { fill: '#e86017' } }
-							d="M15.22 14.41h108.07v108.07H15.22z"
-						/>
-						<path
-							transform="rotate(-30 69.68 69.19)"
-							style={ { fill: '#e86117' } }
-							d="M15.65 15.16h108.07v108.07H15.65z"
-						/>
-						<path
-							transform="rotate(-30 70.118 69.948)"
-							style={ { fill: '#e96217' } }
-							d="M16.09 15.92h108.07v108.07H16.09z"
-						/>
-						<path
-							transform="rotate(-30 70.555 70.705)"
-							style={ { fill: '#e96317' } }
-							d="M16.52 16.67h108.07v108.07H16.52z"
-						/>
-						<path
-							transform="rotate(-30 70.992 71.462)"
-							style={ { fill: '#e96418' } }
-							d="M16.96 17.42h108.07v108.07H16.96z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -26.54 45.39)"
-							style={ { fill: '#e96518' } }
-							d="M17.39 18.18h108.07v108.07H17.39z"
-						/>
-						<path
-							transform="rotate(-30 71.871 72.958)"
-							style={ { fill: '#ea6618' } }
-							d="M17.82 18.93h108.07V127H17.82z"
-						/>
-						<path
-							transform="rotate(-30 72.29 73.71)"
-							style={ { fill: '#ea6718' } }
-							d="M18.26 19.68h108.07v108.07H18.26z"
-						/>
-						<path
-							transform="rotate(-30 72.727 74.467)"
-							style={ { fill: '#ea6818' } }
-							d="M18.69 20.43h108.07V128.5H18.69z"
-						/>
-						<path
-							transform="rotate(-30 73.164 75.224)"
-							style={ { fill: '#ea6918' } }
-							d="M19.13 21.19H127.2v108.07H19.13z"
-						/>
-						<path
-							transform="rotate(-30 73.6 75.981)"
-							style={ { fill: '#eb6a18' } }
-							d="M19.56 21.94h108.07v108.07H19.56z"
-						/>
-						<path
-							transform="rotate(-30 74.038 76.738)"
-							style={ { fill: '#eb6b19' } }
-							d="M20 22.69h108.07v108.07H20z"
-						/>
-						<path
-							transform="rotate(-30 74.48 77.477)"
-							style={ { fill: '#eb6c19' } }
-							d="M20.43 23.45H128.5v108.07H20.43z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -29.08 47.93)"
-							style={ { fill: '#eb6d19' } }
-							d="M20.87 24.2h108.07v108.07H20.87z"
-						/>
-						<path
-							transform="rotate(-30 75.336 78.986)"
-							style={ { fill: '#ec6e19' } }
-							d="M21.3 24.95h108.07v108.07H21.3z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -29.72 48.57)"
-							style={ { fill: '#ec6f19' } }
-							d="M21.74 25.71h108.07v108.07H21.74z"
-						/>
-						<path
-							transform="rotate(-30 76.21 80.5)"
-							style={ { fill: '#ec7019' } }
-							d="M22.17 26.46h108.07v108.07H22.17z"
-						/>
-						<path
-							transform="rotate(-30 76.647 81.258)"
-							style={ { fill: '#ec7119' } }
-							d="M22.61 27.21h108.07v108.07H22.61z"
-						/>
-						<path
-							transform="rotate(-30 77.084 82.015)"
-							style={ { fill: '#ed7219' } }
-							d="M23.04 27.97h108.07v108.07H23.04z"
-						/>
-						<path
-							transform="rotate(-30 77.508 82.748)"
-							style={ { fill: '#ed731a' } }
-							d="M23.48 28.72h108.07v108.07H23.48z"
-						/>
-						<path
-							transform="rotate(-30 77.945 83.505)"
-							style={ { fill: '#ed741a' } }
-							d="M23.91 29.47h108.07v108.07H23.91z"
-						/>
-						<path
-							transform="rotate(-30 78.382 84.262)"
-							style={ { fill: '#ed751a' } }
-							d="M24.35 30.23h108.07V138.3H24.35z"
-						/>
-						<path
-							transform="rotate(-30 78.82 85.02)"
-							style={ { fill: '#ee761a' } }
-							d="M24.78 30.98h108.07v108.07H24.78z"
-						/>
-						<path
-							transform="rotate(-30 79.256 85.777)"
-							style={ { fill: '#ee771a' } }
-							d="M25.22 31.73h108.07V139.8H25.22z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -32.59 51.44)"
-							style={ { fill: '#ee781a' } }
-							d="M25.65 32.49h108.07v108.07H25.65z"
-						/>
-						<path
-							transform="rotate(-30 80.117 87.267)"
-							style={ { fill: '#ef791a' } }
-							d="M26.09 33.24h108.07v108.07H26.09z"
-						/>
-						<path
-							transform="rotate(-30 80.554 88.024)"
-							style={ { fill: '#ef7a1b' } }
-							d="M26.52 33.99h108.07v108.07H26.52z"
-						/>
-						<path
-							transform="rotate(-30 80.991 88.781)"
-							style={ { fill: '#ef7b1b' } }
-							d="M26.96 34.75h108.07v108.07H26.96z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -33.86 52.71)"
-							style={ { fill: '#ef7c1b' } }
-							d="M27.39 35.5h108.07v108.07H27.39z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -34.18 53.03)"
-							style={ { fill: '#f07d1b' } }
-							d="M27.83 36.25H135.9v108.07H27.83z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -34.5 53.35)"
-							style={ { fill: '#f07e1b' } }
-							d="M28.26 37.01h108.07v108.07H28.26z"
-						/>
-						<path
-							transform="rotate(-30 82.745 91.791)"
-							style={ { fill: '#f07f1b' } }
-							d="M28.7 37.76h108.07v108.07H28.7z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -35.13 53.98)"
-							style={ { fill: '#f0801b' } }
-							d="M29.13 38.51H137.2v108.07H29.13z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -35.45 54.3)"
-							style={ { fill: '#f1801c' } }
-							d="M29.57 39.27h108.07v108.07H29.57z"
-						/>
-						<path
-							transform="rotate(-30 84.037 94.058)"
-							style={ { fill: '#f1811c' } }
-							d="M30 40.02h108.07v108.07H30z"
-						/>
-						<path
-							transform="rotate(-30 84.474 94.815)"
-							style={ { fill: '#f1821c' } }
-							d="M30.44 40.77h108.07v108.07H30.44z"
-						/>
-						<path
-							transform="rotate(-30 84.912 95.572)"
-							style={ { fill: '#f1831c' } }
-							d="M30.87 41.53h108.07V149.6H30.87z"
-						/>
-						<path
-							transform="rotate(-30 85.354 96.31)"
-							style={ { fill: '#f2841c' } }
-							d="M31.31 42.28h108.07v108.07H31.31z"
-						/>
-						<path
-							transform="rotate(-30 85.772 97.063)"
-							style={ { fill: '#f2851c' } }
-							d="M31.74 43.03h108.07V151.1H31.74z"
-						/>
-						<path
-							transform="rotate(-30 86.21 97.82)"
-							style={ { fill: '#f2861c' } }
-							d="M32.18 43.79h108.07v108.07H32.18z"
-						/>
-						<path
-							transform="rotate(-30 86.646 98.577)"
-							style={ { fill: '#f2871d' } }
-							d="M32.61 44.54h108.07v108.07H32.61z"
-						/>
-						<path
-							transform="rotate(-30 87.084 99.334)"
-							style={ { fill: '#f3881d' } }
-							d="M33.05 45.29h108.07v108.07H33.05z"
-						/>
-						<path
-							transform="rotate(-30 87.52 100.091)"
-							style={ { fill: '#f3891d' } }
-							d="M33.48 46.04h108.07v108.07H33.48z"
-						/>
-						<path
-							transform="rotate(-30 87.963 100.83)"
-							style={ { fill: '#f38a1d' } }
-							d="M33.92 46.8h108.07v108.07H33.92z"
-						/>
-						<path
-							transform="rotate(-30 88.381 101.582)"
-							style={ { fill: '#f48b1d' } }
-							d="M34.35 47.55h108.07v108.07H34.35z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -39.27 58.12)"
-							style={ { fill: '#f48c1d' } }
-							d="M34.79 48.3h108.07v108.07H34.79z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -39.59 58.44)"
-							style={ { fill: '#f48d1d' } }
-							d="M35.22 49.06h108.07v108.07H35.22z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -39.91 58.76)"
-							style={ { fill: '#f48e1e' } }
-							d="M35.66 49.81h108.07v108.07H35.66z"
-						/>
-						<path
-							transform="rotate(-30 90.13 104.61)"
-							style={ { fill: '#f58f1e' } }
-							d="M36.09 50.56h108.07v108.07H36.09z"
-						/>
-						<path
-							transform="rotate(-30 90.572 105.349)"
-							style={ { fill: '#f5901e' } }
-							d="M36.52 51.32h108.07v108.07H36.52z"
-						/>
-						<path
-							transform="rotate(-30 90.99 106.1)"
-							style={ { fill: '#f5911e' } }
-							d="M36.96 52.07h108.07v108.07H36.96z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -41.18 60.03)"
-							style={ { fill: '#f5921e' } }
-							d="M37.39 52.82h108.07v108.07H37.39z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -41.5 60.35)"
-							style={ { fill: '#f6931e' } }
-							d="M37.83 53.58H145.9v108.07H37.83z"
-						/>
-						<path
-							transform="rotate(-30 92.302 108.372)"
-							style={ { fill: '#f6941e' } }
-							d="M38.26 54.33h108.07V162.4H38.26z"
-						/>
-						<path
-							transform="rotate(-30 92.739 109.13)"
-							style={ { fill: '#f6951e' } }
-							d="M38.7 55.08h108.07v108.07H38.7z"
-						/>
-						<path
-							transform="rotate(-30 93.181 109.868)"
-							style={ { fill: '#f6961f' } }
-							d="M39.13 55.84H147.2v108.07H39.13z"
-						/>
-						<path
-							transform="rotate(-30 93.6 110.62)"
-							style={ { fill: '#f7971f' } }
-							d="M39.57 56.59h108.07v108.07H39.57z"
-						/>
-						<path
-							transform="rotate(-30 94.037 111.377)"
-							style={ { fill: '#f7981f' } }
-							d="M40 57.34h108.07v108.07H40z"
-						/>
-						<path
-							transform="rotate(-30 94.474 112.134)"
-							style={ { fill: '#f7991f' } }
-							d="M40.44 58.1h108.07v108.07H40.44z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -43.73 62.58)"
-							style={ { fill: '#f79a1f' } }
-							d="M40.87 58.85h108.07v108.07H40.87z"
-						/>
-						<path
-							transform="rotate(-30 95.348 113.648)"
-							style={ { fill: '#f89b1f' } }
-							d="M41.31 59.6h108.07v108.07H41.31z"
-						/>
-						<path
-							transform="rotate(-30 95.79 114.387)"
-							style={ { fill: '#f89c1f' } }
-							d="M41.74 60.36h108.07v108.07H41.74z"
-						/>
-						<path
-							transform="matrix(.87 -.5 .5 .87 -44.68 63.53)"
-							style={ { fill: '#f89d20' } }
-							d="M42.18 61.11h108.07v108.07H42.18z"
-						/>
-						<path
-							transform="rotate(-30 96.646 115.896)"
-							style={ { fill: '#f89e20' } }
-							d="M42.61 61.86h108.07v108.07H42.61z"
-						/>
-						<path
-							transform="rotate(-30 97.083 116.653)"
-							style={ { fill: '#f99f20' } }
-							d="M43.05 62.62h108.07v108.07H43.05z"
-						/>
-						<path
-							transform="rotate(-30 97.52 117.41)"
-							style={ { fill: '#f9a020' } }
-							d="M43.48 63.37h108.07v108.07H43.48z"
-						/>
-					</g>
+				<path
+					fill="#0a0300"
+					d="M47.42 24.68a6.68 6.68 0 1 1-6.68 6.68 6.68 6.68 0 0 1 6.68-6.68Z"
+				/>
+				<path
+					fill="#0c0300"
+					d="M47.43 24.69a6.68 6.68 0 1 1-6.68 6.68 6.67 6.67 0 0 1 6.68-6.68Z"
+				/>
+				<path
+					fill="#0e0400"
+					d="M47.43 24.7a6.68 6.68 0 0 1 6.71 6.64 6.75 6.75 0 0 1-6.71 6.72 6.68 6.68 0 0 1 0-13.36Z"
+				/>
+				<path
+					fill="#0f0400"
+					d="M47.44 24.71a6.67 6.67 0 0 1 6.71 6.63 6.75 6.75 0 0 1-6.71 6.73 6.68 6.68 0 0 1 0-13.36Z"
+				/>
+				<path
+					fill="#110500"
+					d="M47.45 24.73a6.67 6.67 0 0 1 6.71 6.62 6.76 6.76 0 0 1-6.71 6.73 6.68 6.68 0 1 1 0-13.35Z"
+				/>
+				<path
+					fill="#130500"
+					d="M47.45 24.74a6.66 6.66 0 0 1 6.72 6.61 6.76 6.76 0 0 1-6.72 6.74 6.68 6.68 0 0 1 0-13.35Z"
+				/>
+				<path
+					fill="#140600"
+					d="M47.46 24.75a6.66 6.66 0 0 1 6.72 6.6 6.76 6.76 0 0 1-6.72 6.74 6.67 6.67 0 1 1 0-13.34Z"
+				/>
+				<path
+					fill="#160600"
+					d="M47.47 24.76a6.66 6.66 0 0 1 6.72 6.6 6.78 6.78 0 0 1-6.72 6.74 6.67 6.67 0 0 1 0-13.34Z"
+				/>
+				<path
+					fill="#180600"
+					d="M47.48 24.77a6.66 6.66 0 0 1 6.72 6.59 6.78 6.78 0 0 1-6.72 6.75 6.67 6.67 0 0 1 0-13.34Z"
+				/>
+				<path
+					fill="#190700"
+					d="M47.48 24.78a6.66 6.66 0 0 1 6.73 6.59 6.79 6.79 0 0 1-6.73 6.75 6.67 6.67 0 1 1 0-13.34Z"
+				/>
+				<path
+					fill="#1b0700"
+					d="M47.49 24.79a6.66 6.66 0 0 1 6.73 6.58 6.8 6.8 0 0 1-6.73 6.76 6.67 6.67 0 0 1 0-13.34Z"
+				/>
+				<path
+					fill="#1d0800"
+					d="M47.5 24.8a6.65 6.65 0 0 1 6.73 6.57 6.81 6.81 0 0 1-6.73 6.77 6.67 6.67 0 0 1 0-13.34Z"
+				/>
+				<path
+					fill="#1f0800"
+					d="M47.5 24.81a6.66 6.66 0 0 1 6.74 6.57 6.81 6.81 0 0 1-6.74 6.77 6.67 6.67 0 0 1 0-13.34Z"
+				/>
+				<path
+					fill="#200900"
+					d="M47.51 24.82a6.66 6.66 0 0 1 6.74 6.56 6.82 6.82 0 0 1-6.74 6.77 6.67 6.67 0 0 1 0-13.33Z"
+				/>
+				<path
+					fill="#220900"
+					d="M47.52 24.84a6.65 6.65 0 0 1 6.74 6.55 6.82 6.82 0 0 1-6.74 6.77 6.66 6.66 0 0 1 0-13.32Z"
+				/>
+				<path
+					fill="#240a00"
+					d="M47.53 24.85a6.64 6.64 0 0 1 6.74 6.54 6.83 6.83 0 0 1-6.74 6.78 6.66 6.66 0 1 1 0-13.32Z"
+				/>
+				<path
+					fill="#250a00"
+					d="M47.53 24.86a6.65 6.65 0 0 1 6.75 6.54 6.84 6.84 0 0 1-6.75 6.78 6.66 6.66 0 0 1 0-13.32Z"
+				/>
+				<path
+					fill="#270b00"
+					d="M47.54 24.87a6.65 6.65 0 0 1 6.75 6.53 6.85 6.85 0 0 1-6.75 6.79 6.66 6.66 0 1 1 0-13.32Z"
+				/>
+				<path
+					fill="#290b00"
+					d="M47.55 24.88a6.65 6.65 0 0 1 6.75 6.52 6.86 6.86 0 0 1-6.75 6.8 6.66 6.66 0 1 1 0-13.32Z"
+				/>
+				<path
+					fill="#2a0c00"
+					d="M47.55 24.89a6.65 6.65 0 0 1 6.76 6.52 6.86 6.86 0 0 1-6.76 6.8 6.66 6.66 0 0 1 0-13.32Z"
+				/>
+				<path
+					fill="#2c0c00"
+					d="M47.56 24.9a6.64 6.64 0 0 1 6.76 6.51 6.86 6.86 0 0 1-6.76 6.8 6.66 6.66 0 0 1 0-13.31Z"
+				/>
+				<path
+					fill="#2e0d00"
+					d="M47.57 24.91a6.64 6.64 0 0 1 6.76 6.51 6.88 6.88 0 0 1-6.76 6.8 6.66 6.66 0 1 1 0-13.31Z"
+				/>
+				<path
+					fill="#300d00"
+					d="M47.58 24.92a6.65 6.65 0 0 1 6.76 6.5 6.88 6.88 0 0 1-6.76 6.81 6.66 6.66 0 1 1 0-13.31Z"
+				/>
+				<path
+					fill="#310d00"
+					d="M47.58 24.94a6.63 6.63 0 0 1 6.77 6.49 6.89 6.89 0 0 1-6.77 6.81 6.65 6.65 0 1 1 0-13.3Z"
+				/>
+				<path
+					fill="#330e00"
+					d="M47.59 25a6.63 6.63 0 0 1 6.77 6.48 6.9 6.9 0 0 1-6.77 6.82 6.65 6.65 0 0 1 0-13.3Z"
+				/>
+				<path
+					fill="#350e00"
+					d="M47.6 25a6.63 6.63 0 0 1 6.77 6.47 6.91 6.91 0 0 1-6.77 6.83 6.65 6.65 0 0 1 0-13.3Z"
+				/>
+				<path
+					fill="#360f00"
+					d="M47.6 25a6.63 6.63 0 0 1 6.78 6.47 6.91 6.91 0 0 1-6.78 6.83 6.65 6.65 0 0 1 0-13.3Z"
+				/>
+				<path
+					fill="#380f00"
+					d="M47.61 25a6.63 6.63 0 0 1 6.78 6.46 6.93 6.93 0 0 1-6.78 6.84 6.65 6.65 0 1 1 0-13.3Z"
+				/>
+				<path
+					fill="#3a1000"
+					d="M47.62 25a6.63 6.63 0 0 1 6.78 6.46 6.92 6.92 0 0 1-6.78 6.83 6.65 6.65 0 1 1 0-13.29Z"
+				/>
+				<path
+					fill="#3b1000"
+					d="M47.63 25a6.62 6.62 0 0 1 6.78 6.45 6.93 6.93 0 0 1-6.78 6.84 6.65 6.65 0 1 1 0-13.29Z"
+				/>
+				<path
+					fill="#3d1100"
+					d="M47.63 25a6.63 6.63 0 0 1 6.79 6.45 6.94 6.94 0 0 1-6.79 6.84 6.65 6.65 0 0 1 0-13.29Z"
+				/>
+				<path
+					fill="#3f1100"
+					d="M47.64 25a6.63 6.63 0 0 1 6.79 6.44 7 7 0 0 1-6.79 6.85 6.65 6.65 0 0 1 0-13.29Z"
+				/>
+				<path
+					fill="#411200"
+					d="M47.65 25a6.64 6.64 0 0 1 6.79 6.43 7 7 0 0 1-6.79 6.86 6.65 6.65 0 0 1 0-13.29Z"
+				/>
+				<path
+					fill="#421200"
+					d="M47.65 25.05a6.62 6.62 0 0 1 6.8 6.42 7 7 0 0 1-6.8 6.86 6.64 6.64 0 0 1 0-13.28Z"
+				/>
+				<path
+					fill="#441300"
+					d="M47.66 25.06a6.62 6.62 0 0 1 6.8 6.41 7 7 0 0 1-6.8 6.87 6.64 6.64 0 1 1 0-13.28Z"
+				/>
+				<path
+					fill="#461300"
+					d="M47.67 25.07a6.62 6.62 0 0 1 6.8 6.41 7 7 0 0 1-6.8 6.86 6.64 6.64 0 1 1 0-13.27Z"
+				/>
+				<path
+					fill="#471300"
+					d="M47.68 25.08a6.62 6.62 0 0 1 6.8 6.4 7 7 0 0 1-6.8 6.87 6.64 6.64 0 1 1 0-13.27Z"
+				/>
+				<path
+					fill="#491400"
+					d="M47.68 25.09a6.61 6.61 0 0 1 6.81 6.39 7 7 0 0 1-6.81 6.88 6.64 6.64 0 0 1 0-13.27Z"
+				/>
+				<path
+					fill="#4b1400"
+					d="M47.69 25.1a6.61 6.61 0 0 1 6.81 6.39 7 7 0 0 1-6.81 6.88 6.64 6.64 0 0 1 0-13.27Z"
+				/>
+				<path
+					fill="#4c1500"
+					d="M47.7 25.11a6.61 6.61 0 0 1 6.81 6.38 7 7 0 0 1-6.81 6.89 6.64 6.64 0 1 1 0-13.27Z"
+				/>
+				<path
+					fill="#4e1500"
+					d="M47.7 25.12c3.67.06 6.82 2.71 6.82 6.38a7 7 0 0 1-6.82 6.89 6.64 6.64 0 0 1 0-13.27Z"
+				/>
+				<path
+					fill="#501600"
+					d="M47.71 25.13c3.66.06 6.82 2.71 6.82 6.37a7 7 0 0 1-6.82 6.9 6.64 6.64 0 0 1 0-13.27Z"
+				/>
+				<path
+					fill="#521600"
+					d="M47.72 25.15c3.66.05 6.82 2.69 6.82 6.36a7 7 0 0 1-6.82 6.9 6.63 6.63 0 0 1 0-13.26Z"
+				/>
+				<path
+					fill="#531700"
+					d="M47.73 25.16c3.66.05 6.82 2.69 6.82 6.35a7 7 0 0 1-6.82 6.9 6.63 6.63 0 1 1 0-13.25Z"
+				/>
+				<path
+					fill="#551700"
+					d="M47.73 25.17c3.66 0 6.83 2.68 6.83 6.34a7 7 0 0 1-6.83 6.91 6.63 6.63 0 0 1 0-13.25Z"
+				/>
+				<path
+					fill="#571800"
+					d="M47.74 25.18c3.66.06 6.83 2.68 6.83 6.34a7 7 0 0 1-6.83 6.91 6.63 6.63 0 1 1 0-13.25Z"
+				/>
+				<path
+					fill="#581800"
+					d="M47.75 25.19c3.66.06 6.83 2.67 6.83 6.33a7.06 7.06 0 0 1-6.83 6.92 6.63 6.63 0 1 1 0-13.25Z"
+				/>
+				<path
+					fill="#5a1900"
+					d="M47.75 25.2c3.66.06 6.84 2.67 6.84 6.33a7.07 7.07 0 0 1-6.84 6.92 6.63 6.63 0 0 1 0-13.25Z"
+				/>
+				<path
+					fill="#5c1900"
+					d="M47.76 25.21c3.66.06 6.84 2.66 6.84 6.32a7.08 7.08 0 0 1-6.84 6.93 6.63 6.63 0 0 1 0-13.25Z"
+				/>
+				<path
+					fill="#5e1a00"
+					d="M47.77 25.22c3.65.06 6.84 2.66 6.84 6.32a7.08 7.08 0 0 1-6.84 6.93 6.63 6.63 0 0 1 0-13.25Z"
+				/>
+				<path
+					fill="#5f1a00"
+					d="M47.78 25.23c3.65.07 6.84 2.65 6.84 6.31a7.08 7.08 0 0 1-6.84 6.93 6.62 6.62 0 0 1 0-13.24Z"
+				/>
+				<path
+					fill="#611a00"
+					d="M47.78 25.24c3.66.07 6.85 2.65 6.85 6.3a7.09 7.09 0 0 1-6.85 6.94 6.62 6.62 0 0 1 0-13.24Z"
+				/>
+				<path
+					fill="#631b00"
+					d="M47.79 25.26c3.65.06 6.85 2.63 6.85 6.29a7.1 7.1 0 0 1-6.85 6.94 6.62 6.62 0 1 1 0-13.23Z"
+				/>
+				<path
+					fill="#641b00"
+					d="M47.8 25.27c3.65.06 6.85 2.63 6.85 6.28a7.11 7.11 0 0 1-6.85 7 6.62 6.62 0 1 1 0-13.23Z"
+				/>
+				<path
+					fill="#661c00"
+					d="M47.8 25.28c3.66.07 6.86 2.62 6.86 6.28a7.12 7.12 0 0 1-6.86 7 6.62 6.62 0 0 1 0-13.23Z"
+				/>
+				<path
+					fill="#681c00"
+					d="M47.81 25.29c3.65.07 6.86 2.62 6.86 6.27a7.14 7.14 0 0 1-6.86 7 6.62 6.62 0 0 1 0-13.23Z"
+				/>
+				<path
+					fill="#691d00"
+					d="M47.82 25.3c3.65.07 6.86 2.61 6.86 6.27a7.14 7.14 0 0 1-6.86 7 6.62 6.62 0 0 1 0-13.23Z"
+				/>
+				<path
+					fill="#6b1d00"
+					d="M47.83 25.31c3.65.07 6.86 2.61 6.86 6.26a7.14 7.14 0 0 1-6.86 7 6.57 6.57 0 0 1-6.62-6.61 6.64 6.64 0 0 1 6.62-6.65Z"
+				/>
+				<path
+					fill="#6d1e00"
+					d="M47.83 25.32c3.65.07 6.87 2.6 6.87 6.25a7.15 7.15 0 0 1-6.87 7 6.57 6.57 0 0 1-6.61-6.61 6.64 6.64 0 0 1 6.61-6.64Z"
+				/>
+				<path
+					fill="#6f1e00"
+					d="M47.84 25.33c3.65.08 6.87 2.6 6.87 6.25a7.15 7.15 0 0 1-6.87 7 6.57 6.57 0 0 1-6.61-6.61 6.64 6.64 0 0 1 6.61-6.64Z"
+				/>
+				<path
+					fill="#701f00"
+					d="M47.85 25.34c3.65.08 6.87 2.59 6.87 6.24a7.16 7.16 0 0 1-6.87 7A6.57 6.57 0 0 1 41.24 32a6.64 6.64 0 0 1 6.61-6.66Z"
+				/>
+				<path
+					fill="#721f00"
+					d="M47.85 25.35c3.65.08 6.88 2.59 6.88 6.24a7.17 7.17 0 0 1-6.88 7 6.56 6.56 0 0 1-6.6-6.59 6.63 6.63 0 0 1 6.6-6.65Z"
+				/>
+				<path
+					fill="#742000"
+					d="M47.86 25.37c3.65.07 6.88 2.57 6.88 6.22a7.18 7.18 0 0 1-6.88 7 6.61 6.61 0 0 1 0-13.21Z"
+				/>
+				<path
+					fill="#752000"
+					d="M47.87 25.38c3.65.08 6.88 2.57 6.88 6.22a7.19 7.19 0 0 1-6.88 7 6.61 6.61 0 1 1 0-13.21Z"
+				/>
+				<path
+					fill="#772000"
+					d="M47.88 25.39c3.64.08 6.88 2.56 6.88 6.21a7.2 7.2 0 0 1-6.88 7 6.61 6.61 0 1 1 0-13.21Z"
+				/>
+				<path
+					fill="#792100"
+					d="M47.88 25.4c3.65.08 6.89 2.56 6.89 6.2a7.2 7.2 0 0 1-6.89 7 6.56 6.56 0 0 1-6.6-6.6 6.62 6.62 0 0 1 6.6-6.6Z"
+				/>
+				<path
+					fill="#7a2100"
+					d="M47.89 25.41c3.64.08 6.89 2.55 6.89 6.2a7.2 7.2 0 0 1-6.89 7 6.55 6.55 0 0 1-6.6-6.6 6.62 6.62 0 0 1 6.6-6.6Z"
+				/>
+				<path
+					fill="#7c2200"
+					d="M47.9 25.42c3.64.08 6.89 2.55 6.89 6.19a7.21 7.21 0 0 1-6.89 7A6.54 6.54 0 0 1 41.3 32a6.62 6.62 0 0 1 6.6-6.58Z"
+				/>
+				<path
+					fill="#7e2200"
+					d="M47.9 25.43c3.65.09 6.89 2.54 6.89 6.19a7.21 7.21 0 0 1-6.89 7A6.55 6.55 0 0 1 41.31 32a6.62 6.62 0 0 1 6.59-6.57Z"
+				/>
+				<path
+					fill="#802300"
+					d="M47.91 25.44c3.64.09 6.89 2.54 6.89 6.18a7.22 7.22 0 0 1-6.89 7 6.56 6.56 0 0 1-6.6-6.6 6.64 6.64 0 0 1 6.6-6.58Z"
+				/>
+				<path
+					fill="#812300"
+					d="M47.92 25.45c3.64.09 6.89 2.53 6.89 6.17a7.23 7.23 0 0 1-6.89 7 6.56 6.56 0 0 1-6.6-6.6 6.64 6.64 0 0 1 6.6-6.57Z"
+				/>
+				<path
+					fill="#832400"
+					d="M47.93 25.47c3.64.08 6.89 2.52 6.89 6.16a7.24 7.24 0 0 1-6.89 7 6.55 6.55 0 0 1-6.6-6.6 6.62 6.62 0 0 1 6.6-6.56Z"
+				/>
+				<path
+					fill="#852400"
+					d="M47.93 25.48c3.64.08 6.9 2.51 6.9 6.15a7.24 7.24 0 0 1-6.9 7 6.54 6.54 0 0 1-6.59-6.59 6.61 6.61 0 0 1 6.59-6.56Z"
+				/>
+				<path
+					fill="#862500"
+					d="M47.94 25.49c3.64.09 6.9 2.51 6.9 6.15a7.25 7.25 0 0 1-6.9 7 6.54 6.54 0 0 1-6.59-6.59 6.61 6.61 0 0 1 6.59-6.56Z"
+				/>
+				<path
+					fill="#882500"
+					d="M48 25.5c3.64.09 6.9 2.5 6.9 6.14a7.26 7.26 0 0 1-6.9 7 6.53 6.53 0 0 1-6.59-6.59A6.61 6.61 0 0 1 48 25.5Z"
+				/>
+				<path
+					fill="#8a2600"
+					d="M48 25.51c3.64.09 6.91 2.5 6.91 6.14a7.27 7.27 0 0 1-6.91 7 6.54 6.54 0 0 1-6.59-6.59A6.63 6.63 0 0 1 48 25.51Z"
+				/>
+				<path
+					fill="#8b2600"
+					d="M48 25.52c3.64.09 6.91 2.49 6.91 6.13A7.28 7.28 0 0 1 48 38.7a6.54 6.54 0 0 1-6.59-6.59A6.63 6.63 0 0 1 48 25.52Z"
+				/>
+				<path
+					fill="#8d2600"
+					d="M48 25.53c3.64.1 6.91 2.49 6.91 6.12A7.28 7.28 0 0 1 48 38.71a6.54 6.54 0 0 1-6.59-6.59A6.63 6.63 0 0 1 48 25.53Z"
+				/>
+				<path
+					fill="#8f2700"
+					d="M48 25.54c3.63.1 6.91 2.48 6.91 6.12A7.29 7.29 0 0 1 48 38.72a6.53 6.53 0 0 1-6.59-6.59A6.62 6.62 0 0 1 48 25.54Z"
+				/>
+				<path
+					fill="#912700"
+					d="M48 25.55c3.64.1 6.92 2.48 6.92 6.11A7.3 7.3 0 0 1 48 38.73a6.53 6.53 0 0 1-6.58-6.59A6.62 6.62 0 0 1 48 25.55Z"
+				/>
+				<path
+					fill="#922800"
+					d="M48 25.56c3.63.1 6.92 2.47 6.92 6.11A7.3 7.3 0 0 1 48 38.73a6.54 6.54 0 0 1-6.59-6.58A6.63 6.63 0 0 1 48 25.56Z"
+				/>
+				<path
+					fill="#942800"
+					d="M48 25.58c3.63.09 6.92 2.46 6.92 6.09A7.31 7.31 0 0 1 48 38.74a6.54 6.54 0 0 1-6.59-6.58A6.62 6.62 0 0 1 48 25.58Z"
+				/>
+				<path
+					fill="#962900"
+					d="M48 25.59c3.64.1 6.93 2.45 6.93 6.09A7.32 7.32 0 0 1 48 38.75a6.53 6.53 0 0 1-6.58-6.58A6.62 6.62 0 0 1 48 25.59Z"
+				/>
+				<path
+					fill="#972900"
+					d="M48 25.6c3.63.1 6.93 2.45 6.93 6.08A7.33 7.33 0 0 1 48 38.76a6.53 6.53 0 0 1-6.58-6.58A6.62 6.62 0 0 1 48 25.6Z"
+				/>
+				<path
+					fill="#992a00"
+					d="M48 25.61c3.63.1 6.93 2.44 6.93 6.07A7.34 7.34 0 0 1 48 38.77a6.52 6.52 0 0 1-6.58-6.58A6.62 6.62 0 0 1 48 25.61Z"
+				/>
+				<path
+					fill="#9b2a00"
+					d="M48 25.62c3.63.1 6.93 2.43 6.93 6.07A7.35 7.35 0 0 1 48 38.78a6.52 6.52 0 0 1-6.58-6.58A6.62 6.62 0 0 1 48 25.62Z"
+				/>
+				<path
+					fill="#9c2b00"
+					d="M48 25.63c3.63.11 6.94 2.43 6.94 6.06a7.35 7.35 0 0 1-6.94 7.1 6.53 6.53 0 0 1-6.58-6.58A6.62 6.62 0 0 1 48 25.63Z"
+				/>
+				<path
+					fill="#9e2b00"
+					d="M48 25.64c3.63.11 6.94 2.42 6.94 6.06A7.35 7.35 0 0 1 48 38.79a6.52 6.52 0 0 1-6.58-6.57A6.62 6.62 0 0 1 48 25.64Z"
+				/>
+				<path
+					fill="#a02c00"
+					d="M48.05 25.65c3.63.11 6.94 2.42 6.94 6.05a7.36 7.36 0 0 1-6.94 7.1 6.52 6.52 0 0 1-6.58-6.57 6.62 6.62 0 0 1 6.58-6.58Z"
+				/>
+				<path
+					fill="#a22c00"
+					d="M48.05 25.66c3.63.11 7 2.41 7 6.05a7.37 7.37 0 0 1-7 7.1 6.51 6.51 0 0 1-6.57-6.57 6.61 6.61 0 0 1 6.57-6.58Z"
+				/>
+				<path
+					fill="#a32d00"
+					d="M48.06 25.67c3.63.11 6.95 2.41 6.95 6a7.38 7.38 0 0 1-6.95 7.11 6.51 6.51 0 0 1-6.57-6.57 6.61 6.61 0 0 1 6.57-6.54Z"
+				/>
+				<path
+					fill="#a52d00"
+					d="M48.07 25.69c3.63.11 7 2.39 7 6a7.39 7.39 0 0 1-7 7.12 6.51 6.51 0 0 1-6.57-6.57 6.6 6.6 0 0 1 6.57-6.55Z"
+				/>
+				<path
+					fill="#a72d00"
+					d="M48.08 25.7c3.62.11 7 2.39 7 6a7.4 7.4 0 0 1-7 7.12 6.53 6.53 0 0 1-6.58-6.57 6.61 6.61 0 0 1 6.58-6.55Z"
+				/>
+				<path
+					fill="#a82e00"
+					d="M48.08 25.71c3.63.11 7 2.38 7 6a7.41 7.41 0 0 1-7 7.13 6.51 6.51 0 0 1-6.57-6.57 6.6 6.6 0 0 1 6.57-6.56Z"
+				/>
+				<path
+					fill="#aa2e00"
+					d="M48.09 25.72c3.63.11 7 2.38 7 6a7.41 7.41 0 0 1-7 7.12 6.5 6.5 0 0 1-6.57-6.56 6.6 6.6 0 0 1 6.57-6.56Z"
+				/>
+				<path
+					fill="#ac2f00"
+					d="M48.1 25.73c3.62.11 7 2.37 7 6a7.42 7.42 0 0 1-7 7.13 6.5 6.5 0 0 1-6.57-6.56 6.6 6.6 0 0 1 6.57-6.57Z"
+				/>
+				<path
+					fill="#ad2f00"
+					d="M48.1 25.74c3.63.12 7 2.37 7 6a7.43 7.43 0 0 1-7 7.14 6.5 6.5 0 0 1-6.56-6.56 6.6 6.6 0 0 1 6.56-6.58Z"
+				/>
+				<path
+					fill="#af3000"
+					d="M48.11 25.75c3.62.12 7 2.36 7 6a7.44 7.44 0 0 1-7 7.14 6.49 6.49 0 0 1-6.56-6.56 6.6 6.6 0 0 1 6.56-6.58Z"
+				/>
+				<path
+					fill="#b13000"
+					d="M48.12 25.76c3.62.12 7 2.36 7 6a7.44 7.44 0 0 1-7 7.15 6.5 6.5 0 0 1-6.57-6.56 6.62 6.62 0 0 1 6.57-6.59Z"
+				/>
+				<path
+					fill="#b33100"
+					d="M48.13 25.77c3.62.12 7 2.35 7 6s-3.35 7.15-7 7.15a6.5 6.5 0 0 1-6.57-6.56 6.62 6.62 0 0 1 6.57-6.59Z"
+				/>
+				<path
+					fill="#b43100"
+					d="M48.13 25.79c3.62.12 7 2.34 7 6a7.46 7.46 0 0 1-7 7.16 6.5 6.5 0 0 1-6.56-6.56 6.59 6.59 0 0 1 6.56-6.6Z"
+				/>
+				<path
+					fill="#b63200"
+					d="M48.14 25.8c3.62.12 7 2.33 7 6s-3.36 7.16-7 7.16a6.49 6.49 0 0 1-6.56-6.56 6.59 6.59 0 0 1 6.56-6.6Z"
+				/>
+				<path
+					fill="#b83200"
+					d="M48.15 25.81c3.62.12 7 2.33 7 6s-3.36 7.16-7 7.16a6.48 6.48 0 0 1-6.56-6.55 6.59 6.59 0 0 1 6.56-6.61Z"
+				/>
+				<path
+					fill="#b93300"
+					d="M48.15 25.82c3.62.12 7 2.32 7 5.94s-3.36 7.17-7 7.17a6.48 6.48 0 0 1-6.55-6.55 6.59 6.59 0 0 1 6.55-6.56Z"
+				/>
+				<path
+					fill="#b30"
+					d="M55.15 31.77c0 3.62-3.37 7.17-7 7.17a6.49 6.49 0 0 1-6.56-6.55 6.61 6.61 0 0 1 6.56-6.56c3.63.17 7 2.32 7 5.94Z"
+				/>
+				<path
+					fill="#bc3502"
+					d="M48.17 25.84c3.62.12 7 2.32 7 5.94s-3.38 7.22-7 7.22a6.49 6.49 0 0 1-6.56-6.56 6.6 6.6 0 0 1 6.56-6.6Z"
+					opacity={0.99}
+				/>
+				<path
+					fill="#bc3705"
+					d="M48.17 25.85c3.62.12 7 2.33 7 5.95s-3.38 7.2-7 7.2a6.49 6.49 0 0 1-6.55-6.56 6.59 6.59 0 0 1 6.55-6.59Z"
+					opacity={0.98}
+				/>
+				<path
+					fill="#bd3907"
+					d="M48.18 25.85c3.62.13 7 2.34 7 6S51.8 39 48.18 39a6.49 6.49 0 0 1-6.56-6.56 6.61 6.61 0 0 1 6.56-6.59Z"
+					opacity={0.97}
+				/>
+				<path
+					fill="#bd3a09"
+					d="M48.18 25.86c3.62.12 7 2.34 7 6a7.45 7.45 0 0 1-7 7.15 6.48 6.48 0 0 1-6.55-6.55 6.6 6.6 0 0 1 6.55-6.6Z"
+					opacity={0.96}
+				/>
+				<path
+					fill="#be3c0c"
+					d="M48.19 25.87c3.62.12 7 2.35 7 6a7.44 7.44 0 0 1-7 7.14 6.49 6.49 0 0 1-6.56-6.55 6.61 6.61 0 0 1 6.56-6.59Z"
+					opacity={0.95}
+				/>
+				<path
+					fill="#bf3e0e"
+					d="M48.19 25.88c3.62.12 7 2.35 7 6a7.44 7.44 0 0 1-7 7.14 6.48 6.48 0 0 1-6.55-6.55 6.6 6.6 0 0 1 6.55-6.59Z"
+					opacity={0.95}
+				/>
+				<path
+					fill="#bf4010"
+					d="M48.2 25.89c3.62.12 7 2.36 7 6a7.42 7.42 0 0 1-7 7.13 6.5 6.5 0 0 1-6.56-6.56 6.6 6.6 0 0 1 6.56-6.57Z"
+					opacity={0.94}
+				/>
+				<path
+					fill="#c04213"
+					d="M48.21 25.9c3.61.11 6.95 2.36 6.95 6a7.42 7.42 0 0 1-6.95 7.1 6.5 6.5 0 0 1-6.56-6.56 6.59 6.59 0 0 1 6.56-6.54Z"
+					opacity={0.93}
+				/>
+				<path
+					fill="#c14415"
+					d="M48.21 25.9c3.62.12 7 2.37 7 6a7.42 7.42 0 0 1-7 7.13 6.49 6.49 0 0 1-6.55-6.56 6.59 6.59 0 0 1 6.55-6.57Z"
+					opacity={0.92}
+				/>
+				<path
+					fill="#c14617"
+					d="M48.22 25.91c3.62.12 7 2.38 7 6a7.41 7.41 0 0 1-7 7.12 6.5 6.5 0 0 1-6.56-6.56 6.59 6.59 0 0 1 6.56-6.56Z"
+					opacity={0.91}
+				/>
+				<path
+					fill="#c24719"
+					d="M48.22 25.92c3.62.11 7 2.38 7 6a7.4 7.4 0 0 1-7 7.11 6.48 6.48 0 0 1-6.55-6.55 6.59 6.59 0 0 1 6.55-6.56Z"
+					opacity={0.9}
+				/>
+				<path
+					fill="#c2491c"
+					d="M48.23 25.93c3.62.11 6.94 2.39 6.94 6A7.38 7.38 0 0 1 48.23 39a6.49 6.49 0 0 1-6.56-6.55 6.59 6.59 0 0 1 6.56-6.52Z"
+					opacity={0.89}
+				/>
+				<path
+					fill="#c34b1e"
+					d="M48.23 25.94c3.62.11 6.94 2.39 6.94 6a7.38 7.38 0 0 1-6.94 7.1 6.49 6.49 0 0 1-6.55-6.56 6.59 6.59 0 0 1 6.55-6.54Z"
+					opacity={0.88}
+				/>
+				<path
+					fill="#c44d20"
+					d="M48.24 26c3.62.11 6.93 2.39 6.93 6a7.36 7.36 0 0 1-6.93 7.1 6.51 6.51 0 0 1-6.56-6.56A6.59 6.59 0 0 1 48.24 26Z"
+					opacity={0.87}
+				/>
+				<path
+					fill="#c44f23"
+					d="M48.25 26c3.61.1 6.93 2.4 6.93 6a7.36 7.36 0 0 1-6.93 7.09 6.49 6.49 0 0 1-6.56-6.56A6.59 6.59 0 0 1 48.25 26Z"
+					opacity={0.86}
+				/>
+				<path
+					fill="#c55125"
+					d="M48.25 26c3.62.11 6.93 2.41 6.93 6a7.36 7.36 0 0 1-6.93 7.09 6.51 6.51 0 0 1-6.56-6.56A6.61 6.61 0 0 1 48.25 26Z"
+					opacity={0.85}
+				/>
+				<path
+					fill="#c65327"
+					d="M48.26 26c3.62.11 6.92 2.42 6.92 6a7.33 7.33 0 0 1-6.92 7.07 6.5 6.5 0 0 1-6.56-6.55A6.6 6.6 0 0 1 48.26 26Z"
+					opacity={0.85}
+				/>
+				<path
+					fill="#c6542a"
+					d="M48.26 26c3.62.11 6.92 2.42 6.92 6a7.33 7.33 0 0 1-6.92 7.07 6.48 6.48 0 0 1-6.55-6.55A6.59 6.59 0 0 1 48.26 26Z"
+					opacity={0.84}
+				/>
+				<path
+					fill="#c7562c"
+					d="M48.27 26c3.62.1 6.91 2.42 6.91 6a7.32 7.32 0 0 1-6.91 7.07 6.51 6.51 0 0 1-6.56-6.56A6.59 6.59 0 0 1 48.27 26Z"
+					opacity={0.83}
+				/>
+				<path
+					fill="#c7582e"
+					d="M48.27 26c3.62.1 6.91 2.43 6.91 6a7.31 7.31 0 0 1-6.91 7.06 6.5 6.5 0 0 1-6.55-6.56 6.58 6.58 0 0 1 6.55-6.5Z"
+					opacity={0.82}
+				/>
+				<path
+					fill="#c85a31"
+					d="M48.28 26c3.62.1 6.91 2.43 6.91 6.05a7.31 7.31 0 0 1-6.91 7.06 6.51 6.51 0 0 1-6.56-6.56A6.59 6.59 0 0 1 48.28 26Z"
+					opacity={0.81}
+				/>
+				<path
+					fill="#c95c33"
+					d="M48.28 26c3.62.1 6.91 2.45 6.91 6.07a7.3 7.3 0 0 1-6.91 7.05 6.5 6.5 0 0 1-6.55-6.56A6.59 6.59 0 0 1 48.28 26Z"
+					opacity={0.8}
+				/>
+				<path
+					fill="#c95e35"
+					d="M48.29 26c3.62.1 6.9 2.45 6.9 6.07a7.29 7.29 0 0 1-6.9 7 6.5 6.5 0 0 1-6.56-6.55A6.6 6.6 0 0 1 48.29 26Z"
+					opacity={0.79}
+				/>
+				<path
+					fill="#ca6038"
+					d="M48.3 26c3.61.1 6.89 2.45 6.89 6.07a7.27 7.27 0 0 1-6.89 7 6.49 6.49 0 0 1-6.56-6.55A6.59 6.59 0 0 1 48.3 26Z"
+					opacity={0.78}
+				/>
+				<path
+					fill="#ca613a"
+					d="M48.3 26c3.62.1 6.89 2.46 6.89 6.08a7.27 7.27 0 0 1-6.89 7 6.51 6.51 0 0 1-6.56-6.56A6.59 6.59 0 0 1 48.3 26Z"
+					opacity={0.77}
+				/>
+				<path
+					fill="#cb633c"
+					d="M48.31 26.05c3.62.09 6.88 2.46 6.88 6.08a7.26 7.26 0 0 1-6.88 7 6.51 6.51 0 0 1-6.56-6.56 6.59 6.59 0 0 1 6.56-6.52Z"
+					opacity={0.76}
+				/>
+				<path
+					fill="#cc653f"
+					d="M48.31 26.06c3.62.09 6.89 2.47 6.89 6.09a7.26 7.26 0 0 1-6.89 7 6.5 6.5 0 0 1-6.55-6.56 6.58 6.58 0 0 1 6.55-6.53Z"
+					opacity={0.75}
+				/>
+				<path
+					fill="#cc6741"
+					d="M48.32 26.06c3.62.1 6.88 2.48 6.88 6.1a7.25 7.25 0 0 1-6.88 7 6.51 6.51 0 0 1-6.56-6.56 6.6 6.6 0 0 1 6.56-6.54Z"
+					opacity={0.75}
+				/>
+				<path
+					fill="#cd6943"
+					d="M48.32 26.07c3.62.09 6.88 2.48 6.88 6.1a7.23 7.23 0 0 1-6.88 7 6.5 6.5 0 0 1-6.55-6.55 6.59 6.59 0 0 1 6.55-6.55Z"
+					opacity={0.74}
+				/>
+				<path
+					fill="#ce6b46"
+					d="M48.33 26.08c3.62.09 6.87 2.49 6.87 6.11a7.22 7.22 0 0 1-6.87 7 6.5 6.5 0 0 1-6.56-6.55 6.59 6.59 0 0 1 6.56-6.56Z"
+					opacity={0.73}
+				/>
+				<path
+					fill="#ce6c48"
+					d="M48.33 26.09c3.62.09 6.87 2.49 6.87 6.11a7.22 7.22 0 0 1-6.87 7 6.5 6.5 0 0 1-6.55-6.55 6.58 6.58 0 0 1 6.55-6.56Z"
+					opacity={0.72}
+				/>
+				<path
+					fill="#cf6e4a"
+					d="M48.34 26.1c3.62.09 6.86 2.5 6.86 6.12a7.21 7.21 0 0 1-6.86 7 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.71}
+				/>
+				<path
+					fill="#cf704c"
+					d="M48.35 26.11c3.62.08 6.86 2.5 6.86 6.12a7.21 7.21 0 0 1-6.86 7 6.51 6.51 0 0 1-6.56-6.56 6.58 6.58 0 0 1 6.56-6.56Z"
+					opacity={0.7}
+				/>
+				<path
+					fill="#d0724f"
+					d="M48.35 26.11c3.62.09 6.86 2.51 6.86 6.13a7.2 7.2 0 0 1-6.86 7 6.5 6.5 0 0 1-6.55-6.56 6.58 6.58 0 0 1 6.55-6.57Z"
+					opacity={0.69}
+				/>
+				<path
+					fill="#d17451"
+					d="M48.36 26.12c3.62.09 6.85 2.52 6.85 6.14a7.19 7.19 0 0 1-6.85 7 6.52 6.52 0 0 1-6.56-6.56 6.59 6.59 0 0 1 6.56-6.58Z"
+					opacity={0.68}
+				/>
+				<path
+					fill="#d17653"
+					d="M48.36 26.13c3.62.09 6.85 2.52 6.85 6.14a7.18 7.18 0 0 1-6.85 7 6.5 6.5 0 0 1-6.55-6.55 6.58 6.58 0 0 1 6.55-6.59Z"
+					opacity={0.67}
+				/>
+				<path
+					fill="#d27856"
+					d="M48.37 26.14c3.62.08 6.84 2.53 6.84 6.15a7.16 7.16 0 0 1-6.84 7 6.51 6.51 0 0 1-6.56-6.55 6.59 6.59 0 0 1 6.56-6.6Z"
+					opacity={0.66}
+				/>
+				<path
+					fill="#d27958"
+					d="M48.37 26.15c3.62.08 6.84 2.53 6.84 6.15a7.16 7.16 0 0 1-6.84 7 6.51 6.51 0 0 1-6.55-6.56 6.58 6.58 0 0 1 6.55-6.59Z"
+					opacity={0.65}
+				/>
+				<path
+					fill="#d37b5a"
+					d="M48.38 26.16c3.62.08 6.84 2.53 6.84 6.15a7.15 7.15 0 0 1-6.84 7 6.52 6.52 0 0 1-6.56-6.56 6.59 6.59 0 0 1 6.56-6.59Z"
+					opacity={0.65}
+				/>
+				<path
+					fill="#d47d5d"
+					d="M48.38 26.16c3.62.08 6.84 2.55 6.84 6.17a7.14 7.14 0 0 1-6.84 7 6.51 6.51 0 0 1-6.55-6.56 6.59 6.59 0 0 1 6.55-6.61Z"
+					opacity={0.64}
+				/>
+				<path
+					fill="#d47f5f"
+					d="M48.39 26.17c3.62.08 6.83 2.55 6.83 6.17a7.14 7.14 0 0 1-6.83 6.95 6.53 6.53 0 0 1-6.56-6.56 6.6 6.6 0 0 1 6.56-6.56Z"
+					opacity={0.63}
+				/>
+				<path
+					fill="#d58161"
+					d="M48.4 26.18c3.62.08 6.82 2.56 6.82 6.18a7.12 7.12 0 0 1-6.82 6.93 6.51 6.51 0 0 1-6.56-6.55 6.58 6.58 0 0 1 6.56-6.56Z"
+					opacity={0.62}
+				/>
+				<path
+					fill="#d68364"
+					d="M48.4 26.19c3.62.08 6.82 2.56 6.82 6.18a7.12 7.12 0 0 1-6.82 6.93 6.5 6.5 0 0 1-6.55-6.55 6.57 6.57 0 0 1 6.55-6.56Z"
+					opacity={0.61}
+				/>
+				<path
+					fill="#d68566"
+					d="M48.41 26.2c3.62.07 6.81 2.56 6.81 6.18a7.1 7.1 0 0 1-6.81 6.93 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.6}
+				/>
+				<path
+					fill="#d78668"
+					d="M48.41 26.21c3.62.07 6.82 2.57 6.82 6.19a7.11 7.11 0 0 1-6.82 6.92 6.56 6.56 0 0 1 0-13.11Z"
+					opacity={0.59}
+				/>
+				<path
+					fill="#d7886b"
+					d="M48.42 26.22c3.62.07 6.81 2.57 6.81 6.19a7.1 7.1 0 0 1-6.81 6.92 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.58}
+				/>
+				<path
+					fill="#d88a6d"
+					d="M48.42 26.22c3.62.08 6.81 2.59 6.81 6.21a7.08 7.08 0 0 1-6.81 6.91 6.51 6.51 0 0 1-6.55-6.56 6.59 6.59 0 0 1 6.55-6.56Z"
+					opacity={0.57}
+				/>
+				<path
+					fill="#d98c6f"
+					d="M48.43 26.23c3.62.07 6.8 2.59 6.8 6.21a7.07 7.07 0 0 1-6.8 6.9 6.52 6.52 0 0 1-6.56-6.55 6.59 6.59 0 0 1 6.56-6.56Z"
+					opacity={0.56}
+				/>
+				<path
+					fill="#d98e72"
+					d="M48.44 26.24c3.62.07 6.79 2.59 6.79 6.21a7.05 7.05 0 0 1-6.79 6.9 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.55}
+				/>
+				<path
+					fill="#da9074"
+					d="M48.44 26.25c3.62.07 6.79 2.6 6.79 6.22a7.05 7.05 0 0 1-6.79 6.89 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.55}
+				/>
+				<path
+					fill="#db9276"
+					d="M48.45 26.26c3.62.06 6.78 2.6 6.78 6.22a7 7 0 0 1-6.78 6.89 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.54}
+				/>
+				<path
+					fill="#db9379"
+					d="M48.45 26.27c3.62.06 6.79 2.61 6.79 6.23a7 7 0 0 1-6.79 6.88 6.56 6.56 0 0 1 0-13.11Z"
+					opacity={0.53}
+				/>
+				<path
+					fill="#dc957b"
+					d="M48.46 26.27c3.62.07 6.78 2.62 6.78 6.24a7 7 0 0 1-6.78 6.88 6.56 6.56 0 0 1 0-13.12Z"
+					opacity={0.52}
+				/>
+				<path
+					fill="#dc977d"
+					d="M48.46 26.28c3.62.07 6.78 2.62 6.78 6.24a7 7 0 0 1-6.78 6.87 6.56 6.56 0 0 1 0-13.11Z"
+					opacity={0.51}
+				/>
+				<path
+					fill="#dd9980"
+					d="M48.47 26.29c3.62.06 6.77 2.63 6.77 6.25a7 7 0 0 1-6.77 6.86 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.5}
+				/>
+				<path
+					fill="#de9b82"
+					d="M48.47 26.3c3.62.06 6.77 2.63 6.77 6.25a7 7 0 0 1-6.77 6.86 6.56 6.56 0 0 1 0-13.11Z"
+					opacity={0.49}
+				/>
+				<path
+					fill="#de9d84"
+					d="M48.48 26.31c3.62.06 6.76 2.64 6.76 6.26a7 7 0 0 1-6.76 6.85 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.48}
+				/>
+				<path
+					fill="#df9f86"
+					d="M48.49 26.32c3.62.06 6.76 2.64 6.76 6.26a7 7 0 0 1-6.76 6.85 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.47}
+				/>
+				<path
+					fill="#dfa089"
+					d="M48.49 26.32c3.62.06 6.76 2.65 6.76 6.27a7 7 0 0 1-6.76 6.85 6.56 6.56 0 0 1 0-13.12Z"
+					opacity={0.46}
+				/>
+				<path
+					fill="#e0a28b"
+					d="M48.5 26.33c3.62.06 6.75 2.66 6.75 6.28a7 7 0 0 1-6.75 6.83 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.45}
+				/>
+				<path
+					fill="#e1a48d"
+					d="M48.5 26.34c3.62.06 6.75 2.66 6.75 6.28a7 7 0 0 1-6.75 6.83 6.56 6.56 0 0 1 0-13.11Z"
+					opacity={0.45}
+				/>
+				<path
+					fill="#e1a690"
+					d="M48.51 26.35c3.62 0 6.74 2.67 6.74 6.29a7 7 0 0 1-6.74 6.82 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.44}
+				/>
+				<path
+					fill="#e2a892"
+					d="M48.51 26.36c3.62.05 6.74 2.67 6.74 6.29a7 7 0 0 1-6.74 6.82 6.56 6.56 0 0 1 0-13.11Z"
+					opacity={0.43}
+				/>
+				<path
+					fill="#e3aa94"
+					d="M48.52 26.37c3.62.05 6.74 2.67 6.74 6.29a6.94 6.94 0 0 1-6.74 6.82 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.42}
+				/>
+				<path
+					fill="#e3ac97"
+					d="M48.52 26.37c3.62.06 6.74 2.69 6.74 6.31a6.93 6.93 0 0 1-6.74 6.81 6.56 6.56 0 0 1 0-13.12Z"
+					opacity={0.41}
+				/>
+				<path
+					fill="#e4ad99"
+					d="M48.53 26.38a6.54 6.54 0 0 1 6.73 6.31 6.93 6.93 0 0 1-6.73 6.81 6.56 6.56 0 0 1 0-13.12Z"
+					opacity={0.4}
+				/>
+				<path
+					fill="#e4af9b"
+					d="M48.54 26.39a6.54 6.54 0 0 1 6.72 6.32 6.9 6.9 0 0 1-6.72 6.79 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.39}
+				/>
+				<path
+					fill="#e5b19e"
+					d="M48.54 26.4a6.54 6.54 0 0 1 6.72 6.32 6.9 6.9 0 0 1-6.72 6.79 6.56 6.56 0 0 1 0-13.11Z"
+					opacity={0.38}
+				/>
+				<path
+					fill="#e6b3a0"
+					d="M48.55 26.41a6.52 6.52 0 0 1 6.71 6.32 6.89 6.89 0 0 1-6.71 6.79 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.37}
+				/>
+				<path
+					fill="#e6b5a2"
+					d="M48.55 26.42a6.54 6.54 0 0 1 6.72 6.33 6.89 6.89 0 0 1-6.72 6.78 6.56 6.56 0 0 1 0-13.11Z"
+					opacity={0.36}
+				/>
+				<path
+					fill="#e7b7a5"
+					d="M48.56 26.43a6.53 6.53 0 0 1 6.71 6.33 6.88 6.88 0 0 1-6.71 6.78 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.35}
+				/>
+				<path
+					fill="#e8b9a7"
+					d="M48.56 26.43a6.55 6.55 0 0 1 6.71 6.35 6.88 6.88 0 0 1-6.71 6.77 6.56 6.56 0 0 1 0-13.12Z"
+					opacity={0.35}
+				/>
+				<path
+					fill="#e8baa9"
+					d="M48.57 26.44a6.54 6.54 0 0 1 6.7 6.35 6.86 6.86 0 0 1-6.7 6.76 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.34}
+				/>
+				<path
+					fill="#e9bcac"
+					d="M48.58 26.45a6.53 6.53 0 0 1 6.69 6.35 6.84 6.84 0 0 1-6.69 6.76 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.33}
+				/>
+				<path
+					fill="#e9beae"
+					d="M48.58 26.46a6.53 6.53 0 0 1 6.69 6.36 6.84 6.84 0 0 1-6.69 6.75 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.32}
+				/>
+				<path
+					fill="#eac0b0"
+					d="M48.59 26.47a6.53 6.53 0 0 1 6.69 6.36 6.84 6.84 0 0 1-6.69 6.75 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.31}
+				/>
+				<path
+					fill="#ebc2b3"
+					d="M48.59 26.48a6.54 6.54 0 0 1 6.69 6.37 6.83 6.83 0 0 1-6.69 6.74 6.56 6.56 0 0 1 0-13.11Z"
+					opacity={0.3}
+				/>
+				<path
+					fill="#ebc4b5"
+					d="M48.6 26.48a6.54 6.54 0 0 1 6.68 6.38 6.82 6.82 0 0 1-6.68 6.74 6.56 6.56 0 0 1 0-13.12Z"
+					opacity={0.29}
+				/>
+				<path
+					fill="#ecc6b7"
+					d="M48.6 26.49a6.54 6.54 0 0 1 6.68 6.38 6.81 6.81 0 0 1-6.68 6.73 6.56 6.56 0 0 1 0-13.11Z"
+					opacity={0.28}
+				/>
+				<path
+					fill="#ecc7b9"
+					d="M48.61 26.5a6.54 6.54 0 0 1 6.67 6.39 6.8 6.8 0 0 1-6.67 6.72 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.27}
+				/>
+				<path
+					fill="#edc9bc"
+					d="M48.61 26.51a6.54 6.54 0 0 1 6.67 6.39 6.79 6.79 0 0 1-6.67 6.72 6.56 6.56 0 0 1 0-13.11Z"
+					opacity={0.26}
+				/>
+				<path
+					fill="#eecbbe"
+					d="M48.62 26.52a6.54 6.54 0 0 1 6.67 6.4 6.79 6.79 0 0 1-6.67 6.71 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.25}
+				/>
+				<path
+					fill="#eecdc0"
+					d="M48.63 26.53a6.53 6.53 0 0 1 6.66 6.4 6.78 6.78 0 0 1-6.66 6.71 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.25}
+				/>
+				<path
+					fill="#efcfc3"
+					d="M48.63 26.53a6.54 6.54 0 0 1 6.66 6.41 6.77 6.77 0 0 1-6.66 6.71 6.56 6.56 0 0 1 0-13.12Z"
+					opacity={0.24}
+				/>
+				<path
+					fill="#f0d1c5"
+					d="M48.64 26.54A6.54 6.54 0 0 1 55.29 33a6.75 6.75 0 0 1-6.65 6.69 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.23}
+				/>
+				<path
+					fill="#f0d2c7"
+					d="M48.64 26.55A6.54 6.54 0 0 1 55.29 33a6.75 6.75 0 0 1-6.65 6.69 6.56 6.56 0 0 1 0-13.11Z"
+					opacity={0.22}
+				/>
+				<path
+					fill="#f1d4ca"
+					d="M48.65 26.56A6.54 6.54 0 0 1 55.29 33a6.74 6.74 0 0 1-6.64 6.68 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.21}
+				/>
+				<path
+					fill="#f1d6cc"
+					d="M48.65 26.57A6.54 6.54 0 0 1 55.3 33a6.74 6.74 0 0 1-6.65 6.68 6.56 6.56 0 0 1 0-13.11Z"
+					opacity={0.2}
+				/>
+				<path
+					fill="#f2d8ce"
+					d="M48.66 26.58A6.54 6.54 0 0 1 55.3 33a6.73 6.73 0 0 1-6.64 6.68 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.19}
+				/>
+				<path
+					fill="#f3dad1"
+					d="M48.66 26.58A6.55 6.55 0 0 1 55.3 33a6.72 6.72 0 0 1-6.64 6.67 6.56 6.56 0 0 1 0-13.12Z"
+					opacity={0.18}
+				/>
+				<path
+					fill="#f3dcd3"
+					d="M48.67 26.59A6.54 6.54 0 0 1 55.3 33a6.71 6.71 0 0 1-6.63 6.66 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.17}
+				/>
+				<path
+					fill="#f4ded5"
+					d="M48.68 26.6a6.54 6.54 0 0 1 6.62 6.46 6.69 6.69 0 0 1-6.62 6.65 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.16}
+				/>
+				<path
+					fill="#f4dfd8"
+					d="M48.68 26.61a6.54 6.54 0 0 1 6.62 6.46 6.69 6.69 0 0 1-6.62 6.65 6.56 6.56 0 0 1 0-13.11Z"
+					opacity={0.15}
+				/>
+				<path
+					fill="#f5e1da"
+					d="M48.69 26.62a6.54 6.54 0 0 1 6.62 6.46 6.69 6.69 0 0 1-6.62 6.65 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.15}
+				/>
+				<path
+					fill="#f6e3dc"
+					d="M48.69 26.63a6.55 6.55 0 0 1 6.62 6.47 6.68 6.68 0 0 1-6.62 6.64 6.56 6.56 0 0 1 0-13.11Z"
+					opacity={0.14}
+				/>
+				<path
+					fill="#f6e5df"
+					d="M48.7 26.63a6.55 6.55 0 0 1 6.61 6.48 6.68 6.68 0 0 1-6.61 6.64 6.56 6.56 0 0 1 0-13.12Z"
+					opacity={0.13}
+				/>
+				<path
+					fill="#f7e7e1"
+					d="M48.7 26.64a6.56 6.56 0 0 1 6.61 6.49 6.67 6.67 0 0 1-6.61 6.63 6.56 6.56 0 0 1 0-13.12Z"
+					opacity={0.12}
+				/>
+				<path
+					fill="#f8e9e3"
+					d="M48.71 26.65a6.55 6.55 0 0 1 6.6 6.49 6.65 6.65 0 0 1-6.6 6.62 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.11}
+				/>
+				<path
+					fill="#f8ebe6"
+					d="M48.72 26.66a6.54 6.54 0 0 1 6.59 6.49 6.63 6.63 0 0 1-6.59 6.62 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.1}
+				/>
+				<path
+					fill="#f9ece8"
+					d="M48.72 26.67a6.55 6.55 0 0 1 6.6 6.5 6.64 6.64 0 0 1-6.6 6.61 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.09}
+				/>
+				<path
+					fill="#f9eeea"
+					d="M48.73 26.68a6.54 6.54 0 0 1 6.59 6.5 6.63 6.63 0 0 1-6.59 6.61 6.56 6.56 0 1 1 0-13.11Z"
+					opacity={0.08}
+				/>
+				<path
+					fill="#faf0ec"
+					d="M48.73 26.69a6.55 6.55 0 0 1 6.59 6.51 6.62 6.62 0 0 1-6.59 6.6 6.56 6.56 0 0 1 0-13.11Z"
+					opacity={0.07}
+				/>
+				<path
+					fill="#fbf2ef"
+					d="M48.74 26.69a6.56 6.56 0 1 1-6.56 6.56 6.55 6.55 0 0 1 6.56-6.56Z"
+					opacity={0.06}
+				/>
+				<path
+					fill="#fbf4f1"
+					d="M48.74 26.7a6.56 6.56 0 1 1-6.55 6.56 6.55 6.55 0 0 1 6.55-6.56Z"
+					opacity={0.05}
+				/>
+				<path
+					fill="#fcf6f3"
+					d="M48.75 26.71a6.56 6.56 0 1 1-6.56 6.56 6.55 6.55 0 0 1 6.56-6.56Z"
+					opacity={0.05}
+				/>
+				<path
+					fill="#fdf8f6"
+					d="M48.75 26.72a6.56 6.56 0 1 1-6.55 6.55 6.55 6.55 0 0 1 6.55-6.55Z"
+					opacity={0.04}
+				/>
+				<path
+					fill="#fdf9f8"
+					d="M48.76 26.73a6.56 6.56 0 1 1-6.56 6.55 6.56 6.56 0 0 1 6.56-6.55Z"
+					opacity={0.03}
+				/>
+				<path
+					fill="#fefbfa"
+					d="M48.77 26.74a6.56 6.56 0 1 1-6.56 6.55 6.55 6.55 0 0 1 6.56-6.55Z"
+					opacity={0.02}
+				/>
+				<path
+					fill="#fefdfd"
+					d="M48.77 26.74a6.56 6.56 0 1 1-6.55 6.56 6.56 6.56 0 0 1 6.55-6.56Z"
+					opacity={0.01}
+				/>
+			</g>
+			<g clipPath="url(#d)">
+				<g clipPath="url(#e)">
+					<path
+						fill="#e04e15"
+						d="m-25.94 11.65 93.592-54.036 54.035 93.592-93.592 54.035z"
+					/>
+					<path
+						fill="#e04f15"
+						d="m-24.806 13.593 93.59-54.035L122.82 53.15l-93.591 54.035z"
+					/>
+					<path
+						fill="#e15015"
+						d="m-23.707 15.477 94.021-54.035 54.035 94.021-94.02 54.035z"
+					/>
+					<path
+						fill="#e15115"
+						d="m-22.562 17.487 93.59-54.035 54.036 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#e25215"
+						d="m-21.444 19.377 94.02-54.035 54.036 94.02-94.021 54.035z"
+					/>
+					<path
+						fill="#e25316"
+						d="m-20.314 21.39 93.591-54.036 54.035 93.592-93.59 54.035z"
+					/>
+					<path
+						fill="#e35416"
+						d="m-19.189 23.338 93.591-54.035 54.035 93.591-93.59 54.035z"
+					/>
+					<path
+						fill="#e35516"
+						d="m-18.055 25.282 93.591-54.035 54.035 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#e35616"
+						d="m-16.93 27.213 94.022-54.035 54.035 94.02-94.021 54.035z"
+					/>
+					<path
+						d="M-15.812 29.176 77.78-24.86l54.035 93.591-93.59 54.035z"
+						className="cls-240"
+					/>
+					<path
+						d="M-14.687 31.124 78.904-22.91l54.035 93.592-93.59 54.035z"
+						className="cls-240"
+					/>
+					<path
+						fill="#e55816"
+						d="m-13.539 33.073 94.021-54.035 54.035 94.021-94.02 54.035z"
+					/>
+					<path
+						fill="#e55916"
+						d="m-12.438 35.027 93.591-54.035 54.035 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#e55a16"
+						d="m-11.305 36.97 93.591-54.034 54.035 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#e65b16"
+						d="M-10.152 38.948 83.87-15.087l54.035 94.02-94.02 54.036z"
+					/>
+					<path
+						fill="#e65c17"
+						d="m-9.014 40.9 94.021-54.036 54.035 94.021-94.02 54.035z"
+					/>
+					<path
+						fill="#e75d17"
+						d="m-7.89 42.857 94.022-54.035 54.035 94.021-94.021 54.035z"
+					/>
+					<path
+						fill="#e75e17"
+						d="m-6.812 44.767 93.59-54.035 54.036 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#e85f17"
+						d="M-5.688 46.716 87.903-7.32l54.035 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#e86017"
+						d="M-4.554 48.67 89.037-5.365l54.035 93.59-93.591 54.036z"
+					/>
+					<path
+						fill="#e86117"
+						d="M-4.126 49.414 89.465-4.621 143.5 88.97l-93.591 54.035z"
+					/>
+					<path
+						fill="#e96217"
+						d="M-3.686 50.172 89.906-3.863 143.94 89.73l-93.592 54.035z"
+					/>
+					<path
+						fill="#e96317"
+						d="M-3.258 50.927 90.333-3.108l54.035 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#e96418"
+						d="M-2.822 51.676 90.769-2.359l54.035 93.592-93.591 54.035z"
+					/>
+					<path
+						fill="#e96518"
+						d="M-2.32 52.512 91.7-1.523l54.035 94.02-94.02 54.036z"
+					/>
+					<path
+						fill="#ea6618"
+						d="M-1.953 53.194 91.64-.841l54.035 93.591-93.592 54.035z"
+					/>
+					<path
+						fill="#ea6718"
+						d="M-1.516 53.934 92.075-.101 146.11 93.49l-93.591 54.035z"
+					/>
+					<path
+						fill="#ea6818"
+						d="M-1.089 54.688 92.502.653l54.035 93.591-93.59 54.035z"
+					/>
+					<path
+						fill="#ea6918"
+						d="M-.648 55.446 92.944 1.411l54.035 93.592-93.592 54.035z"
+					/>
+					<path
+						fill="#eb6a18"
+						d="M-.22 56.2 93.37 2.165l54.036 93.591-93.592 54.035z"
+					/>
+					<path
+						fill="#eb6b19"
+						d="M.216 56.95 93.807 2.915l54.035 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#eb6c19"
+						d="M.658 57.713 94.249 3.678l54.035 93.592-93.591 54.035z"
+					/>
+					<path
+						fill="#eb6d19"
+						d="m1.177 58.549 94.02-54.035 54.036 94.02-94.021 54.036z"
+					/>
+					<path
+						fill="#ec6e19"
+						d="M1.521 59.207 95.113 5.172l54.035 93.592-93.592 54.035z"
+					/>
+					<path
+						fill="#ec6f19"
+						d="m2.049 60.068 94.02-54.035 54.036 94.02-94.021 54.036z"
+					/>
+					<path
+						fill="#ec7019"
+						d="M2.39 60.72 95.981 6.685l54.035 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#ec7119"
+						d="M2.826 61.47 96.417 7.435l54.035 93.59-93.591 54.036z"
+					/>
+					<path
+						fill="#ed7219"
+						d="M3.258 62.233 96.849 8.198l54.035 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#ed731a"
+						d="M3.704 62.972 97.296 8.937l54.035 93.592-93.592 54.035z"
+					/>
+					<path
+						fill="#ed741a"
+						d="M4.132 63.727 97.723 9.692l54.035 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#ed751a"
+						d="M4.573 64.485 98.164 10.45l54.036 93.59-93.591 54.035z"
+					/>
+					<path
+						fill="#ee761a"
+						d="m5 65.24 93.591-54.035 54.035 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#ee771a"
+						d="m5.436 65.989 93.591-54.035 54.035 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#ee781a"
+						d="m5.97 66.881 94.021-54.035 54.035 94.021-94.02 54.035z"
+					/>
+					<path
+						fill="#ef791a"
+						d="m6.315 67.492 93.591-54.035 54.035 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#ef7a1b"
+						d="m6.742 68.246 93.592-54.035 54.035 93.592-93.592 54.035z"
+					/>
+					<path
+						fill="#ef7b1b"
+						d="m7.183 69.004 93.592-54.035 54.035 93.592-93.592 54.035z"
+					/>
+					<path
+						fill="#ef7c1b"
+						d="m7.72 69.9 94.02-54.035 54.035 94.02-94.02 54.036z"
+					/>
+					<path
+						fill="#f07d1b"
+						d="m8.157 70.653 94.021-54.036 54.035 94.021-94.02 54.035z"
+					/>
+					<path
+						fill="#f07e1b"
+						d="m8.591 71.419 94.021-54.035 54.035 94.02-94.02 54.036z"
+					/>
+					<path
+						fill="#f07f1b"
+						d="m8.925 72.021 93.592-54.035 54.035 93.592-93.592 54.035z"
+					/>
+					<path
+						fill="#f0801b"
+						d="m9.468 72.919 94.021-54.035 54.035 94.02-94.02 54.036z"
+					/>
+					<path
+						fill="#f1801c"
+						d="m9.91 73.68 94.022-54.035 54.035 94.02-94.021 54.036z"
+					/>
+					<path
+						fill="#f1811c"
+						d="m10.22 74.278 93.592-54.035 54.035 93.592-93.591 54.035z"
+					/>
+					<path
+						fill="#f1821c"
+						d="m10.657 75.028 93.591-54.035 54.035 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#f1831c"
+						d="m11.09 75.791 93.59-54.035 54.036 93.592-93.592 54.035z"
+					/>
+					<path
+						fill="#f2841c"
+						d="m11.536 76.54 93.59-54.034 54.036 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#f2851c"
+						d="m11.962 77.285 93.592-54.035 54.035 93.591-93.592 54.035z"
+					/>
+					<path
+						fill="#f2861c"
+						d="m12.404 78.044 93.591-54.035L160.03 117.6l-93.591 54.035z"
+					/>
+					<path
+						fill="#f2871d"
+						d="m12.831 78.798 93.591-54.035 54.035 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#f3881d"
+						d="m13.267 79.548 93.592-54.035 54.035 93.59-93.592 54.036z"
+					/>
+					<path
+						fill="#f3891d"
+						d="m13.694 80.301 93.592-54.035 54.035 93.592-93.592 54.035z"
+					/>
+					<path
+						fill="#f38a1d"
+						d="m14.145 81.06 93.592-54.035 54.035 93.592-93.592 54.035z"
+					/>
+					<path
+						fill="#f48b1d"
+						d="m14.573 81.804 93.591-54.035 54.035 93.592-93.591 54.035z"
+					/>
+					<path
+						fill="#f48c1d"
+						d="m15.147 82.746 94.021-54.035 54.035 94.02-94.02 54.036z"
+					/>
+					<path
+						fill="#f48d1d"
+						d="m15.581 83.512 94.021-54.035 54.035 94.021-94.02 54.035z"
+					/>
+					<path
+						fill="#f48e1e"
+						d="m16.02 84.265 94.02-54.035 54.035 94.02-94.02 54.036z"
+					/>
+					<path
+						fill="#f58f1e"
+						d="m16.305 84.821 93.591-54.035 54.035 93.592-93.591 54.035z"
+					/>
+					<path
+						fill="#f5901e"
+						d="m16.747 85.585 93.591-54.035 54.035 93.59-93.59 54.036z"
+					/>
+					<path
+						fill="#f5911e"
+						d="m17.184 86.324 93.591-54.035 54.035 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#f5921e"
+						d="m17.76 87.288 94.02-54.035 54.035 94.021-94.02 54.035z"
+					/>
+					<path
+						fill="#f6931e"
+						d="m18.202 88.05 94.021-54.035 54.035 94.02-94.02 54.035z"
+					/>
+					<path
+						fill="#f6941e"
+						d="m18.48 88.591 93.59-54.035 54.036 93.592-93.592 54.035z"
+					/>
+					<path
+						fill="#f6951e"
+						d="m18.915 89.34 93.591-54.034 54.035 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#f6961f"
+						d="m19.357 90.104 93.592-54.035 54.035 93.591-93.592 54.035z"
+					/>
+					<path
+						fill="#f7971f"
+						d="m19.794 90.844 93.591-54.035L167.42 130.4l-93.591 54.035z"
+					/>
+					<path
+						fill="#f7981f"
+						d="m20.221 91.598 93.591-54.035 54.035 93.591-93.59 54.035z"
+					/>
+					<path
+						fill="#f7991f"
+						d="m20.662 92.356 93.592-54.035 54.035 93.592-93.592 54.035z"
+					/>
+					<path
+						fill="#f79a1f"
+						d="m21.252 93.344 94.02-54.034 54.036 94.02-94.021 54.035z"
+					/>
+					<path
+						fill="#f89b1f"
+						d="m21.526 93.86 93.591-54.035 54.035 93.591-93.591 54.035z"
+					/>
+					<path
+						fill="#f89c1f"
+						d="m21.968 94.623 93.591-54.035 54.035 93.592-93.591 54.035z"
+					/>
+					<path
+						fill="#f89d20"
+						d="m22.572 95.606 94.02-54.035 54.035 94.02-94.02 54.036z"
+					/>
+					<path
+						fill="#f89e20"
+						d="m22.831 96.117 93.592-54.035 54.035 93.592-93.592 54.035z"
+					/>
+					<path
+						fill="#f99f20"
+						d="m23.273 96.876 93.59-54.035 54.036 93.59-93.591 54.036z"
+					/>
+					<path
+						fill="#f9a020"
+						d="m23.7 97.63 93.591-54.035 54.035 93.591-93.591 54.035z"
+					/>
 				</g>
 			</g>
 		</g>
 	</svg>
+	/* eslint-enable */
 );
