+++ packages/js/onboarding/src/images/cards/giropay.js
