+++ packages/js/onboarding/src/images/cards/googlepay.js
@@ -3,44 +3,25 @@
  */
 import { createElement } from '@wordpress/element';
 
-export default () => (
-	<svg
-		width="64"
-		height="40"
-		viewBox="0 0 64 40"
-		fill="none"
-		xmlns="http://www.w3.org/2000/svg"
-	>
-		<rect width="64" height="40" fill="white" />
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M30.9 25.07V19.79H33.76C34.8522 19.8223 35.914 19.4277 36.72 18.69C37.118 18.3482 37.4356 17.9226 37.65 17.4438C37.8643 16.9649 37.9702 16.4446 37.96 15.92C37.9659 15.399 37.858 14.883 37.6438 14.408C37.4296 13.933 37.1143 13.5105 36.72 13.17C35.9175 12.4232 34.856 12.0179 33.76 12.04H29.18V25.07H30.9ZM30.9 18.19V13.64H33.8C34.45 13.63 35.07 13.88 35.51 14.34C35.7261 14.5446 35.8981 14.7912 36.0157 15.0646C36.1334 15.3379 36.194 15.6324 36.194 15.93C36.194 16.2276 36.1334 16.5221 36.0157 16.7954C35.8981 17.0688 35.7261 17.3153 35.51 17.52C35.06 17.97 34.44 18.22 33.8 18.19H30.9ZM44.97 16.86C44.23 16.19 43.22 15.86 41.94 15.86C40.3 15.86 39.07 16.45 38.23 17.61L39.75 18.54C40.0021 18.1647 40.3454 17.8595 40.7477 17.6531C41.15 17.4467 41.5981 17.3458 42.05 17.36C42.65 17.35 43.25 17.56 43.71 17.96C44.16 18.33 44.43 18.88 44.42 19.45V19.83C43.6359 19.4379 42.7661 19.2488 41.89 19.28C40.69 19.28 39.73 19.56 39.01 20.1C38.6643 20.3509 38.3851 20.6825 38.1969 21.0659C38.0087 21.4494 37.917 21.873 37.93 22.3C37.91 23.14 38.29 23.95 38.95 24.49C39.63 25.07 40.48 25.36 41.49 25.36C42.69 25.36 43.63 24.86 44.35 23.83H44.42V25.07H46.08V19.57C46.08 18.42 45.71 17.52 44.97 16.86ZM40.28 23.43C40.1036 23.3082 39.9596 23.1454 39.8603 22.9554C39.761 22.7655 39.7094 22.5543 39.71 22.34C39.71 21.86 39.94 21.45 40.41 21.12C40.89 20.79 41.47 20.62 42.17 20.62C43.14 20.62 43.89 20.83 44.43 21.25C44.43 21.95 44.14 22.57 43.57 23.09C43.0458 23.5932 42.3466 23.8729 41.62 23.87C41.14 23.87 40.66 23.72 40.28 23.43ZM55.56 16.16L49.8 29H48.02L50.16 24.5L46.36 16.16H48.24L50.99 22.56H51L53.67 16.16H55.55H55.56Z"
-			fill="#5F6368"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M23.92 18.66C23.92 18.15 23.88 17.64 23.79 17.13H16.5V20.02H20.68C20.5904 20.4842 20.4067 20.9251 20.1401 21.3155C19.8735 21.7059 19.5297 22.0376 19.13 22.29V24.16H21.63C22.3839 23.4607 22.9792 22.6079 23.376 21.6592C23.7727 20.7105 23.9616 19.6878 23.93 18.66H23.92Z"
-			fill="#4285F4"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M16.5 25.97C18.59 25.97 20.34 25.31 21.62 24.17L19.13 22.29C18.44 22.75 17.55 23.01 16.5 23.01C15.5459 23.0062 14.6166 22.7057 13.8408 22.1503C13.065 21.5949 12.4811 20.812 12.17 19.91H9.60001V21.85C10.2645 23.1012 11.2588 24.1465 12.4751 24.8728C13.6914 25.5991 15.0834 25.9785 16.5 25.97Z"
-			fill="#34A853"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M12.17 19.92C11.8453 18.9909 11.8453 17.9791 12.17 17.05V15.12H9.6C9.05843 16.159 8.77563 17.3133 8.77563 18.485C8.77563 19.6567 9.05843 20.811 9.6 21.85L12.17 19.92Z"
-			fill="#FBBC04"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M16.5 13.96C17.6 13.95 18.67 14.35 19.47 15.09L21.67 12.95C20.2545 11.6709 18.4077 10.9744 16.5 11C13.58 11 10.91 12.6 9.60001 15.12L12.17 17.05C12.4828 16.1499 13.0674 15.3692 13.8431 14.8156C14.6187 14.2621 15.5471 13.9631 16.5 13.96Z"
-			fill="#EA4335"
-		/>
+export const GooglePay = () => (
+	/* eslint-disable */
+	<svg viewBox="0 0 64 40" fill="none" xmlns="http://www.w3.org/2000/svg">
+		<rect width="64" height="40" fill="white"/>
+		<path fillRule="evenodd" clipRule="evenodd"
+			  d="M30.9 25.07V19.79H33.76C34.8522 19.8223 35.914 19.4277 36.72 18.69C37.118 18.3482 37.4356 17.9226 37.65 17.4438C37.8643 16.9649 37.9702 16.4446 37.96 15.92C37.9659 15.399 37.858 14.883 37.6438 14.408C37.4296 13.933 37.1143 13.5105 36.72 13.17C35.9175 12.4232 34.856 12.0179 33.76 12.04H29.18V25.07H30.9ZM30.9 18.19V13.64H33.8C34.45 13.63 35.07 13.88 35.51 14.34C35.7261 14.5446 35.8981 14.7912 36.0157 15.0646C36.1334 15.3379 36.194 15.6324 36.194 15.93C36.194 16.2276 36.1334 16.5221 36.0157 16.7954C35.8981 17.0688 35.7261 17.3153 35.51 17.52C35.06 17.97 34.44 18.22 33.8 18.19H30.9ZM44.97 16.86C44.23 16.19 43.22 15.86 41.94 15.86C40.3 15.86 39.07 16.45 38.23 17.61L39.75 18.54C40.0021 18.1647 40.3454 17.8595 40.7477 17.6531C41.15 17.4467 41.5981 17.3458 42.05 17.36C42.65 17.35 43.25 17.56 43.71 17.96C44.16 18.33 44.43 18.88 44.42 19.45V19.83C43.6359 19.4379 42.7661 19.2488 41.89 19.28C40.69 19.28 39.73 19.56 39.01 20.1C38.6643 20.3509 38.3851 20.6825 38.1969 21.0659C38.0087 21.4494 37.917 21.873 37.93 22.3C37.91 23.14 38.29 23.95 38.95 24.49C39.63 25.07 40.48 25.36 41.49 25.36C42.69 25.36 43.63 24.86 44.35 23.83H44.42V25.07H46.08V19.57C46.08 18.42 45.71 17.52 44.97 16.86ZM40.28 23.43C40.1036 23.3082 39.9596 23.1454 39.8603 22.9554C39.761 22.7655 39.7094 22.5543 39.71 22.34C39.71 21.86 39.94 21.45 40.41 21.12C40.89 20.79 41.47 20.62 42.17 20.62C43.14 20.62 43.89 20.83 44.43 21.25C44.43 21.95 44.14 22.57 43.57 23.09C43.0458 23.5932 42.3466 23.8729 41.62 23.87C41.14 23.87 40.66 23.72 40.28 23.43ZM55.56 16.16L49.8 29H48.02L50.16 24.5L46.36 16.16H48.24L50.99 22.56H51L53.67 16.16H55.55H55.56Z"
+			  fill="#5F6368"/>
+		<path fillRule="evenodd" clipRule="evenodd"
+			  d="M23.92 18.66C23.92 18.15 23.88 17.64 23.79 17.13H16.5V20.02H20.68C20.5904 20.4842 20.4067 20.9251 20.1401 21.3155C19.8735 21.7059 19.5297 22.0376 19.13 22.29V24.16H21.63C22.3839 23.4607 22.9792 22.6079 23.376 21.6592C23.7727 20.7105 23.9616 19.6878 23.93 18.66H23.92Z"
+			  fill="#4285F4"/>
+		<path fillRule="evenodd" clipRule="evenodd"
+			  d="M16.5 25.97C18.59 25.97 20.34 25.31 21.62 24.17L19.13 22.29C18.44 22.75 17.55 23.01 16.5 23.01C15.5459 23.0062 14.6166 22.7057 13.8408 22.1503C13.065 21.5949 12.4811 20.812 12.17 19.91H9.60001V21.85C10.2645 23.1012 11.2588 24.1465 12.4751 24.8728C13.6914 25.5991 15.0834 25.9785 16.5 25.97Z"
+			  fill="#34A853"/>
+		<path fillRule="evenodd" clipRule="evenodd"
+			  d="M12.17 19.92C11.8453 18.9909 11.8453 17.9791 12.17 17.05V15.12H9.6C9.05843 16.159 8.77563 17.3133 8.77563 18.485C8.77563 19.6567 9.05843 20.811 9.6 21.85L12.17 19.92Z"
+			  fill="#FBBC04"/>
+		<path fillRule="evenodd" clipRule="evenodd"
+			  d="M16.5 13.96C17.6 13.95 18.67 14.35 19.47 15.09L21.67 12.95C20.2545 11.6709 18.4077 10.9744 16.5 11C13.58 11 10.91 12.6 9.60001 15.12L12.17 17.05C12.4828 16.1499 13.0674 15.3692 13.8431 14.8156C14.6187 14.2621 15.5471 13.9631 16.5 13.96Z"
+			  fill="#EA4335"/>
 	</svg>
+	/* eslint-enable */
 );
