+++ packages/js/onboarding/src/images/cards/index.js
@@ -1,17 +1,12 @@
-export * from './affirm';
-export * from './afterpay';
 export * from './amex';
 export * from './applepay';
 export * from './cb';
-export * from './clearpay';
+export * from './diners';
 export * from './discover';
 export * from './giropay';
-export * from './gpay';
-export * from './ideal';
+export * from './googlepay';
 export * from './jcb';
-export * from './klarna';
 export * from './maestro';
 export * from './mastercard';
 export * from './unionpay';
 export * from './visa';
-export * from './woopay';
