+++ packages/js/onboarding/src/images/cards/mastercard.js
@@ -3,38 +3,24 @@
  */
 import { createElement } from '@wordpress/element';
 
-export default () => (
-	<svg
-		width="64"
-		height="40"
-		viewBox="0 0 64 40"
-		fill="none"
-		xmlns="http://www.w3.org/2000/svg"
-	>
-		<rect width="64" height="40" fill="white" />
+export const MasterCard = () => (
+	/* eslint-disable */
+	<svg viewBox="0 0 64 40" fill="none" xmlns="http://www.w3.org/2000/svg">
+		<rect width="64" height="40" fill="white"/>
 		<g clipPath="url(#clip0_1_132)">
-			<path
-				d="M37.376 10.284H26.56V29.716H37.376V10.284Z"
-				fill="#FF5A00"
-			/>
+			<path d="M37.376 10.284H26.56V29.716H37.376V10.284Z" fill="#FF5A00"/>
 			<path
 				d="M27.28 20C27.28 16.052 29.136 12.548 31.984 10.284C29.888 8.636 27.244 7.64 24.36 7.64C17.528 7.64 12 13.168 12 20C12 26.832 17.528 32.36 24.36 32.36C27.244 32.36 29.888 31.364 31.984 29.716C29.132 27.484 27.28 23.948 27.28 20Z"
-				fill="#EB001B"
-			/>
+				fill="#EB001B"/>
 			<path
 				d="M51.968 20C51.968 26.832 46.44 32.36 39.608 32.36C36.724 32.36 34.08 31.364 31.984 29.716C34.868 27.448 36.688 23.948 36.688 20C36.688 16.052 34.832 12.548 31.984 10.284C34.076 8.636 36.72 7.64 39.604 7.64C46.44 7.64 51.968 13.204 51.968 20Z"
-				fill="#F79E1B"
-			/>
+				fill="#F79E1B"/>
 		</g>
 		<defs>
 			<clipPath id="clip0_1_132">
-				<rect
-					width="40"
-					height="24.72"
-					fill="white"
-					transform="translate(12 7.64)"
-				/>
+				<rect width="40" height="24.72" fill="white" transform="translate(12 7.64)"/>
 			</clipPath>
 		</defs>
 	</svg>
+	/* eslint-enable */
 );
