+++ packages/js/onboarding/src/images/cards/unionpay.js
@@ -3,112 +3,29 @@
  */
 import { createElement } from '@wordpress/element';
 
-export default () => (
-	<svg
-		width="52"
-		height="35"
-		viewBox="0 0 52 35"
-		fill="none"
-		xmlns="http://www.w3.org/2000/svg"
-	>
-		<rect
-			x="0.878906"
-			y="0.5"
-			width="50"
-			height="34"
-			rx="3.5"
-			fill="white"
-			stroke="#F3F3F3"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M44.0545 5.25735L34.3353 5.25488C34.3341 5.25488 34.3328 5.25488 34.3328 5.25488C34.3253 5.25488 34.3179 5.2562 34.3106 5.2562C32.9754 5.29641 31.3124 6.34915 31.0096 7.64726L26.4132 27.6401C26.1104 28.9503 26.9343 30.0165 28.2599 30.0361H38.4703C39.7756 29.9726 41.044 28.932 41.3417 27.6486L45.9382 7.65564C46.2459 6.33208 45.402 5.25735 44.0545 5.25735Z"
-			fill="#01798A"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M26.4134 27.6401L31.0097 7.64729C31.3126 6.34917 32.9755 5.29643 34.3107 5.25622L30.4464 5.25376L23.484 5.25244C22.1451 5.27936 20.4605 6.33949 20.1577 7.64729L15.5601 27.6401C15.2561 28.9503 16.0813 30.0165 17.4059 30.0361H28.26C26.9345 30.0165 26.1105 28.9503 26.4134 27.6401"
-			fill="#024381"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M15.5602 27.64L20.1578 7.64714C20.4606 6.33934 22.1452 5.27922 23.4841 5.2523L14.5649 5.25C13.2185 5.25 11.4923 6.32227 11.1846 7.64714L6.58694 27.64C6.55896 27.762 6.54344 27.8815 6.53418 27.9986V28.3695C6.62418 29.3246 7.36619 30.0201 8.43278 30.036H17.406C16.0814 30.0163 15.2562 28.9502 15.5602 27.64Z"
-			fill="#DD0228"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M23.6716 19.8205H23.8404C23.9955 19.8205 24.0999 19.7693 24.1488 19.668L24.5874 19.0227H25.762L25.5171 19.4472H26.9254L26.7467 20.0975H25.0709C24.8779 20.3829 24.6403 20.5171 24.3547 20.5012H23.4818L23.6716 19.8205H23.6716ZM23.4788 20.7527H26.5643L26.3676 21.4591H25.1268L24.9374 22.1409H26.1449L25.9482 22.8473H24.7407L24.4602 23.8548C24.3908 24.0232 24.4821 24.099 24.7327 24.0818H25.7168L25.5345 24.7382H23.6451C23.287 24.7382 23.1641 24.5368 23.2765 24.1331L23.6351 22.8473H22.8633L23.0593 22.1409H23.8313L24.0205 21.4591H23.2827L23.4788 20.7527H23.4788ZM28.4035 19.018L28.355 19.4315C28.355 19.4315 28.937 19.002 29.4656 19.002H31.4189L30.6719 21.6601C30.61 21.964 30.3443 22.1151 29.8752 22.1151H27.6612L27.1426 23.9817C27.1128 24.0817 27.155 24.133 27.2667 24.133H27.7023L27.5422 24.7124H26.4347C26.0096 24.7124 25.8328 24.5867 25.903 24.3343L27.3684 19.018H28.4035H28.4035ZM30.0576 19.7693H28.3141L28.1056 20.4866C28.1056 20.4866 28.3959 20.2805 28.8811 20.2731C29.365 20.2657 29.9173 20.2731 29.9173 20.2731L30.0576 19.7693ZM29.4261 21.4333C29.555 21.4504 29.6271 21.4003 29.6358 21.282L29.7425 20.9039H27.9964L27.85 21.4333H29.4261ZM28.2483 22.2921H29.2547L29.236 22.7203H29.504C29.6394 22.7203 29.7065 22.6776 29.7065 22.5935L29.7858 22.3166H30.6223L30.5106 22.7203C30.4161 23.057 30.1656 23.2327 29.7586 23.2499H29.2225L29.22 23.9817C29.2101 24.0989 29.318 24.1587 29.54 24.1587H30.0439L29.8813 24.7381H28.6727C28.3339 24.754 28.1678 24.5953 28.1713 24.2587L28.2483 22.2921V22.2921Z"
-			fill="white"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M16.0529 15.4764C15.9164 16.1339 15.6 16.639 15.1091 16.9976C14.6227 17.3502 13.9954 17.527 13.2273 17.527C12.5044 17.527 11.9745 17.3465 11.6364 16.9841C11.4018 16.7267 11.2852 16.3998 11.2852 16.0045C11.2852 15.8411 11.3051 15.6654 11.3448 15.4764L12.1631 11.5972H13.3991L12.5919 15.4325C12.5671 15.5386 12.5571 15.6374 12.5584 15.7265C12.5571 15.9229 12.6068 16.0839 12.7073 16.2095C12.8537 16.3962 13.0914 16.4889 13.4221 16.4889C13.8024 16.4889 14.1158 16.3974 14.359 16.2132C14.6022 16.0302 14.761 15.7704 14.8324 15.4325L15.6422 11.5972H16.8719L16.0529 15.4764Z"
-			fill="white"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M21.2436 13.9502H22.2116L21.4534 17.4123H20.4873L21.2436 13.9502ZM21.5482 12.689H22.5248L22.3424 13.5293H21.3659L21.5482 12.689Z"
-			fill="white"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M23.0688 17.1487C22.8156 16.9109 22.6878 16.59 22.6865 16.1826C22.6865 16.113 22.6908 16.0338 22.7002 15.9471C22.7095 15.8592 22.7214 15.7739 22.738 15.6946C22.8528 15.1323 23.0973 14.6858 23.4739 14.3564C23.8499 14.0258 24.3036 13.8599 24.8347 13.8599C25.2696 13.8599 25.6145 13.9794 25.8672 14.2185C26.1196 14.4589 26.2462 14.7833 26.2462 15.1957C26.2462 15.2664 26.2407 15.3481 26.2313 15.436C26.2201 15.525 26.2066 15.6104 26.1909 15.6946C26.0787 16.2484 25.8349 16.69 25.4583 17.0134C25.0816 17.3391 24.6293 17.5012 24.1019 17.5012C23.6651 17.5012 23.3213 17.3841 23.0688 17.1487M24.9136 16.4631C25.0843 16.2814 25.2065 16.0056 25.2809 15.6385C25.2921 15.5812 25.302 15.5214 25.3082 15.4616C25.3143 15.403 25.3168 15.3482 25.3168 15.2981C25.3168 15.0846 25.2616 14.9188 25.1506 14.8016C25.0402 14.6833 24.8832 14.6248 24.6804 14.6248C24.4122 14.6248 24.1939 14.7174 24.0227 14.9029C23.8501 15.0884 23.7279 15.3689 23.6509 15.7422C23.6404 15.7995 23.6317 15.8569 23.6237 15.913C23.6175 15.9703 23.6157 16.024 23.6168 16.0728C23.6168 16.285 23.6721 16.4485 23.7831 16.5644C23.8935 16.6803 24.0498 16.7376 24.2553 16.7376C24.5246 16.7376 24.743 16.6461 24.9136 16.4631Z"
-			fill="white"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M32.5262 19.8496L32.7596 19.0421H33.9397L33.8888 19.3385C33.8888 19.3385 34.4918 19.0421 34.9261 19.0421C35.3606 19.0421 36.3854 19.0421 36.3854 19.0421L36.1535 19.8496H35.9239L34.8231 23.6582H35.0527L34.8343 24.4146H34.6047L34.5092 24.7427H33.3664L33.4617 24.4146H31.207L31.4268 23.6582H31.6527L32.7544 19.8496H32.5262H32.5262ZM33.7993 19.8498L33.4989 20.8805C33.4989 20.8805 34.0128 20.6866 34.4558 20.6318C34.5536 20.2718 34.6815 19.8498 34.6815 19.8498H33.7993V19.8498ZM33.3598 21.3637L33.0585 22.4433C33.0585 22.4433 33.628 22.1676 34.0188 22.1444C34.1317 21.7271 34.2447 21.3637 34.2447 21.3637H33.3598V21.3637ZM33.5808 23.6583L33.8067 22.8751H32.9258L32.6987 23.6583H33.5808ZM36.4352 18.9922H37.5447L37.5918 19.3946C37.5844 19.4971 37.6463 19.546 37.7779 19.546H37.9739L37.7756 20.2279H36.9601C36.6487 20.2437 36.4886 20.1267 36.4738 19.8741L36.4352 18.9922ZM36.1102 20.4548H39.7039L39.493 21.1868H38.3488L38.1526 21.8673H39.2957L39.0835 22.5981H37.8104L37.5224 23.0264H38.1455L38.2894 23.8839C38.3066 23.9693 38.3836 24.0108 38.5151 24.0108H38.7086L38.5053 24.717H37.8202C37.4653 24.7342 37.2818 24.6171 37.2667 24.3646L37.1016 23.5814L36.5346 24.4146C36.4005 24.65 36.1945 24.7599 35.9167 24.7427H34.8705L35.074 24.0363H35.4004C35.5345 24.0363 35.646 23.9778 35.7465 23.8595L36.634 22.5981H35.4898L35.7018 21.8673H36.9428L37.1402 21.1868H35.898L36.1102 20.4548Z"
-			fill="white"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M17.1915 13.9492H18.0645L17.9647 14.4493L18.0899 14.3066C18.3729 14.009 18.7166 13.8613 19.1224 13.8613C19.4898 13.8613 19.7547 13.9663 19.921 14.1773C20.0847 14.3884 20.1294 14.6799 20.0519 15.0544L19.571 17.4137H18.6738L19.1081 15.2752C19.1529 15.0544 19.1405 14.8897 19.0715 14.7836C19.0033 14.6774 18.873 14.625 18.685 14.625C18.4542 14.625 18.26 14.6957 18.1017 14.8361C17.9429 14.9776 17.8381 15.174 17.7865 15.424L17.3863 17.4137H16.4873L17.1915 13.9492Z"
-			fill="white"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M27.2021 13.9492H28.0758L27.9767 14.4493L28.1006 14.3066C28.3837 14.009 28.7287 13.8613 29.1332 13.8613C29.5005 13.8613 29.766 13.9663 29.931 14.1773C30.0937 14.3884 30.1408 14.6799 30.0614 15.0544L29.5823 17.4137H28.6839L29.1184 15.2752C29.1629 15.0544 29.1506 14.8897 29.0823 14.7836C29.0115 14.6774 28.8836 14.625 28.6964 14.625C28.4655 14.625 28.272 14.6957 28.1119 14.8361C27.953 14.9776 27.8476 15.174 27.798 15.424L27.396 17.4137H26.498L27.2021 13.9492"
-			fill="white"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M31.5212 11.8018H34.0577C34.5454 11.8018 34.9225 11.9104 35.1818 12.1238C35.44 12.3398 35.5692 12.6497 35.5692 13.0534V13.0656C35.5692 13.1424 35.564 13.229 35.5567 13.3229C35.5441 13.4157 35.5279 13.5095 35.5071 13.6072C35.3954 14.1415 35.1359 14.571 34.7352 14.8967C34.333 15.2211 33.8567 15.3846 33.3082 15.3846H31.9479L31.5274 17.4133H30.3496L31.5212 11.8018M32.1554 14.4087H33.2835C33.5776 14.4087 33.8108 14.3415 33.9809 14.2086C34.1497 14.0744 34.2614 13.8695 34.3234 13.5914C34.3332 13.54 34.3394 13.4937 34.3469 13.451C34.3508 13.4108 34.3556 13.3704 34.3556 13.3315C34.3556 13.1326 34.2838 12.9887 34.1397 12.8984C33.9958 12.8068 33.7701 12.763 33.4572 12.763H32.4991L32.1554 14.4087"
-			fill="white"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M40.8406 18.0833C40.4683 18.8615 40.1135 19.3152 39.9051 19.5263C39.6964 19.735 39.2833 20.2205 38.2881 20.1839L38.3737 19.5898C39.2112 19.3361 39.6642 18.1929 39.9223 17.6867L39.6146 13.9587L40.2624 13.9502H40.8059L40.8643 16.2888L41.8829 13.9502H42.9143L40.8406 18.0833Z"
-			fill="white"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M37.9561 14.232L37.5464 14.509C37.1183 14.1796 36.7274 13.9759 35.9731 14.3199C34.9454 14.7883 34.0868 18.381 36.9161 17.1976L37.0774 17.3855L38.1905 17.4135L38.9215 14.1491L37.9561 14.232M37.3233 16.0168C37.1445 16.5353 36.7451 16.8781 36.4324 16.7805C36.1196 16.6853 36.008 16.1851 36.1891 15.6655C36.3678 15.1458 36.7698 14.8042 37.08 14.9018C37.3927 14.997 37.5056 15.4971 37.3233 16.0168Z"
-			fill="white"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M34.3328 5.26107L30.4463 5.25342L34.3106 5.26981C34.318 5.26981 34.3253 5.26107 34.3328 5.26107"
-			fill="#E02F41"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M30.4467 5.27406L23.5378 5.25C23.5204 5.25 23.5024 5.25765 23.4844 5.26531L30.4467 5.27406"
-			fill="#2E4F7D"
-		/>
+export const UnionPay = () => (
+	/* eslint-disable */
+	<svg viewBox="0 0 64 40" fill="none" xmlns="http://www.w3.org/2000/svg">
+		<rect width="64" height="40" fill="white"/>
+		<g clipPath="url(#clip0_17_2784)">
+			<path
+				d="M17.3827 5H29.3414C31.0102 5 32.0489 6.36125 31.6589 8.03563L26.0939 31.94C25.7002 33.6088 24.0314 34.97 22.3589 34.97H10.4002C8.73142 34.97 7.69267 33.6088 8.08267 31.94L13.6514 8.03563C14.0414 6.36125 15.712 5 17.3827 5Z"
+				fill="#E21836"/>
+			<path
+				d="M28.3456 5H42.0988C43.7675 5 43.0138 6.36125 42.6238 8.03563L37.055 31.94C36.665 33.6088 36.7869 34.97 35.1144 34.97H21.3613C19.6888 34.97 18.6538 33.6088 19.0475 31.94L24.6144 8.03563C25.0063 6.36125 26.675 5 28.3438 5"
+				fill="#00447C"/>
+			<path
+				d="M41.5531 5H53.5119C55.1825 5 56.2212 6.36125 55.8294 8.03563L50.2606 31.94C49.8687 33.6088 48.1981 34.97 46.5256 34.97H34.5725C32.9 34.97 31.8631 33.6088 32.2531 31.94L37.8219 8.03563C38.21 6.36125 39.8806 5 41.5531 5Z"
+				fill="#007B84"/>
+			<path
+				d="M20.5063 12.6613C19.2763 12.6725 18.9125 12.6613 18.7963 12.6331C18.7532 12.845 17.9225 16.6738 17.9207 16.6756C17.7407 17.45 17.6113 18.0031 17.1707 18.3594C16.9229 18.5649 16.6093 18.674 16.2875 18.6669C15.7438 18.6669 15.425 18.395 15.3707 17.8813L15.3613 17.705L15.5263 16.6625C15.5263 16.6625 16.3982 13.175 16.5538 12.7138L16.5669 12.6613C14.8719 12.6763 14.57 12.6613 14.5494 12.6331C14.5382 12.6706 14.4969 12.8863 14.4969 12.8863L13.6063 16.8181L13.5294 17.1519L13.3813 18.2431C13.3725 18.525 13.4378 18.8043 13.5707 19.0531C13.9757 19.7619 15.1325 19.8688 15.785 19.8688C16.6288 19.8688 17.4182 19.6888 17.9525 19.3625C18.8788 18.815 19.1225 17.9581 19.3382 17.1969L19.4375 16.8069C19.4375 16.8069 20.3357 13.1844 20.4875 12.7138C20.49 12.6954 20.4957 12.6776 20.5044 12.6613H20.5063ZM23.5588 15.5825C23.3432 15.5825 22.9475 15.635 22.5913 15.8094C22.4638 15.8769 22.3419 15.9519 22.2125 16.0288L22.3288 15.6088L22.265 15.5394C21.5132 15.6913 21.3444 15.7119 20.6488 15.8094L20.5907 15.8469C20.51 16.5181 20.4369 17.0206 20.1388 18.3369C20.0244 18.8225 19.9063 19.3119 19.7882 19.7956L19.82 19.8575C20.5325 19.82 20.7482 19.82 21.3669 19.8294L21.4175 19.775C21.4963 19.3719 21.5075 19.2781 21.68 18.4625C21.7625 18.0744 21.9332 17.225 22.0175 16.9213C22.1573 16.8428 22.3118 16.7943 22.4713 16.7788C22.8182 16.7788 22.7769 17.0825 22.7619 17.2025C22.6956 17.6842 22.6054 18.1622 22.4919 18.635L22.4057 19.0006C22.3457 19.2725 22.2782 19.535 22.2182 19.8031L22.2444 19.8575C22.9475 19.82 23.1613 19.82 23.7613 19.8294L23.8325 19.775C23.9394 19.1469 23.9713 18.9781 24.1644 18.0613L24.2619 17.6394C24.4494 16.8144 24.5432 16.3981 24.4007 16.0569C24.2507 15.6763 23.8907 15.5825 23.5569 15.5825H23.5588ZM26.9694 16.445C26.5944 16.5181 26.3563 16.565 26.12 16.595C25.8857 16.6344 25.6569 16.67 25.295 16.7188L25.2669 16.745L25.2407 16.7638C25.2032 17.0338 25.1769 17.2663 25.1282 17.54C25.085 17.8213 25.0213 18.1419 24.9144 18.6031C24.8725 18.8071 24.8155 19.0077 24.7438 19.2031C24.6988 19.3288 24.6482 19.4488 24.5563 19.7956L24.5769 19.8275L24.5957 19.8575C24.9332 19.8406 25.1544 19.8294 25.3794 19.8275C25.6063 19.82 25.8407 19.8275 26.2044 19.8294L26.2363 19.805L26.27 19.775C26.3225 19.4619 26.33 19.3775 26.3638 19.2256C26.3938 19.0625 26.45 18.8356 26.5832 18.23C26.6469 17.945 26.7182 17.6619 26.7838 17.3713L26.9919 16.5125L26.9825 16.4788L26.9694 16.4469V16.445ZM26.9769 15.2825C26.6375 15.0819 26.0413 15.1456 25.6419 15.4231C25.2407 15.6931 25.1957 16.0794 25.5332 16.2819C25.8669 16.4769 26.465 16.4188 26.8625 16.1394C27.2619 15.8619 27.3125 15.4813 26.975 15.2825H26.9769ZM29.0282 19.9194C29.7144 19.9194 30.4194 19.7319 30.9482 19.1694C31.355 18.7138 31.5425 18.0369 31.6063 17.7575C31.8182 16.8331 31.6532 16.4019 31.4469 16.1394C31.1338 15.7381 30.5807 15.6088 30.0069 15.6088C29.6619 15.6088 28.8407 15.6444 28.1975 16.235C27.7363 16.6625 27.5225 17.2419 27.395 17.7969C27.2657 18.3631 27.1175 19.3831 28.0532 19.7619C28.3438 19.8856 28.76 19.9194 29.0282 19.9194ZM28.9757 17.84C29.1332 17.1406 29.3207 16.5519 29.7969 16.5519C30.1719 16.5519 30.1982 16.9888 30.0313 17.6919C30.0013 17.8475 29.8644 18.4269 29.6807 18.6744C29.5513 18.8563 29.3994 18.9669 29.2307 18.9669C29.18 18.9669 28.8819 18.9669 28.8782 18.5244C28.8763 18.305 28.9194 18.0819 28.9757 17.84ZM33.3238 19.8294L33.3763 19.775C33.4513 19.3719 33.4644 19.2781 33.6332 18.4625C33.7175 18.0744 33.8919 17.225 33.9744 16.9213C34.13 16.85 34.28 16.7788 34.43 16.7788C34.775 16.7788 34.7319 17.0825 34.7188 17.2025C34.7057 17.405 34.58 18.065 34.4488 18.635L34.3663 19.0006C34.3025 19.2725 34.235 19.535 34.175 19.8031L34.2013 19.8594C34.9063 19.82 35.1125 19.82 35.7144 19.8294L35.7894 19.775C35.8944 19.145 35.9225 18.9781 36.1213 18.0613L36.215 17.6394C36.4025 16.8144 36.5 16.3981 36.3594 16.0569C36.2057 15.6763 35.8419 15.5825 35.5157 15.5825C35.0324 15.59 34.5631 15.7458 34.1713 16.0288L34.28 15.6088L34.2219 15.5394C33.4682 15.6913 33.2975 15.7119 32.6019 15.8094L32.5494 15.8469C32.4482 16.6851 32.2973 17.5166 32.0975 18.3369C31.9832 18.8225 31.865 19.3119 31.7469 19.7956L31.7788 19.8575C32.4913 19.82 32.705 19.82 33.3238 19.8294ZM38.5007 19.8575C38.5457 19.6419 38.8082 18.3575 38.8119 18.3575C38.8119 18.3575 39.0369 17.42 39.0481 17.3844C39.0481 17.3844 39.1194 17.2869 39.1907 17.2475H39.2938C40.2725 17.2475 41.3788 17.2475 42.245 16.61C42.8481 16.1492 43.2643 15.486 43.4169 14.7425C43.4638 14.5513 43.4994 14.3225 43.4994 14.0938C43.4994 13.7938 43.4375 13.4938 43.2632 13.2631C42.8225 12.6444 41.9432 12.6331 40.9269 12.6294L40.4263 12.6331C39.125 12.65 38.6038 12.6444 38.39 12.6181L38.3375 12.8806L37.8725 15.0444L36.7044 19.85C37.8407 19.835 38.3057 19.835 38.5007 19.8575ZM39.3632 16.0231L39.8582 13.8763L39.8732 13.7656L39.8807 13.6813L40.0775 13.7019L41.1238 13.7919C41.5269 13.9475 41.6919 14.3506 41.5775 14.8756C41.5297 15.1045 41.4315 15.3197 41.2901 15.5059C41.1487 15.6921 40.9676 15.8444 40.76 15.9519C40.4319 16.1169 40.0288 16.13 39.6125 16.13H39.3444L39.3632 16.0231ZM42.4475 17.8794C42.3163 18.4381 42.1663 19.4563 43.0982 19.82C43.3963 19.9456 43.6625 19.985 43.9325 19.97C44.2194 19.955 44.4838 19.8106 44.7294 19.6044L44.6638 19.8594L44.705 19.9156C45.3763 19.8856 45.5844 19.8856 46.3119 19.8913L46.3775 19.8406C46.4825 19.2181 46.5838 18.6125 46.8594 17.4181C46.9925 16.8463 47.1275 16.2819 47.2644 15.7119L47.2438 15.65C46.4938 15.7888 46.2932 15.8188 45.5713 15.92L45.515 15.965L45.4944 16.1338C45.3731 15.9348 45.1881 15.7828 44.9694 15.7025C44.6469 15.575 43.8932 15.7381 43.2444 16.3288C42.7907 16.7488 42.5713 17.3263 42.4475 17.8794ZM44.0225 17.9131C44.1838 17.2269 44.3694 16.6438 44.8475 16.6438C45.1475 16.6438 45.3088 16.925 45.275 17.3994C45.2507 17.5175 45.2225 17.6431 45.1888 17.7838C45.1419 17.99 45.0894 18.1906 45.0388 18.3931C44.9969 18.5195 44.9375 18.6394 44.8625 18.7494C44.7388 18.9256 44.4463 19.0325 44.2794 19.0325C44.2307 19.0325 43.9363 19.0325 43.925 18.5975C43.9317 18.3663 43.965 18.1367 44.0244 17.9131H44.0225ZM52.2538 15.6444L52.1938 15.5769C51.4513 15.7269 51.3181 15.7513 50.6356 15.8431L50.585 15.8938L50.5775 15.9256L50.5757 15.9144C50.0675 17.0863 50.0825 16.8331 49.6681 17.7556L49.6644 17.6431L49.5613 15.6444L49.4956 15.5769C48.7175 15.7269 48.6988 15.7513 47.9806 15.8431L47.9244 15.8938C47.9169 15.9181 47.9169 15.9444 47.9131 15.9725L47.9169 15.9838C48.0069 16.4413 47.9863 16.34 48.0763 17.0638C48.1175 17.42 48.1738 17.7763 48.215 18.1269C48.2863 18.7138 48.3275 19.0044 48.4119 19.9006C47.9281 20.6994 47.8119 21.0031 47.345 21.7063L47.3488 21.7119L47.0207 22.2331C46.9832 22.2875 46.9494 22.325 46.9007 22.34C46.8321 22.3668 46.7584 22.3777 46.685 22.3719H46.5032L46.2332 23.2719L47.1613 23.2888C47.705 23.285 48.05 23.03 48.2338 22.6888L48.8188 21.6875H48.8094L48.8694 21.6163C49.2632 20.7725 52.2519 15.6444 52.2519 15.6444H52.2538ZM42.4475 27.4719H42.0538L43.5125 22.6475H43.9963L44.15 22.1506L44.165 22.7038C44.1463 23.045 44.4144 23.3488 45.1213 23.2981H45.9369L46.2182 22.37H45.9107C45.7344 22.37 45.6519 22.325 45.6631 22.2294L45.6481 21.6669H44.1369V21.6706C43.6475 21.68 42.1869 21.7175 41.8925 21.7963C41.6302 21.878 41.3822 22 41.1575 22.1581L41.3057 21.6613H39.8919L39.5975 22.6475L38.1182 27.5469H37.8313L37.55 28.4694H40.3682L40.2744 28.7769H41.6619L41.7538 28.4694H42.1419L42.4475 27.4719ZM41.2925 23.6281C41.0719 23.701 40.8554 23.7855 40.6438 23.8813L41.0188 22.6475H42.1438L41.8719 23.5475C41.8719 23.5475 41.525 23.5663 41.2907 23.6281H41.2925ZM41.3132 25.3906C41.3132 25.3906 40.9607 25.4338 40.7282 25.4863C40.4994 25.5556 40.0719 25.775 40.0719 25.775L40.4582 24.4925H41.5869L41.3132 25.3906ZM40.6832 27.4831H39.5582L39.8844 26.4013H41.0075L40.6832 27.4831ZM43.3981 24.4925H45.0238L44.7894 25.2481H43.1431L42.8957 26.0731H44.3375L43.2463 27.6106C43.1713 27.7231 43.1019 27.7625 43.025 27.7944C42.95 27.8319 42.8488 27.8788 42.7325 27.8788H42.3332L42.0575 28.7844H43.1038C43.6475 28.7844 43.9663 28.5369 44.2044 28.2125L44.9525 27.1888L45.1138 28.2275C45.1272 28.3036 45.1585 28.3754 45.2051 28.437C45.2517 28.4986 45.3123 28.5483 45.3819 28.5819C45.4869 28.6344 45.5938 28.7244 45.7475 28.7375C45.9107 28.745 46.0288 28.7506 46.1075 28.7506H46.6213L46.9307 27.7381H46.7282C46.6119 27.7381 46.4094 27.7175 46.3775 27.6819C46.3438 27.6369 46.3438 27.5694 46.325 27.4644L46.1619 26.4219H45.4944L45.7869 26.075H47.4294L47.6825 25.2481H46.1619L46.3981 24.4925H47.9169L48.1981 23.5588H43.6757L43.3981 24.4925ZM29.6769 27.695L30.0557 26.4331H31.6138L31.8988 25.4956H30.3388L30.5769 24.7175H32.1013L32.3825 23.81H28.5688L28.2932 24.7175H29.1594L28.9288 25.4938H28.0588L27.7719 26.45H28.6382L28.1319 28.1188C28.0644 28.3381 28.1638 28.4225 28.2275 28.5256C28.2932 28.625 28.3588 28.6925 28.5069 28.73C28.6374 28.764 28.7715 28.7823 28.9063 28.7844H30.6632L30.9763 27.7438L30.1982 27.8525C30.0482 27.8525 29.63 27.8338 29.6769 27.695ZM29.855 21.6556L29.4613 22.3681C29.4086 22.4846 29.3296 22.5873 29.2307 22.6681C29.1707 22.7056 29.0507 22.7206 28.8782 22.7206H28.6719L28.3963 23.6338H29.0807C29.4107 23.6338 29.6619 23.5119 29.7838 23.4519C29.9132 23.3825 29.9469 23.4219 30.0463 23.3263L30.2788 23.1256H32.4163L32.6975 22.175H31.1357L31.4075 21.6556H29.855ZM33.0107 27.7119C32.9732 27.6613 32.9994 27.5675 33.0557 27.3744L33.6388 25.4431H35.7182C36.02 25.4375 36.2394 25.4338 36.3819 25.4244C36.5338 25.4075 36.7007 25.3531 36.8807 25.2556C37.0298 25.1784 37.1554 25.0624 37.2444 24.92C37.4057 24.6699 37.5266 24.3959 37.6025 24.1081L38.3375 21.6613L36.1813 21.6744C36.1813 21.6744 35.5175 21.7719 35.225 21.8806C34.9288 22.0006 34.5088 22.3381 34.5088 22.3381L34.7038 21.6669H33.3725L31.505 27.8506C31.453 28.0207 31.4129 28.1943 31.385 28.37C31.3813 28.4825 31.5275 28.595 31.6213 28.6794C31.7338 28.7638 31.8988 28.7506 32.0563 28.7638C32.225 28.7769 32.4594 28.7825 32.7875 28.7825H33.8113L34.1244 27.7231L33.2075 27.8094C33.11 27.8094 33.0388 27.7569 33.0107 27.7119ZM34.0157 24.1363H36.1982L36.0594 24.5713C36.0407 24.5806 35.9938 24.5488 35.7725 24.575H33.8825L34.0157 24.1363ZM34.4544 22.6775H36.6538L36.4963 23.2025C36.4963 23.2025 35.4594 23.1913 35.2925 23.2213C34.5613 23.3488 34.1357 23.7388 34.1357 23.7388L34.4544 22.6775ZM36.11 26.0281C36.0913 26.0919 36.0632 26.1331 36.0219 26.1613C35.9788 26.1913 35.9057 26.2006 35.7988 26.2006H35.4894L35.5082 25.6719H34.2144L34.16 28.2594C34.16 28.4469 34.1769 28.5556 34.3138 28.6419C34.4507 28.7506 34.8707 28.7638 35.4388 28.7638H36.2488L36.5413 27.7944L35.8363 27.8319L35.6019 27.8469C35.57 27.8338 35.5382 27.8206 35.5044 27.7869C35.4744 27.7569 35.4257 27.7756 35.4332 27.5881L35.4388 26.9244L36.1794 26.8944C36.5788 26.8944 36.7494 26.7631 36.8938 26.6394C37.0344 26.5213 37.0794 26.3844 37.1319 26.2006L37.2557 25.6138H36.2394L36.1082 26.0263L36.11 26.0281Z"
+				fill="#FEFEFE"/>
+		</g>
+		<defs>
+			<clipPath id="clip0_17_2784">
+				<rect width="48" height="30" fill="white" transform="translate(8 5)"/>
+			</clipPath>
+		</defs>
 	</svg>
+	/* eslint-enable */
 );
