+++ packages/js/onboarding/src/images/cards/visa.js
@@ -3,34 +3,20 @@
  */
 import { createElement } from '@wordpress/element';
 
-export default () => (
-	<svg
-		width="64"
-		height="40"
-		viewBox="0 0 64 40"
-		fill="none"
-		xmlns="http://www.w3.org/2000/svg"
-	>
-		<rect width="64" height="40" rx="3" fill="white" />
-		<path
-			d="M28.6564 27.2545H24.7949L27.2102 13.2538H31.0714L28.6564 27.2545Z"
-			fill="#1C34C3"
-		/>
+export const Visa = () => (
+	/* eslint-disable */
+	<svg viewBox="0 0 64 40" fill="none" xmlns="http://www.w3.org/2000/svg">
+		<rect width="64" height="40" rx="3" fill="white"/>
+		<path d="M28.6564 27.2545H24.7949L27.2102 13.2538H31.0714L28.6564 27.2545Z" fill="#1C34C3"/>
 		<path
 			d="M42.6537 13.5961C41.8921 13.3128 40.684 13 39.1903 13C35.3769 13 32.6916 14.9064 32.6751 17.6319C32.6435 19.6428 34.5977 20.7597 36.0594 21.4302C37.5534 22.1154 38.0612 22.5626 38.0612 23.1733C38.046 24.1112 36.854 24.5436 35.7422 24.5436C34.2006 24.5436 33.3745 24.3207 32.1191 23.7989L31.6107 23.5752L31.0703 26.718C31.976 27.1048 33.6446 27.4481 35.3769 27.4631C39.4287 27.4631 42.0665 25.5863 42.0977 22.6818C42.1131 21.088 41.0812 19.8667 38.8564 18.8688C37.5058 18.2282 36.6787 17.7962 36.6787 17.1408C36.6946 16.5449 37.3783 15.9346 38.9029 15.9346C40.1582 15.9047 41.0806 16.1876 41.7793 16.4707L42.1286 16.6194L42.6537 13.5961Z"
-			fill="#1C34C3"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M49.5665 13.2538H52.5534L55.6686 27.2543H52.0933C52.0933 27.2543 51.7434 25.6456 51.6325 25.1541H46.6747C46.5313 25.5263 45.8642 27.2543 45.8642 27.2543H41.8125L47.5482 14.4154C47.9456 13.5068 48.6454 13.2538 49.5665 13.2538ZM49.3285 18.3773C49.3285 18.3773 48.1049 21.4902 47.7869 22.2945H50.9965C50.8377 21.5945 50.1064 18.2432 50.1064 18.2432L49.8366 17.0368C49.7229 17.3475 49.5586 17.7746 49.4478 18.0626C49.3726 18.2579 49.322 18.3893 49.3285 18.3773Z"
-			fill="#1C34C3"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M8.06356 13.2538H14.2763C15.1184 13.2833 15.8017 13.5365 16.0241 14.4307L17.3743 20.8634C17.3744 20.8638 17.3745 20.8643 17.3747 20.8647L17.7879 22.8009L21.5696 13.2538H25.6528L19.5832 27.2396H15.4998L12.058 15.0743C10.8705 14.4234 9.51527 13.8999 8 13.5367L8.06356 13.2538Z"
-			fill="#1C34C3"
-		/>
+			fill="#1C34C3"/>
+		<path fillRule="evenodd" clipRule="evenodd"
+			  d="M49.5665 13.2538H52.5534L55.6686 27.2543H52.0933C52.0933 27.2543 51.7434 25.6456 51.6325 25.1541H46.6747C46.5313 25.5263 45.8642 27.2543 45.8642 27.2543H41.8125L47.5482 14.4154C47.9456 13.5068 48.6454 13.2538 49.5665 13.2538ZM49.3285 18.3773C49.3285 18.3773 48.1049 21.4902 47.7869 22.2945H50.9965C50.8377 21.5945 50.1064 18.2432 50.1064 18.2432L49.8366 17.0368C49.7229 17.3475 49.5586 17.7746 49.4478 18.0626C49.3726 18.2579 49.322 18.3893 49.3285 18.3773Z"
+			  fill="#1C34C3"/>
+		<path fillRule="evenodd" clipRule="evenodd"
+			  d="M8.06356 13.2538H14.2763C15.1184 13.2833 15.8017 13.5365 16.0241 14.4307L17.3743 20.8634C17.3744 20.8638 17.3745 20.8643 17.3747 20.8647L17.7879 22.8009L21.5696 13.2538H25.6528L19.5832 27.2396H15.4998L12.058 15.0743C10.8705 14.4234 9.51527 13.8999 8 13.5367L8.06356 13.2538Z"
+			  fill="#1C34C3"/>
 	</svg>
+	/* eslint-enable */
 );
