+++ packages/js/onboarding/src/images/icons/index.js
@@ -1,0 +1,1 @@
+export * from './lightbulb';
