+++ packages/js/onboarding/src/images/payment-methods/affirm.js
@@ -3,49 +3,37 @@
  */
 import { createElement } from '@wordpress/element';
 
-export default () => (
+export const Affirm = () => (
+	/* eslint-disable */
 	<svg
-		width="64"
-		height="40"
+		xmlns="http://www.w3.org/2000/svg"
 		viewBox="0 0 64 40"
 		fill="none"
-		xmlns="http://www.w3.org/2000/svg"
 	>
-		<rect width="64" height="40" fill="white" />
-		<path d="M53 10.6201H11V27.3789H53V10.6201Z" fill="white" />
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M13.7539 25.935C13.2345 25.935 12.9699 25.6792 12.9699 25.2587C12.9699 24.4747 13.8441 24.2111 15.4387 24.0415C15.4387 25.0862 14.732 25.935 13.749 25.935H13.7539ZM14.44 20.0596C13.3002 20.0596 11.9899 20.5957 11.2783 21.1631L11.9281 22.5352C12.4985 22.0128 13.4207 21.5659 14.2528 21.5659C15.0437 21.5659 15.4808 21.8305 15.4808 22.3637C15.4808 22.7214 15.1917 22.9027 14.6448 22.9742C12.6004 23.2389 10.998 23.8034 10.998 25.3783C10.998 26.6269 11.887 27.3825 13.2757 27.3825C14.2675 27.3825 15.1495 26.8317 15.569 26.1084V27.1865H17.4174V22.6695C17.4174 20.8074 16.1247 20.0557 14.441 20.0557L14.44 20.0596Z"
-			fill="#101820"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M32.9912 20.263V27.1782H34.969V23.8461C34.969 22.2633 35.9274 21.7978 36.5949 21.7978C36.8953 21.7954 37.1897 21.8824 37.4406 22.0477L37.8023 20.2199C37.522 20.1077 37.222 20.0534 36.9202 20.0601C35.9039 20.0601 35.2649 20.51 34.8435 21.4253V20.263H32.9912Z"
-			fill="#101820"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M46.9673 20.0596C45.9216 20.0596 45.1395 20.677 44.7328 21.2729C44.3555 20.5035 43.5568 20.0596 42.5963 20.0596C41.5516 20.0596 40.8273 20.6398 40.4931 21.3081V20.2624H38.5859V27.1777H40.5656V23.6171C40.5656 22.3431 41.235 21.7266 41.8593 21.7266C42.4248 21.7266 42.9442 22.0922 42.9442 23.036V27.1777H44.92V23.6171C44.92 22.3245 45.5727 21.7266 46.2264 21.7266C46.7497 21.7266 47.3045 22.1069 47.3045 23.0223V27.1777H49.2802V22.397C49.2802 20.8436 48.2345 20.0596 46.9712 20.0596"
-			fill="#101820"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M28.1787 20.2633H26.3872V19.5606C26.3872 18.6452 26.9096 18.3845 27.3594 18.3845C27.6666 18.3885 27.9686 18.464 28.2415 18.605L28.8511 17.2104C28.8511 17.2104 28.2326 16.8066 27.1085 16.8066C25.8452 16.8066 24.4075 17.5191 24.4075 19.7546V20.2633H21.4135V19.5606C21.4135 18.6452 21.9348 18.3845 22.3857 18.3845C22.6933 18.3845 22.9963 18.4602 23.2677 18.605L23.8773 17.2104C23.5137 16.9977 22.9286 16.8066 22.1358 16.8066C20.8725 16.8066 19.4347 17.5191 19.4347 19.7546V20.2633H18.29V21.7882H19.4377V27.1785H21.4135V21.7882H24.4114V27.1785H26.3872V21.7882H28.1787V20.2633Z"
-			fill="#101820"
-		/>
-		<path
-			d="M31.3322 20.2627H29.3584V27.175H31.3322V20.2627Z"
-			fill="#101820"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M29.4307 19.0877H31.3643C32.4923 15.5409 36.3194 12.4234 40.8708 12.4234C46.4061 12.4234 51.1897 16.6376 51.1897 23.198C51.2066 24.5447 51.0216 25.8862 50.6409 27.178H52.5177L52.5363 27.1133C52.8526 25.8345 53.0086 24.5213 53.0009 23.2039C53.0009 15.8879 47.6694 10.6221 40.8747 10.6221C35.5374 10.6221 30.7802 14.3266 29.4316 19.0897L29.4307 19.0877Z"
-			fill="#4A4AF4"
-		/>
+		<path fill="#fff" d="M0 0h64v40H0z"/>
+		<mask
+			id="a"
+			width={42}
+			height={18}
+			x={11}
+			y={10}
+			maskUnits="userSpaceOnUse"
+			style={{
+				maskType: "luminance",
+			}}
+		>
+			<path fill="#fff" d="M53 10.62H11V27.38h42V10.62Z"/>
+		</mask>
+		<g mask="url(#a)">
+			<path
+				fill="#101820"
+				d="M13.754 25.935c-.52 0-.784-.256-.784-.676 0-.784.874-1.048 2.469-1.218 0 1.045-.707 1.894-1.69 1.894h.005Zm.686-5.875c-1.14 0-2.45.536-3.162 1.103l.65 1.372c.57-.522 1.493-.97 2.325-.97.79 0 1.228.265 1.228.799 0 .357-.29.539-.836.61-2.045.265-3.647.83-3.647 2.404 0 1.249.889 2.005 2.278 2.005.992 0 1.873-.551 2.293-1.275v1.078h1.848V22.67c0-1.862-1.292-2.613-2.976-2.613l-.001.004ZM32.991 20.263v6.915h1.978v-3.332c0-1.583.958-2.048 1.626-2.048.3-.003.595.084.846.25l.361-1.828a2.24 2.24 0 0 0-.882-.16c-1.016 0-1.655.45-2.077 1.365v-1.162h-1.852ZM46.967 20.06c-1.045 0-1.828.617-2.234 1.213-.377-.77-1.176-1.213-2.137-1.213-1.044 0-1.769.58-2.103 1.248v-1.046h-1.907v6.916h1.98v-3.56c0-1.275.669-1.891 1.293-1.891.566 0 1.085.365 1.085 1.309v4.142h1.976v-3.56c0-1.293.653-1.891 1.306-1.891.524 0 1.078.38 1.078 1.295v4.156h1.976v-4.781c0-1.553-1.046-2.337-2.309-2.337M28.179 20.263h-1.792v-.702c0-.916.523-1.177.972-1.177.308.005.61.08.882.221l.61-1.395s-.618-.403-1.742-.403c-1.264 0-2.702.712-2.702 2.948v.508h-2.994v-.702c0-.916.522-1.177.973-1.177.307 0 .61.076.882.221l.61-1.395c-.364-.212-.95-.403-1.742-.403-1.264 0-2.701.712-2.701 2.948v.508H18.29v1.525h1.148v5.39h1.976v-5.39h2.997v5.39h1.976v-5.39h1.792v-1.525ZM31.332 20.263h-1.974v6.912h1.974v-6.912Z"
+			/>
+			<path
+				fill="#4A4AF4"
+				d="M29.43 19.088h1.934c1.128-3.547 4.955-6.665 9.507-6.665 5.535 0 10.319 4.215 10.319 10.775a13.48 13.48 0 0 1-.55 3.98h1.878l.018-.065c.317-1.279.473-2.592.465-3.91 0-7.315-5.332-12.58-12.126-12.58-5.338 0-10.095 3.704-11.443 8.467l-.001-.002Z"
+			/>
+		</g>
 	</svg>
+	/* eslint-enable */
 );
