+++ packages/js/onboarding/src/images/payment-methods/ideal.js
@@ -3,34 +3,37 @@
  */
 import { createElement } from '@wordpress/element';
 
-export default () => (
+export const Ideal = () => (
+	/* eslint-disable */
 	<svg
-		width="64"
-		height="40"
+		xmlns="http://www.w3.org/2000/svg"
 		viewBox="0 0 64 40"
 		fill="none"
-		xmlns="http://www.w3.org/2000/svg"
 	>
-		<rect width="64" height="40" fill="white" />
+		<path fill="#fff" d="M0 0h64v40H0z"/>
 		<mask
-			id="mask0_1_73"
+			id="a"
+			width={64}
+			height={40}
+			x={0}
+			y={0}
 			maskUnits="userSpaceOnUse"
-			x="0"
-			y="0"
-			width="64"
-			height="40"
+			style={{
+				maskType: 'luminance',
+			}}
 		>
-			<path d="M0.0249939 0H63.72V40H0.0249939V0Z" fill="white" />
+			<path fill="#fff" d="M.025 0H63.72v40H.025V0Z"/>
 		</mask>
-		<g mask="url(#mask0_1_73)">
+		<g mask="url(#a)">
 			<path
-				d="M34.398 6C38.272 6.07 41.544 7.194 43.874 9.272C45.2188 10.4824 46.2595 11.9925 46.912 13.68C47.632 15.502 48 17.63 48 20C48 24.96 46.606 28.662 43.86 31.01C41.6 32.938 38.416 33.94 34.398 33.994V34H16V6H34.398ZM34.05 7.86H17.852V32.14H34.05C42.19 32.14 46.148 28.17 46.148 20C46.148 15.878 44.968 12.736 42.644 10.664C40.588 8.83 37.616 7.86 34.05 7.86ZM25.99 19.15V27.93H20.7V19.15H25.99ZM23.392 11.766C23.7852 11.766 24.1745 11.8434 24.5378 11.9939C24.901 12.1444 25.2311 12.3649 25.5091 12.6429C25.7871 12.9209 26.0076 13.251 26.1581 13.6142C26.3086 13.9775 26.386 14.3668 26.386 14.76C26.386 15.1532 26.3086 15.5425 26.1581 15.9058C26.0076 16.269 25.7871 16.5991 25.5091 16.8771C25.2311 17.1551 24.901 17.3756 24.5378 17.5261C24.1745 17.6766 23.7852 17.754 23.392 17.754C22.5979 17.754 21.8364 17.4386 21.2749 16.8771C20.7134 16.3156 20.398 15.5541 20.398 14.76C20.398 13.9659 20.7134 13.2044 21.2749 12.6429C21.8364 12.0814 22.5979 11.766 23.392 11.766Z"
-				fill="black"
+				fill="#000"
+				d="M34.398 6c3.874.07 7.146 1.194 9.476 3.272a11.52 11.52 0 0 1 3.038 4.408C47.632 15.502 48 17.63 48 20c0 4.96-1.394 8.662-4.14 11.01-2.26 1.928-5.444 2.93-9.462 2.984V34H16V6h18.398Zm-.348 1.86H17.852v24.28H34.05c8.14 0 12.098-3.97 12.098-12.14 0-4.122-1.18-7.264-3.504-9.336C40.588 8.83 37.616 7.86 34.05 7.86Zm-8.06 11.29v8.78H20.7v-8.78h5.29Zm-2.598-7.384a2.994 2.994 0 1 1 0 5.987 2.994 2.994 0 0 1 0-5.987Z"
 			/>
 			<path
-				d="M33.796 28.152H28.416V11.48H33.796H33.576C38.064 11.48 42.838 13.254 42.838 19.836C42.838 26.794 38.064 28.148 33.578 28.148H33.796V28.152Z"
 				fill="#DB4093"
+				d="M33.796 28.152h-5.38V11.48h5.38-.22c4.488 0 9.262 1.774 9.262 8.356 0 6.958-4.774 8.312-9.26 8.312h.218v.004Z"
 			/>
 		</g>
 	</svg>
+	/* eslint-enable */
 );
