+++ packages/js/onboarding/src/images/payment-methods/index.js
@@ -1,0 +1,6 @@
+export * from './affirm';
+export * from './afterpay';
+export * from './clearpay';
+export * from './ideal';
+export * from './klarna';
+export * from './woopay';
