+++ packages/js/onboarding/src/images/payment-methods/klarna.js
@@ -3,20 +3,13 @@
  */
 import { createElement } from '@wordpress/element';
 
-export default () => (
-	<svg
-		width="64"
-		height="40"
-		viewBox="0 0 64 40"
-		fill="none"
-		xmlns="http://www.w3.org/2000/svg"
-	>
-		<rect width="64" height="40" fill="#FFB3C7" />
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M57.4626 23.0076C56.6136 23.0076 55.9253 23.7021 55.9253 24.5588C55.9253 25.4154 56.6136 26.1101 57.4626 26.1101C58.3116 26.1101 59 25.4155 59 24.5588C59 23.702 58.3116 23.0076 57.4626 23.0076ZM52.4048 21.8083C52.4048 20.6353 51.4112 19.6846 50.1855 19.6846C48.9599 19.6846 47.9662 20.6353 47.9662 21.8083C47.9662 22.9812 48.9597 23.9322 50.1855 23.9322C51.4113 23.9322 52.4048 22.9812 52.4048 21.8083ZM52.413 17.6807H54.8622V25.9355H52.413V25.4079C51.7211 25.8841 50.8858 26.164 49.985 26.164C47.601 26.164 45.6684 24.2139 45.6684 21.8082C45.6684 19.4025 47.601 17.4525 49.985 17.4525C50.8858 17.4525 51.7211 17.7324 52.413 18.2088V17.6807ZM32.8122 18.756V17.6808H30.3045V25.9355H32.8178V22.0815C32.8178 20.7812 34.2145 20.0823 35.1835 20.0823C35.1935 20.0823 35.2027 20.0833 35.2127 20.0834V17.6815C34.218 17.6815 33.3032 18.1111 32.8122 18.756ZM26.5656 21.8083C26.5656 20.6353 25.572 19.6847 24.3463 19.6847C23.1206 19.6847 22.1269 20.6354 22.1269 21.8083C22.1269 22.9813 23.1206 23.9322 24.3463 23.9322C25.572 23.9322 26.5656 22.9813 26.5656 21.8083ZM26.5737 17.6807H29.0231V25.9355H26.5737V25.408C25.8818 25.8841 25.0466 26.1641 24.1458 26.1641C21.7618 26.1641 19.8291 24.2139 19.8291 21.8082C19.8291 19.4026 21.7618 17.4526 24.1458 17.4526C25.0466 17.4526 25.8818 17.7325 26.5737 18.2088V17.6807ZM41.3162 17.4587C40.3378 17.4587 39.4117 17.7652 38.7927 18.6109V17.6812H36.3539V25.9356H38.8226V21.5977C38.8226 20.3424 39.6568 19.7277 40.6613 19.7277C41.7378 19.7277 42.3566 20.3766 42.3566 21.5806V25.9356H44.803V20.6863C44.803 18.7653 43.2894 17.4587 41.3162 17.4587ZM16.2614 25.9356H18.825V14.0022H16.2614V25.9356ZM5 25.9389H7.71458V14H5V25.9389ZM14.4947 14C14.4947 16.5847 13.4965 18.9891 11.7178 20.7762L15.4701 25.9393H12.1173L8.03934 20.3282L9.09178 19.5328C10.8371 18.2136 11.8382 16.197 11.8382 14L14.4947 14Z"
-			fill="#0A0B09"
-		/>
+export const Klarna = () => (
+	/* eslint-disable */
+	<svg viewBox="0 0 64 40" fill="none" xmlns="http://www.w3.org/2000/svg">
+		<rect width="64" height="40" fill="#FFB3C7"/>
+		<path fillRule="evenodd" clipRule="evenodd"
+			  d="M57.4626 23.0076C56.6136 23.0076 55.9253 23.7021 55.9253 24.5588C55.9253 25.4154 56.6136 26.1101 57.4626 26.1101C58.3116 26.1101 59 25.4155 59 24.5588C59 23.702 58.3116 23.0076 57.4626 23.0076ZM52.4048 21.8083C52.4048 20.6353 51.4112 19.6846 50.1855 19.6846C48.9599 19.6846 47.9662 20.6353 47.9662 21.8083C47.9662 22.9812 48.9597 23.9322 50.1855 23.9322C51.4113 23.9322 52.4048 22.9812 52.4048 21.8083ZM52.413 17.6807H54.8622V25.9355H52.413V25.4079C51.7211 25.8841 50.8858 26.164 49.985 26.164C47.601 26.164 45.6684 24.2139 45.6684 21.8082C45.6684 19.4025 47.601 17.4525 49.985 17.4525C50.8858 17.4525 51.7211 17.7324 52.413 18.2088V17.6807ZM32.8122 18.756V17.6808H30.3045V25.9355H32.8178V22.0815C32.8178 20.7812 34.2145 20.0823 35.1835 20.0823C35.1935 20.0823 35.2027 20.0833 35.2127 20.0834V17.6815C34.218 17.6815 33.3032 18.1111 32.8122 18.756ZM26.5656 21.8083C26.5656 20.6353 25.572 19.6847 24.3463 19.6847C23.1206 19.6847 22.1269 20.6354 22.1269 21.8083C22.1269 22.9813 23.1206 23.9322 24.3463 23.9322C25.572 23.9322 26.5656 22.9813 26.5656 21.8083ZM26.5737 17.6807H29.0231V25.9355H26.5737V25.408C25.8818 25.8841 25.0466 26.1641 24.1458 26.1641C21.7618 26.1641 19.8291 24.2139 19.8291 21.8082C19.8291 19.4026 21.7618 17.4526 24.1458 17.4526C25.0466 17.4526 25.8818 17.7325 26.5737 18.2088V17.6807ZM41.3162 17.4587C40.3378 17.4587 39.4117 17.7652 38.7927 18.6109V17.6812H36.3539V25.9356H38.8226V21.5977C38.8226 20.3424 39.6568 19.7277 40.6613 19.7277C41.7378 19.7277 42.3566 20.3766 42.3566 21.5806V25.9356H44.803V20.6863C44.803 18.7653 43.2894 17.4587 41.3162 17.4587ZM16.2614 25.9356H18.825V14.0022H16.2614V25.9356ZM5 25.9389H7.71458V14H5V25.9389ZM14.4947 14C14.4947 16.5847 13.4965 18.9891 11.7178 20.7762L15.4701 25.9393H12.1173L8.03934 20.3282L9.09178 19.5328C10.8371 18.2136 11.8382 16.197 11.8382 14L14.4947 14Z"
+			  fill="#0A0B09"/>
 	</svg>
+	/* eslint-enable */
 );
