--- packages/js/onboarding/src/images/payment-methods/woo.js
@@ -1,37 +1,0 @@
-/**
- * External dependencies
- */
-import { createElement } from '@wordpress/element';
-
-export default () => (
-	<svg
-		width="64"
-		height="40"
-		viewBox="0 0 64 40"
-		fill="none"
-		xmlns="http://www.w3.org/2000/svg"
-	>
-		<path
-			d="M61 0H3C1.34315 0 0 1.34315 0 3V37C0 38.6569 1.34315 40 3 40H61C62.6569 40 64 38.6569 64 37V3C64 1.34315 62.6569 0 61 0Z"
-			fill="#674399"
-		/>
-		<path
-			d="M40.8254 15.7289C41.7593 15.7289 42.508 15.9512 43.0713 16.396C43.6346 16.8407 43.9163 17.4263 43.9163 18.1527C43.9163 18.9161 43.6272 19.5165 43.0565 19.9464C42.4857 20.3763 41.6926 20.5987 40.6771 20.5987H40.3065V23.5858H38.1644V15.7289H40.8254ZM40.314 17.2336V19.0866C40.7958 19.0792 41.1515 19.0051 41.3739 18.8568C41.6037 18.7086 41.7149 18.4788 41.7149 18.1675C41.7149 17.8266 41.6037 17.5893 41.3813 17.4559C41.1664 17.3151 40.8105 17.241 40.314 17.2336Z"
-			fill="white"
-		/>
-		<path
-			d="M50.565 23.5784H48.3414L47.9781 22.2664H45.7175L45.3468 23.5784H43.1676L45.6655 15.7215H48.0671L50.565 23.5784ZM47.5705 20.8582C47.2517 19.687 47.0072 18.5751 46.8515 17.5226H46.8219C46.6513 18.6122 46.4142 19.7241 46.1103 20.8582H47.5705Z"
-			fill="white"
-		/>
-		<path
-			d="M49.2827 15.7289H51.6842C52.2179 16.8555 52.6033 17.7302 52.8405 18.338H52.8701C53.1074 17.708 53.4928 16.8407 54.0116 15.7289H56.4651L53.9375 20.0279V23.5784H51.7732V20.0279L49.2827 15.7289Z"
-			fill="white"
-		/>
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			d="M33.4238 12.9999H10.6165C9.17111 12.9999 8 14.171 8 15.6091V24.3109C8 25.7489 9.17111 26.92 10.6091 26.92H21.4161L26.36 29.6699L25.2407 26.92H33.4238C34.8618 26.92 36.0329 25.7489 36.0329 24.3109V15.6091C36.0329 14.171 34.8618 12.9999 33.4238 12.9999ZM10.1348 15.0087C9.85306 15.0161 9.59364 15.1495 9.41577 15.3644C9.24524 15.5868 9.17853 15.8685 9.2379 16.1501C9.91235 20.4492 10.5424 23.3474 11.128 24.8446C11.3429 25.3931 11.6097 25.6525 11.9211 25.6303C12.3954 25.5932 12.9736 24.9335 13.6481 23.6364C13.7981 23.3364 13.9809 22.9721 14.1956 22.5442L14.197 22.5413C14.5019 21.9336 14.8709 21.1982 15.301 20.338C15.9237 22.495 16.7687 24.1182 17.836 25.2078C18.1399 25.5117 18.4364 25.6525 18.7403 25.6303C18.9997 25.6155 19.2369 25.4524 19.348 25.2152C19.4593 24.9855 19.5112 24.726 19.4815 24.4666C19.4073 23.4214 19.5186 21.9613 19.8002 20.1008C20.0893 18.1737 20.4599 16.795 20.9047 15.9648C20.9936 15.8018 21.0307 15.6091 21.0158 15.4163C21.001 15.1569 20.875 14.9123 20.6674 14.7567C20.4673 14.5861 20.2079 14.5046 19.9484 14.5268C19.6223 14.5417 19.3333 14.7344 19.1924 15.0308C18.4882 16.328 17.9842 18.4331 17.6804 21.3386C17.1985 20.0712 16.8354 18.7592 16.5981 17.425C16.4721 16.7431 16.1608 16.4244 15.6568 16.4615C15.3084 16.4836 15.0268 16.7135 14.797 17.1433L12.2843 21.9168C11.8766 20.2491 11.4912 18.2181 11.1354 15.8314C11.0464 15.2458 10.7129 14.9716 10.1348 15.0087ZM26.8567 17.1656C26.4712 16.4762 25.8042 15.9871 25.0259 15.8388C24.8183 15.7943 24.6108 15.7721 24.4032 15.7721C23.3063 15.7721 22.4093 16.3428 21.7126 17.4843C21.1196 18.4553 20.8083 19.5745 20.8232 20.7086C20.8232 21.5906 21.0085 22.3467 21.3716 22.9767C21.7571 23.6661 22.4242 24.1552 23.2024 24.3035C23.41 24.348 23.6175 24.3702 23.8251 24.3702C24.9295 24.3702 25.8263 23.7995 26.5157 22.658C27.1087 21.6796 27.42 20.5604 27.4052 19.4115C27.4052 18.5294 27.2199 17.7808 26.8567 17.1656ZM25.4113 20.3454C25.2556 21.094 24.9665 21.6573 24.5367 22.0428C24.2031 22.3467 23.8918 22.4653 23.6101 22.4134C23.3284 22.3615 23.1061 22.1095 22.9356 21.6722C22.8096 21.3386 22.7355 20.9902 22.7355 20.6271C22.7355 20.3454 22.7651 20.0638 22.817 19.7895C22.9283 19.3003 23.1283 18.8407 23.4248 18.4257C23.8028 17.8697 24.2031 17.6326 24.6182 17.7215C24.8998 17.7808 25.1222 18.0254 25.2927 18.4627C25.4187 18.7963 25.4928 19.1447 25.4928 19.5004C25.4928 19.7821 25.4706 20.0638 25.4113 20.3454ZM32.2008 15.8388C32.979 15.9871 33.6461 16.4762 34.0316 17.1656C34.3948 17.7808 34.5801 18.5294 34.5801 19.4115C34.5949 20.5604 34.2836 21.6796 33.6907 22.658C33.0013 23.7995 32.1044 24.3702 31 24.3702C30.7925 24.3702 30.5849 24.348 30.3774 24.3035C29.5991 24.1552 28.932 23.6661 28.5466 22.9767C28.1834 22.3467 27.998 21.5906 27.998 20.7086C27.9833 19.5745 28.2946 18.4553 28.8876 17.4843C29.5843 16.3428 30.4812 15.7721 31.5782 15.7721C31.7857 15.7721 31.9932 15.7943 32.2008 15.8388ZM31.7116 22.0428C32.1415 21.6573 32.4306 21.094 32.5862 20.3454C32.6456 20.0638 32.6677 19.7821 32.6677 19.5004C32.6677 19.1447 32.5936 18.7963 32.4676 18.4627C32.2972 18.0254 32.0748 17.7808 31.7931 17.7215C31.378 17.6326 30.9778 17.8697 30.5998 18.4257C30.3032 18.8407 30.1031 19.3003 29.992 19.7895C29.9401 20.0638 29.9104 20.3454 29.9104 20.6271C29.9104 20.9902 29.9845 21.3386 30.1105 21.6722C30.2811 22.1095 30.5034 22.3615 30.7851 22.4134C31.0668 22.4653 31.378 22.3467 31.7116 22.0428Z"
-			fill="white"
-		/>
-	</svg>
-);
