+++ packages/js/onboarding/src/images/payment-methods/woopay.js
