+++ packages/js/onboarding/src/images/wcpay-logo.tsx
@@ -17,6 +17,7 @@ export default () => (
 		/>
 		<path
 			fill="#fff"
-			d="M47.22 78.86A32.28 32.28 0 0 1 71.2 67c19.28-1.24 30.4 7.9 33.38 27.44 11.86 79.6 24.72 147.34 38.31 202.97L226.7 138.2c7.67-14.34 17.06-22 28.68-22.75 16.81-1.23 27.2 9.4 31.4 32.14a724.61 724.61 0 0 0 36.1 130.53C333 181.22 349.81 111 373.3 67.75a29.33 29.33 0 0 1 25.21-16.81 32.68 32.68 0 0 1 23.98 7.66 29.64 29.64 0 0 1 11.62 22c.5 6.43-.74 12.86-3.7 18.3-14.84 27.69-27.2 73.67-36.84 137.95-9.4 62.05-13.1 110.75-10.63 145.61a45.6 45.6 0 0 1-4.45 24.97 24.2 24.2 0 0 1-20.27 13.85c-10.14.74-20.03-3.96-30.17-14.1-35.6-36.34-63.78-90.48-84.55-162.42-24.47 48.95-43.01 85.79-55.13 110.01-22.5 43.27-41.78 65.27-57.6 66.5-10.38.75-19.28-7.9-26.45-26.2-19.53-49.94-40.55-146.6-63.04-290a31.9 31.9 0 0 1 5.93-26.2Zm581.72 60.08a89.18 89.18 0 0 0-61.07-44.25 98.89 98.89 0 0 0-20.76-2.23c-36.6 0-66.5 19.04-89.75 57.1a201.58 201.58 0 0 0-29.66 107.55c0 29.42 6.18 54.64 18.3 75.65a89.19 89.19 0 0 0 61.05 44.25 98.88 98.88 0 0 0 20.77 2.23c36.84 0 66.75-19.04 89.74-57.11a203.52 203.52 0 0 0 29.67-108.28c0-29.42-6.18-54.4-18.3-74.91ZM580.73 245c-5.2 24.97-14.84 43.76-29.17 56.61-11.13 10.14-21.51 14.1-30.9 12.36-9.4-1.73-16.82-10.13-22.5-24.72-4.2-11.13-6.68-22.75-6.68-34.86 0-9.4.99-18.79 2.72-27.93a125.56 125.56 0 0 1 20.27-45.5c12.61-18.54 25.96-26.45 39.8-23.48 9.4 1.98 16.82 10.14 22.5 24.72 4.2 11.13 6.68 22.75 6.68 34.61 0 9.4-.75 18.8-2.72 28.19Zm287.51-106.06a89.18 89.18 0 0 0-61.06-44.25 98.89 98.89 0 0 0-20.77-2.23c-36.58 0-66.5 19.04-89.74 57.1a201.59 201.59 0 0 0-29.66 107.55c0 29.42 6.18 54.64 18.29 75.65a89.19 89.19 0 0 0 61.06 44.25 98.88 98.88 0 0 0 20.77 2.23c36.84 0 66.75-19.04 89.74-57.11a203.52 203.52 0 0 0 29.67-108.28c0-29.42-6.18-54.4-18.3-74.91ZM820.04 245c-5.2 24.97-14.84 43.76-29.18 56.61-11.12 10.14-21.5 14.1-30.9 12.36-9.4-1.73-16.81-10.13-22.5-24.72-4.2-11.13-6.67-22.75-6.67-34.86 0-9.4.99-18.79 2.72-27.93a125.56 125.56 0 0 1 20.27-45.5c12.6-18.54 25.96-26.45 39.8-23.48 9.4 1.98 16.82 10.14 22.5 24.72 4.2 11.13 6.68 22.75 6.68 34.61 0 9.4-.75 18.8-2.72 28.19Z"/>
+			d="M47.22 78.86A32.28 32.28 0 0 1 71.2 67c19.28-1.24 30.4 7.9 33.38 27.44 11.86 79.6 24.72 147.34 38.31 202.97L226.7 138.2c7.67-14.34 17.06-22 28.68-22.75 16.81-1.23 27.2 9.4 31.4 32.14a724.61 724.61 0 0 0 36.1 130.53C333 181.22 349.81 111 373.3 67.75a29.33 29.33 0 0 1 25.21-16.81 32.68 32.68 0 0 1 23.98 7.66 29.64 29.64 0 0 1 11.62 22c.5 6.43-.74 12.86-3.7 18.3-14.84 27.69-27.2 73.67-36.84 137.95-9.4 62.05-13.1 110.75-10.63 145.61a45.6 45.6 0 0 1-4.45 24.97 24.2 24.2 0 0 1-20.27 13.85c-10.14.74-20.03-3.96-30.17-14.1-35.6-36.34-63.78-90.48-84.55-162.42-24.47 48.95-43.01 85.79-55.13 110.01-22.5 43.27-41.78 65.27-57.6 66.5-10.38.75-19.28-7.9-26.45-26.2-19.53-49.94-40.55-146.6-63.04-290a31.9 31.9 0 0 1 5.93-26.2Zm581.72 60.08a89.18 89.18 0 0 0-61.07-44.25 98.89 98.89 0 0 0-20.76-2.23c-36.6 0-66.5 19.04-89.75 57.1a201.58 201.58 0 0 0-29.66 107.55c0 29.42 6.18 54.64 18.3 75.65a89.19 89.19 0 0 0 61.05 44.25 98.88 98.88 0 0 0 20.77 2.23c36.84 0 66.75-19.04 89.74-57.11a203.52 203.52 0 0 0 29.67-108.28c0-29.42-6.18-54.4-18.3-74.91ZM580.73 245c-5.2 24.97-14.84 43.76-29.17 56.61-11.13 10.14-21.51 14.1-30.9 12.36-9.4-1.73-16.82-10.13-22.5-24.72-4.2-11.13-6.68-22.75-6.68-34.86 0-9.4.99-18.79 2.72-27.93a125.56 125.56 0 0 1 20.27-45.5c12.61-18.54 25.96-26.45 39.8-23.48 9.4 1.98 16.82 10.14 22.5 24.72 4.2 11.13 6.68 22.75 6.68 34.61 0 9.4-.75 18.8-2.72 28.19Zm287.51-106.06a89.18 89.18 0 0 0-61.06-44.25 98.89 98.89 0 0 0-20.77-2.23c-36.58 0-66.5 19.04-89.74 57.1a201.59 201.59 0 0 0-29.66 107.55c0 29.42 6.18 54.64 18.29 75.65a89.19 89.19 0 0 0 61.06 44.25 98.88 98.88 0 0 0 20.77 2.23c36.84 0 66.75-19.04 89.74-57.11a203.52 203.52 0 0 0 29.67-108.28c0-29.42-6.18-54.4-18.3-74.91ZM820.04 245c-5.2 24.97-14.84 43.76-29.18 56.61-11.12 10.14-21.5 14.1-30.9 12.36-9.4-1.73-16.81-10.13-22.5-24.72-4.2-11.13-6.67-22.75-6.67-34.86 0-9.4.99-18.79 2.72-27.93a125.56 125.56 0 0 1 20.27-45.5c12.6-18.54 25.96-26.45 39.8-23.48 9.4 1.98 16.82 10.14 22.5 24.72 4.2 11.13 6.68 22.75 6.68 34.61 0 9.4-.75 18.8-2.72 28.19Z"
+		/>
 	</svg>
 );
