+++ packages/js/onboarding/src/index.ts
@@ -1,21 +1,9 @@
 export * from './components/WCPayConnectCard';
-export * from './components/WCPayBenefits';
 export * from './components/RecommendedRibbon';
 export * from './components/SetupRequired';
-export * from './components/WCPayAcceptedMethods';
-export { default as Visa } from './images/cards/visa';
-export { default as MasterCard } from './images/cards/mastercard';
-export { default as Amex } from './images/cards/amex';
-export { default as ApplePay } from './images/cards/applepay';
-export { default as GooglePay } from './images/cards/googlepay';
-export { default as Discover } from './images/cards/discover';
-export { default as Diners } from './images/cards/diners';
-export { default as Klarna } from './images/payment-methods/klarna';
-export { default as Affirm } from './images/payment-methods/affirm';
-export { default as AfterPay } from './images/payment-methods/afterpay';
-export { default as ClearPay } from './images/payment-methods/clearpay';
-export { default as Ideal } from './images/payment-methods/ideal';
-export { default as Woo } from './images/payment-methods/woo';
+export * from './images/cards';
+export * from './images/payment-methods';
+export { WooPay as Woo } from './images/payment-methods/woopay';
 export { default as WCPayLogo } from './images/wcpay-logo';
 export { WooPaymentGatewaySetup } from './components/WooPaymentGatewaySetup';
 export { WooPaymentGatewayConfigure } from './components/WooPaymentGatewayConfigure';
