+++ packages/js/onboarding/src/style.scss
@@ -1,4 +1,3 @@
 @import 'components/WCPayConnectCard/WCPayConnectCard.scss';
-@import 'components/WCPayBenefits/WCPayBenefits.scss';
 @import 'components/RecommendedRibbon/RecommendedRibbon.scss';
 @import 'components/Loader/Loader.scss';
